### 2021-11-19 #3932 @whedon
[PRE REVIEW]: 𝚗𝚛𝙲𝚊𝚜𝚌𝚊𝚍𝚎𝚂𝚒𝚖 -- A simulation tool for nuclear recoil cascades resulting from neutron capture
**Submitting author:** @villaa (<a href="http://orcid.org/0000-0002-3893-7259">Anthony Villano</a>)
**Repository:** <a href="https://github.com/villano-lab/nrCascadeSim" target ="_blank">https://github.com/villano-lab/nrCascadeSim</a>
**Version:** v1.2.2
