### 2021-11-19 #3932 @arfon
@villaa – I see this is a resubmission based on the feedback from [an earlier review](https://github.com/openjournals/joss-reviews/issues/3174). Before we assign an editor here, could I ask you to provide a detailed summary of the changes that you've made since the last review (this should help expedite the process on our part).
