### 2021-11-19 #3932 @villaa
@afron sure, the list of things that were done can be found [here](https://raw.githubusercontent.com/villano-lab/nrCascadeSim/master/RELEASENOTES.md). All the work from v1.1.0. to v1.2.2 has been done in response to the earlier review.
