+++ RELEASE_CHECKLIST.md
@@ -8,3 +8,4 @@ Please be sure to do the following when making a release.
 4. be sure the `stable` build of readthedocs points to the new release
 5. update the front-facing README in the Version history and citations sections
 6. update release notes RELEASENOTES.md
+7. update `docs/source/index.rst` if badges changed
