### 2021-11-26 #3932 @Kevin-Mattheus-Moerman
@arfon okay thanks. I cannot thoroughly check if all functionality and documentation issues are fully resolved wrt the previous review (since this software is outside of my area of expertise) but it seems we may be ready to move on based on @villaa 's comments above and the [release notes](https://github.com/villano-lab/nrCascadeSim/blob/master/RELEASENOTES.md), which seem to tick off most of [issues mentioned](https://github.com/openjournals/joss-reviews/issues/3174#issuecomment-931574877) in the [previous review](https://github.com/openjournals/joss-reviews/issues/3174).

@villaa it would be helpful if rather than the less human readable release notes to summarize the required changes from the previous review and to link to where they were resolved. Although I will proceed to find an editor as this point, I think such a list/summary would be very helpful to whoever has to pick this up to handle as editor. Thanks.
