### 2021-11-26 #3932 @Kevin-Mattheus-Moerman
@pibion okay we do consider that a (perceived) COI so I will look for an alternative editor.  Thanks for getting back to me.
