### 2021-11-26 #3932 @villaa
> @villaa while we wait for an editor you have some time to create the clearest possible description here of what was needed in the former review and how the issues raised were addressed. Does that sound good?

@Kevin-Mattheus-Moerman , sure I can provide that. The previous submission had one reviewer give detailed comments and the authors believe we have addressed those comments in full. In the previous review the second reviewer had not given any comments and the author team had made the mistake of waiting on the first reviewer's comments for when the second reviewer responded. This made the review take too long from the perspective of the first reviewer and they suggested we withdraw and re-submit when their comments were addressed. See the review thread [here](https://github.com/openjournals/joss-reviews/issues/3174).

The key issues that were raised are below, they have been addressed. The reviewer was @matthewfeickert so they can comment on how satisfactory these updates are if they want to.

1. although it is not required the reviewer suggested to use a workflow to include github issues and pull requests. We now work exclusively in this mode.

2. Issue9 addressed the lack of contributing protocols. This was addressed in PR19 and closed.

3. Issue3 addressed the lack of a statement of need. It suggested to put it in the README. The README.md in the repository is now shorter but the paper itself has a clear statement of need. Furthermore, the documentation, which has been migrated to [readthedocs](https://nrcascadesim.readthedocs.io/en/latest/?badge=latest) is clear up-front about the applicability of the program.

4. Issue4 the ROOT citation was. incorrect and that was fixed in PR21

5. Issue5 We use an external (older) version of the Mersenne Twister algorithm which needed to have its license included. We included that with PR13

6. Issue7 pointed out that we had an examples directory with no documentation within it and that the .ipynb file did not count as a demo but only "a nice visualization." We have since included a demo script in PR40

7. Issue8 was for typos in the initial README.md top-level documentation and lack of documentation for the custom input "level" files. The typos along with better documentation for the level files were included in PR20. Furthermore this documentation was migrated and further clarified when we moved to [readthedocs](https://nrcascadesim.readthedocs.io/en/latest/?badge=latest).

8. Issue10 was for us to construct an automated testing system. We did this through [TravisCI ](https://nrcascadesim.readthedocs.io/en/latest/?badge=latest) in PR17. We removed official support for OSX (there are still some lingering issues with that) and only formally support Linux Ubuntu Xenial distro. This is the distro that is used in our [TravisCI](https://nrcascadesim.readthedocs.io/en/latest/?badge=latest).

9. Issue11 requested that the paper be made sorter and for us to remove a figure because it was more documentation. This was fixed in PR25

10. Issue12 said that our main executable `realizeCascades` was overly verbose. We included verbosity options in PR16.

11. Issue14 requested a `requirements.txt` in the example directory, this was added in PR15.

12. Issue23 (not suggested by reviewer) was brought up by us as a way to clean up and expand the documentation by using [readthedocs](https://nrcascadesim.readthedocs.io/en/latest/?badge=latest).  This was resolved in PR38.

13. Issue24 (not suggested by reviewer) was to get a [Zenodo DOI](https://doi.org/10.5281/zenodo.5579857) for the code. This was done in PR30.
