### 2021-12-03 #3932 @matthewfeickert
> The key issues that were raised are below, they have been addressed. The reviewer was @matthewfeickert so they can comment on how satisfactory these updates are if they want to.

Hey sorry just saw the email that got sent today (I missed the first one in my slew of GitHub notifications it seems). I'll take a look at this over the weekend and make any comments here if that works for everyone.
