### 2021-12-04 #3932 @lucydot
Hi @matthewfeickert I've just taken a look at the [previous review thread](https://github.com/openjournals/joss-reviews/issues/3174) and can see you have invested some time into reviewing this piece of work - if you are available to give your thoughts on the changes made for this re-submission that would be very useful.
