### 2021-12-10 #3932 @matthewfeickert
> Hi @matthewfeickert I've just taken a look at the [previous review thread](https://github.com/openjournals/joss-reviews/issues/3174) and can see you have invested some time into reviewing this piece of work - if you are available to give your thoughts on the changes made for this re-submission that would be very useful.

:wave: Hi @lucydot. Sure, I ran through a mini-review tonight following my previous set of review comments. Below are my notes on [`nrCascadeSim` `v1.2.3`](https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.2.3).

The TL;DR though is that the team has made lots of great improvements here since their last submission and I think that all of my comments that I would raise in the review could be addressed very quickly, so I would recommend that this submission move forward to review now.

I've reviewed this software 3 times now (including these notes), so I'm happy to serve as a reviewer again as it would probably speed up the process as I have very targeted suggestions at this point. I'll let you and the team decide if that sounds good.

## Notes

As always, these notes are following the [JOSS paper review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html).
For brevity, I'm making comments only on major points that were brought up in the previous rounds of review.

### General Checks

* License issues are resolved RE: Msersenne Twister prime algorithm. :+1:

### Functionality

   - No. The Makefile has hardcoded assumptions on the C++ standard that will be used, while they should instead rely on the information provided by `root-config --cflags`. Here the [Makefile has hardcoded `c++14`](https://github.com/villano-lab/nrCascadeSim/blob/59d756c592dc0801340876e803c4a5393d6b0cc0/Makefile#L35). If this software wants to make the decision to enforce an uppper bound on the C++ standards they support that's not fundamentally wrong, but it needs to be made explicitly clear to the user in advance.

Reproducible example of failing with ROOT compiled against `c++17`:

```console
$ docker run --rm -ti -v $PWD:$PWD -w $PWD atlasamglab/stats-base:root6.24.06
# git clone --depth 1 https://github.com/villano-lab/nrCascadeSim.git --branch v1.2.3 --single-branch
# cd nrCascadeSim/
# gcc --version | head -n 1
gcc (Debian 10.2.1-6) 10.2.1 20210110
# root --version
ROOT Version: 6.24/06
Built for linuxx8664gcc on Oct 19 2021, 19:44:00
From tags/v6-24-06@v6-24-06
# root-config --cflags
-pthread -std=c++17 -m64 -I/usr/local/venv/include
# make
g++ -fPIC -c -D__GIT_VERSION=\"v1.2.3\" -IMersenne -Iinc src//isotope_info.c `root-config --cflags --glibs` -std=c++14 -L. -L/tmp/nrCascadeSim/bin/lib
mv isotope_info.o /tmp/nrCascadeSim/bin/lib/
g++ -fPIC -c -D__GIT_VERSION=\"v1.2.3\" -IMersenne -Iinc src//weisskopf.c `root-config --cflags --glibs` -std=c++14 -L. -L/tmp/nrCascadeSim/bin/lib
mv weisskopf.o /tmp/nrCascadeSim/bin/lib/
g++ -fPIC -c -D__GIT_VERSION=\"v1.2.3\" -IMersenne -Iinc src//lindhard.c `root-config --cflags --glibs` -std=c++14 -L. -L/tmp/nrCascadeSim/bin/lib
mv lindhard.o /tmp/nrCascadeSim/bin/lib/
g++ -fPIC -c -D__GIT_VERSION=\"v1.2.3\" -IMersenne -Iinc src//cascadeProd.c `root-config --cflags --glibs` -std=c++14 -L. -L/tmp/nrCascadeSim/bin/lib
src//cascadeProd.c: In function ‘cri* Cascade(int, int, double, int, double*, double*, double, MTRand*)’:
src//cascadeProd.c:1113:1: warning: control reaches end of non-void function [-Wreturn-type]
 1113 | }
      | ^
mv cascadeProd.o /tmp/nrCascadeSim/bin/lib/
g++ -fPIC -c -D__GIT_VERSION=\"v1.2.3\" -IMersenne -Iinc src//edepmath.c `root-config --cflags --glibs` -std=c++14 -L. -L/tmp/nrCascadeSim/bin/lib
mv edepmath.o /tmp/nrCascadeSim/bin/lib/
g++ -fPIC -c -D__GIT_VERSION=\"v1.2.3\" -IMersenne -Iinc src//rootUtil.c `root-config --cflags --glibs` -std=c++14 -L. -L/tmp/nrCascadeSim/bin/lib
In file included from /usr/local/venv/include/TString.h:29,
                 from /usr/local/venv/include/TNamed.h:26,
                 from /usr/local/venv/include/TDictionary.h:44,
                 from /usr/local/venv/include/TClass.h:23,
                 from /usr/local/venv/include/TTree.h:36,
                 from inc/rootUtil.h:3,
                 from src//rootUtil.c:8:
/usr/local/venv/include/ROOT/RStringView.hxx:84:17: error: expected type-specifier
   84 |        operator std::string_view() const { return std::string_view(fData,fLength); }
      |                 ^~~
In file included from /usr/local/venv/include/TNamed.h:26,
                 from /usr/local/venv/include/TDictionary.h:44,
                 from /usr/local/venv/include/TClass.h:23,
                 from /usr/local/venv/include/TTree.h:36,
                 from inc/rootUtil.h:3,
                 from src//rootUtil.c:8:
/usr/local/venv/include/TString.h:115:13: error: expected type-specifier
  115 |    operator std::string_view() const { return std::string_view(Data(),fExtent); }
      |             ^~~
/usr/local/venv/include/TString.h:280:32: error: ‘string_view’ in namespace ‘std’ does not name a type
  280 |    explicit TString(const std::string_view &sub);
      |                                ^~~~~~~~~~~
/usr/local/venv/include/TString.h:280:27: note: ‘std::string_view’ is only available from C++17 onwards
  280 |    explicit TString(const std::string_view &sub);
      |                           ^~~
/usr/local/venv/include/TString.h:317:37: error: ‘string_view’ in namespace ‘std’ does not name a type
  317 |    TString    &operator=(const std::string_view &s);
      |                                     ^~~~~~~~~~~
/usr/local/venv/include/TString.h:317:32: note: ‘std::string_view’ is only available from C++17 onwards
  317 |    TString    &operator=(const std::string_view &s);
      |                                ^~~
/usr/local/venv/include/TString.h:444:9: error: ‘string_view’ in namespace ‘std’ does not name a type
  444 |    std::string_view View() const { return std::string_view(GetPointer(),Length()); }
      |         ^~~~~~~~~~~
/usr/local/venv/include/TString.h:444:4: note: ‘std::string_view’ is only available from C++17 onwards
  444 |    std::string_view View() const { return std::string_view(GetPointer(),Length()); }
      |    ^~~
In file included from /usr/local/venv/include/TNamed.h:26,
                 from /usr/local/venv/include/TDictionary.h:44,
                 from /usr/local/venv/include/TClass.h:23,
                 from /usr/local/venv/include/TTree.h:36,
                 from inc/rootUtil.h:3,
                 from src//rootUtil.c:8:
/usr/local/venv/include/TString.h:839:53: error: ‘string_view’ in namespace ‘std’ does not name a type
  839 | inline Bool_t operator==(const char *s1, const std::string_view &s2)
      |                                                     ^~~~~~~~~~~
/usr/local/venv/include/TString.h:839:48: note: ‘std::string_view’ is only available from C++17 onwards
  839 | inline Bool_t operator==(const char *s1, const std::string_view &s2)
      |                                                ^~~
/usr/local/venv/include/TString.h:839:15: error: ‘Bool_t operator==(const char*, const int&)’ must have an argument of class or enumerated type
  839 | inline Bool_t operator==(const char *s1, const std::string_view &s2)
      |               ^~~~~~~~
/usr/local/venv/include/TString.h:844:37: error: ‘string_view’ in namespace ‘std’ does not name a type
  844 | inline Bool_t operator==(const std::string_view &s1, const char *s2)
      |                                     ^~~~~~~~~~~
/usr/local/venv/include/TString.h:844:32: note: ‘std::string_view’ is only available from C++17 onwards
  844 | inline Bool_t operator==(const std::string_view &s1, const char *s2)
      |                                ^~~
/usr/local/venv/include/TString.h:844:15: error: ‘Bool_t operator==(const int&, const char*)’ must have an argument of class or enumerated type
  844 | inline Bool_t operator==(const std::string_view &s1, const char *s2)
      |               ^~~~~~~~
/usr/local/venv/include/TString.h:857:37: error: ‘string_view’ in namespace ‘std’ does not name a type
  857 |   std::string printValue(const std::string_view* val);
      |                                     ^~~~~~~~~~~
/usr/local/venv/include/TString.h:857:32: note: ‘std::string_view’ is only available from C++17 onwards
  857 |   std::string printValue(const std::string_view* val);
      |                                ^~~
In file included from inc/rootUtil.h:4,
                 from src//rootUtil.c:8:
/usr/local/venv/include/TFile.h:327:45: error: ‘std::string_view’ has not been declared
  327 |    static Bool_t       SetCacheFileDir(std::string_view cacheDir, Bool_t operateDisconnected = kTRUE,
      |                                             ^~~~~~~~~~~
/usr/local/venv/include/TFile.h: In static member function ‘static Bool_t TFile::SetCacheFileDir(ROOT::Internal::TStringView, Bool_t, Bool_t)’:
/usr/local/venv/include/TFile.h:326:36: error: ‘string_view’ is not a member of ‘std’
  326 |      { return SetCacheFileDir(std::string_view(cacheDir), operateDisconnected, forceCacheread); }
      |                                    ^~~~~~~~~~~
/usr/local/venv/include/TFile.h:326:36: note: ‘std::string_view’ is only available from C++17 onwards
make: *** [Makefile:39: /tmp/nrCascadeSim/bin/lib/rootUtil.o] Error 1
```

If the hardcoded C++ flags are changed to reflect the version of C++ used to compile ROOT then things will pass (this is why the project CI passes, as only `c++14` is tested).

```console
...
# sed -i 's/c++14/c++17/g' Makefile
# make clean && make  # Skipping lots of build warning here for readability
# tree bin
bin
├── lib
│   ├── cascadeProd.o
│   ├── edepmath.o
│   ├── isotope_info.o
│   ├── libncap.so
│   ├── lindhard.o
│   ├── rootUtil.o
│   └── weisskopf.o
├── realizeCascades
└── realizeCascades.cpp

1 directory, 9 files
# command -v realizeCascades  # Just showing that nothing is in PATH yet as expected
# make install
cp /tmp/nrCascadeSim/bin/realizeCascades /usr/local/bin/
# command -v realizeCascades
/usr/local/bin/realizeCascades
```

This should not be the responsibility of the user, but should be discoverable and covered in tests.

The [`install` Make target](https://github.com/villano-lab/nrCascadeSim/blob/59d756c592dc0801340876e803c4a5393d6b0cc0/Makefile#L65-L66) also hardcodes the install target directory to `/user/local/bin`. Install locations can have defaults, but should allow for user input. Similarly, advocating directly for the use of `sudo` for installation of local software should be avoided.

### Documentation

The docs have improved significantly, so great work there! There are some Sphinx errors in terms of syntax, but I don't think that anyone has perfectly typeset anything in Sphinx in the first go, so that shouldn't be held against anyone. These are easy to fix on an afternoon with a coffee/tea too.

There is an outline of the API now too (another great improvement). It is [hardcoded in the docs .rst files](https://github.com/villano-lab/nrCascadeSim/blob/59d756c592dc0801340876e803c4a5393d6b0cc0/docs/source/03_Libraries.rst) though, which is not great as this is extremely brittle and prone to drift. What would be quite a bit nicer would be to generate the docs from the source code itself (for C++ code the obvious method is Doxygen and there are [tutorials](https://devblogs.microsoft.com/cppblog/clear-functional-c-documentation-with-sphinx-breathe-doxygen-cmake/) and [examples](https://github.com/leimao/Sphinx-CPP-TriangleLib) out there on how to build Doxygen annotated projects in Sphinx.)

The [API documentation](https://joss.readthedocs.io/en/latest/review_criteria.html#api-documentation) section of the review guidelines mentions

> the software API is documented to a suitable level.

So while as a library maintainer I can strongly recommend that this be done, it isn't a requirement for the review &mdash; but it would certainly improve the usefulness of the API docs.

#### Example usage

The first example that a user encounters in the [Executables of `nrCascadeSim`](https://github.com/villano-lab/nrCascadeSim/blob/59d756c592dc0801340876e803c4a5393d6b0cc0/docs/source/02_Executables.rst#required-arguments) section of the docs fails. It will run without producing warning or errors, but if run from the top level of the project repo so that `levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt` is in the given example path the content of the resulting ROOT `TTree` in the file is empty

```console
# realizeCascades --verbose -n 100000 -o output.root levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
Seed used: 339870340
MTRand: 0x55c868a47d70
# ls -lh output.root
-rw-r--r-- 1 root root 4.6K Dec 10 07:08 output.root
root@54bfe788713b:~/feickert/Desktop/tmp/nrCascadeSim# root -l output.root
root [0]
Attaching file output.root as _file0...
(TFile *) 0x55c3f3ee8dc0
root [1] _file0->ls()
TFile**		output.root
 TFile*		output.root
  KEY: TTree	cascade;1	cascade
root [2] cascade->Print()
******************************************************************************
*Tree    :cascade   : cascade                                                *
*Entries :        0 : Total =             329 bytes  File  Size =        188 *
*        :          : Tree compression factor =   1.00                       *
******************************************************************************
root [3]
```

This doesn't seem correct given that the example says it will:

> simulate 100000 events for 28Si and output them to a file...

The [examples section in the docs](https://github.com/villano-lab/nrCascadeSim/blob/59d756c592dc0801340876e803c4a5393d6b0cc0/docs/source/04_Examples.rst) are a nice improvement! They center around an

> level input file called inputfile.txt of the correct format

As the project [includes over 30 levelfiles](https://github.com/villano-lab/nrCascadeSim/tree/v1.2.3/levelfiles) it would be nice to show specific run examples and output in the docs as well rather than just example source.

- **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

   - There's CI on Travis-CI against one version of ROOT now. That's a big improvement. It would be nice if there was a known lower bound version of ROOT tested in terms of API dependence and a latest version of ROOT tested as well. Also at the moment the tests are testing the examples, but there isn't coverage reported for the codebase. This isn't a requirement, but it is nice to have so users can have some first heuristic of confidence in the strength of the tests / API stability.

   - There is a `CONTRIBUTING.md` and `CODE_OF_CONDUCT.md` now! :tada:

### Software paper

I didn't review the paper tonight. That can be done in the review.
