### 2021-12-10 #3932 @ziotom78
> @lucydot I have added the following to the pool of potential reviewers that I've cultivated: @danehkar, @smsharma, @AstroBarker, @ziotom78, @manodeep, @kostunin, @aureliocarnero .
>
> Would any of those people be available for this review?

It's outside my domain as well, and I am currently involved in two other reviews for JOSS, so I have to skip this one, sorry…
