### 2021-12-10 #3932 @matthewfeickert
> I am not a fan of the C language and I am not familiar with the review process here, but I can review code both in terms of particle physics and software design.

@altavir, forgive me if I'm wrong, but it doesn't look like you are a registered JOSS reviewer. If you are interested in reviewing for JOSS (that's great!) please first read the [Reviewing for JOSS](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) overview and then [register to become a JOSS reviewer](https://joss.theoj.org/reviewer-signup.html). (Also hi from a fellow particle physicist! :wave: :slightly_smiling_face:)
