### 2021-12-11 #3932 @altavir
@matthewfeickert OK, I never knew about the opportunity before Dmitriy tagged me. I think it is an interesting idea in general - I mean the community to make better scientific software. Though I think it requires more serious measures (code quality expertise, not only documentation quality). I've sent my registration info.
