### 2021-12-14 villano-lab/nrCascadeSim#45 @matthewfeickert
Coming here from https://github.com/villano-lab/nrCascadeSim/issues/48#issuecomment-993035515, I guess this was already know to you given:

https://github.com/villano-lab/nrCascadeSim/blob/59d756c592dc0801340876e803c4a5393d6b0cc0/Makefile#L20-L23
