### 2021-12-14 villano-lab/nrCascadeSim#45 @matthewfeickert
As `nrCascadeSim` can currently only be compiled (I think, so let me know if you have a `c++17` version I should be using) against ROOT releases that were also compiled against `c++14` can you recommend an CVMFS LCG view that you've used to test this that has `c++14` ROOT so I can keep testing now while `c++17` support is considered?

The LCG view that I'm most used to setting up is [LCG view `98python3`](http://lcginfo.cern.ch/release/98python3/) with [`x86_64-centos7-gcc8-opt`](http://lcginfo.cern.ch/platform/x86_64-centos7-gcc8-opt/) but that gives me a `c++17` ROOT:

```console
$ . /cvmfs/sft.cern.ch/lcg/views/LCG_98python3/x86_64-centos7-gcc8-opt/setup.sh
$ command -v root
/cvmfs/sft.cern.ch/lcg/views/LCG_98python3/x86_64-centos7-gcc8-opt/bin/root
$ root --version
ROOT Version: 6.22/00
Built for linuxx8664gcc on Jun 14 2020, 15:54:05
From tags/v6-22-00@v6-22-00
$ root-config --cflags
-pthread -std=c++17 -m64 -I/cvmfs/sft.cern.ch/lcg/releases/ROOT/v6.22.00-be0a0/x86_64-centos7-gcc8-opt/include
```
