### 2021-12-14 #3932 @lucydot
Hi @villaa -  ok it seems we can assign two reviewers to this: @altavir who can give a viewpoint on particle physics and software design, and @matthewfeickert who seems to have the C side of things under control 😉

I'll ask our bot Whedon to start a review issue thread (this is the pre-review thread) - the review guidelines will be listed at the top of that new thread.

@altavir welcome to JOSS! Please ping me in the new review thread with any questions you have about the process. There are review guidelines and a checklist at the new thread (the link will appear below), and there is further reviewer documentation here: [https://joss.readthedocs.io/en/latest/reviewer_guidelines.html](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

We're asking for reviews to be complete within a 4-6 week period however I think given that many of us will be taking a break over Christmas a 1-2 week delay might be expected and is reasonable. As the review is usually an iterative process (rather than something you do in one sitting) it's best for reviews to start earlier in that timeline.
