### 2021-12-14 #3993 @whedon
[REVIEW]: 𝚗𝚛𝙲𝚊𝚜𝚌𝚊𝚍𝚎𝚂𝚒𝚖 -- A simulation tool for nuclear recoil cascades resulting from neutron capture
**Submitting author:** @villaa (<a href="http://orcid.org/0000-0002-3893-7259">Anthony Villano</a>)
**Repository:** https://github.com/villano-lab/nrCascadeSim
**Branch with paper.md** (empty if default branch):
**Version:** v1.4.0
