### 2021-12-14 #3993 @matthewfeickert
@lucydot @altavir As I already have done a few reviews of the project, and @villaa and the rest of the dev team are working on [Issues that I raised in the pre review](https://github.com/openjournals/joss-reviews/issues/3932#issuecomment-990735982) RE: release [`v1.2.3`](https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.2.3) (c.f. https://github.com/openjournals/joss-reviews/issues/3932#issuecomment-993030180) I'm going to hold of on starting my review until they finish those and release [`v1.2.4`](https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.2.4) and notify us here :+1:  (once that is done can we also get whedon to update the Version number in the Issue body?). I suspect that this will only take a few days as the team seemed quite eager to get the review started, which is great.

I'll of course take a look at the paper itself now, but my memory from [the previous review](https://github.com/openjournals/joss-reviews/issues/3174) was that it was basically fine, so I'm not expecting to have many, if any, comments there.
