### 2021-12-15 23ec766
Remove all references to old licensed MT implementation and start working on translation to built-in.
