+++ inc/cascadeProd.h
@@ -1,4 +1,4 @@
-#include "MersenneTwister.h"
+// #include "MersenneTwister.h" // should no longer be needed
 #include <string>
 #include <vector>
 
