+++ src/lindhard.c
@@ -2,7 +2,7 @@
 #include <string>
 #include <iostream>
 #include <cmath>
-#include "MersenneTwister.h"
+// #include "MersenneTwister.h"
 #include "edepmath.h"
 
 using namespace std;
