+++ docs/source/06_Citations.rst
@@ -8,7 +8,7 @@ How to cite
 
 If you decide to use this code, or if you want to add a reference to it, please cite the latest archived version,
 
-    Villano, A.N., Harris, K., Brown, S., 2021, nrCascadeSim - A tool for generating nuclear recoil spectra resulting from neutron capture [Code, v1.2.3] [DOI:10.5281/zenodo.5579857]
+    Villano, A.N., Harris, K., Brown, S., 2021, nrCascadeSim - A tool for generating nuclear recoil spectra resulting from neutron capture [Code, v1.2.4] [DOI:10.5281/zenodo.5579857]
 
 .. raw:: html
 
@@ -19,10 +19,10 @@ If you decide to use this code, or if you want to add a reference to it, please 
 
     @software{nrcascadesim,
     author = {Villano, A.N. and Harris, K. and Brown S.},
-    title = {{nrCascadeSim - A tool for generating nuclear recoil spectra resulting from neutron capture [Code, v1.2.3]}},
+    title = {{nrCascadeSim - A tool for generating nuclear recoil spectra resulting from neutron capture [Code, v1.2.4]}},
     year         = {2021},
     publisher    = {Zenodo},
-    version      = {v1.2.3},
+    version      = {v1.2.4},
     doi          = {DOI:10.5281/zenodo.5579857},
     url          = {https://doi.org/10.5281/zenodo.5579857},
     howpublished={The code can be found under \url{https://github.com/villano-lab}.}
