### 2021-12-16 36000f7
[ANV] added more standard --version text
