### 2021-12-16 #3993 @villaa
@matthewfeickert @altavir we have released `v1.2.4` which takes care of an obvious bug while running the `realizeCascades` executable. This should allow better testing for now. The team is also working on the other major thing revealed in the. pre-review, which is related to issue 27 in `nrCascadeSim`. When that is complete we expect to go to `v1.3.0` and I will notify here.
