### 2021-12-16 #3993 @lucydot
That's excellent thanks @villaa and @matthewfeickert for the updates.

@altavir I can see you have also started working through the reviewer checklist which is great. I guess you've probably already seen the link to the previous review of this code, but just in case you missed it - it is here: https://github.com/openjournals/joss-reviews/issues/3174.
