+++ src/cascadeProd.c
@@ -45,7 +45,7 @@ void freecri(cri *cascade_data)
 }
 //do a generalized multi-step cascade (for now just print a table and do one event)
 //eventually: can do n events, put in a yield model function, generalize to other elements 
-cri *geCascade(int n, int cid, double Sn, int nlev, double *Elev, double *taus, double A, MTRand *mtrand)
+cri *geCascade(int n, int cid, double Sn, int nlev, double *Elev, double *taus, double A, mt19937 *mtrand)
 {
   //input:
   //the neutron separation Sn in MeV
@@ -134,7 +134,7 @@ cri *geCascade(int n, int cid, double Sn, int nlev, double *Elev, double *taus, 
   return outinfo;
 }
 //return the Energy after the mid-stop kick. 
-double geDecay(double v, double M, double Egam, MTRand *mtrand)
+double geDecay(double v, double M, double Egam, mt19937 *mtrand)
 {
 
   //assume v in units of c, M in GeV, and Egam in MeV.
@@ -175,7 +175,7 @@ double geDecay(double v, double M, double Egam, MTRand *mtrand)
   return El;
 }
 //return the velocity at a random stopping time
-double *geStop(double E, double M, double tau, MTRand *mtrand)
+double *geStop(double E, double M, double tau, mt19937 *mtrand)
 {
   //assume energy in eV, mass in GeV, tau in fs.
  
@@ -297,7 +297,7 @@ double vgeS2func(double *x,double *par)
 }
 //do a generalized multi-step cascade (for now just print a table and do one event)
 //eventually: can do n events, put in a yield model function, generalize to other elements 
-cri *siCascade(int n, int cid, double Sn, int nlev, double *Elev, double *taus, double A, MTRand *mtrand)
+cri *siCascade(int n, int cid, double Sn, int nlev, double *Elev, double *taus, double A, mt19937 *mtrand)
 {
   //input:
   //the neutron separation Sn in MeV
@@ -386,7 +386,7 @@ cri *siCascade(int n, int cid, double Sn, int nlev, double *Elev, double *taus, 
   return outinfo;
 }
 //return the Energy after the mid-stop kick. 
-double siDecay(double v, double M, double Egam, MTRand *mtrand)
+double siDecay(double v, double M, double Egam, mt19937 *mtrand)
 {
 
   //assume v in units of c, M in GeV, and Egam in MeV.
@@ -427,7 +427,7 @@ double siDecay(double v, double M, double Egam, MTRand *mtrand)
   return El;
 }
 //return the velocity at a random stopping time
-double *siStop(double E, double M, double tau, MTRand *mtrand)
+double *siStop(double E, double M, double tau, mt19937 *mtrand)
 {
   //assume energy in eV, mass in GeV, tau in fs.
  
@@ -549,7 +549,7 @@ double vsiS2func(double *x,double *par)
 }
 //do a generalized multi-step cascade (for now just print a table and do one event)
 //eventually: can do n events, put in a yield model function, generalize to other elements 
-cri *arCascade(int n,int cid, double Sn, int nlev, double *Elev, double *taus, double A, MTRand *mtrand)
+cri *arCascade(int n,int cid, double Sn, int nlev, double *Elev, double *taus, double A, mt19937 *mtrand)
 {
   //input:
   //the neutron separation Sn in MeV
@@ -639,7 +639,7 @@ cri *arCascade(int n,int cid, double Sn, int nlev, double *Elev, double *taus, d
   return outinfo;
 }
 //return the Energy after the mid-stop kick. 
-double arDecay(double v, double M, double Egam, MTRand *mtrand)
[…]
