### 2021-12-18 d1482d0
[ANV] added static html
