+++ docs/build/html/_sources/07_Release_History.rst.txt
@@ -1,0 +1,8 @@
+===============
+Release history
+===============
+
+* 15.12.2021: Release of `v1.2.4 <https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.2.4>`_
+* 19.11.2021: Release of `v1.2.3 <https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.2.3>`_
+* 13.11.2021: Release of `v1.2.0 <https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.2.0>`_
+* 06.11.2021: Release of `v1.1.3 <https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.1.3>`_
