+++ docs/build/html/_sources/03_Libraries.rst.txt
@@ -1,0 +1,274 @@
+========================================
+3. The libraries of *nrCascadeSim*
+========================================
+
+The functionality of the executables of *nrCascadeSim* such as `realizeCascades` is constructed
+from a set of library functions whose code is stored in the `libncap.so` file and whose prototypes
+are stored in several header files. Below we list the different header files with the internal
+functions, data structures, and their uses. 
+
+---------------
+`cascadeProd.h`
+---------------
+
+This header has the prototypes for the functions that generate the statistical realizations of
+each cascade by Monte-Carlo simulation. The structures and functions related to our data
+structures are prototyped in the table below:
+
+.. role:: C(code)
+   :language: C
+   :class: highlight
+
++-------------------------------------------------------------+---------------------------------------------------+
+| prototypes for structs and functions                        |   purpose                                         |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`struct cli`                                            |  the "Cascade Level Info" struct                  |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`bool cli.sucess`                                       |  has the structure been written correctly         |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`int cli.n`                                             |  number of levels in this cascade                 |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`int cli.cid`                                           |  cascade id                                       |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`double cli.frac`                                       |  probability of observing this particular cascade |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`double cli.Sn`                                         |  neutron separation energy                        |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`int cli.A`                                             |  mass number of capturing isotope                 |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`double* cli.Elev`                                      |  array of energy levels in **keV**                |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`double* cli.taus`                                      |  half-lives of energy level in **as**             |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`struct cri`                                            |  the "Cascade Recoil Info" struct                 |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`int cri.n`                                             |  n levels in the cascade that gave this recoil    |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`int cri.cid`                                           |  cascade id for cascade that gave this recoil     |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`double* cri.E`                                         |  recoil energy at beginning of recoil step **eV** |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`double* cri.delE`                                      |  energy deposited in step **eV**                  |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`int* cri.I`                                            |  ionization created in each step in n e/h pairs   |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`double* cri.Ei`                                        |  electron-equivalent ionization energy **eV**     |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`double* cri.time`                                      |  time since capture **fs**                        |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`double* cri.Eg`                                        |  gamma energy emitted in each step **MeV**        |
++-------------------------------------------------------------+---------------------------------------------------+
+|  :C:`void freecli(cli *cascade_levels)`                     |  function to free memory in cli struct            |
[…]
