+++ docs/build/html/_sources/01_Getting_Started.rst.txt
@@ -1,0 +1,132 @@
+==================
+1. Getting started
+==================
+
+------------
+Installation
+------------
+
+*nrCascadeSim* is designed to run in a Unix-based system and is tested via Travis-CI_ using the
+Xenial_ distribution, Ubuntu 16.04 on x86_64 archetecture.
+
+.. _Travis-CI: https://app.travis-ci.com/github/villano-lab/nrCascadeSim
+.. _Xenial: https://docs.travis-ci.com/user/reference/xenial/ 
+
+^^^^^^^^^^^^
+Dependencies
+^^^^^^^^^^^^
+
+""""""""""""""""""""""""""""""""""""
+1. `ROOT <https://root.cern/>`_
+""""""""""""""""""""""""""""""""""""
+
+To install `ROOT` please follow the instructions on the `CERN  <https://root.cern/install/>`_
+website.
+
+It is intended that *nrCascadeSim* is compatible with all versions; known compatibility with
+version 6. 
+
+On Linux machines, you can also install `ROOT` via a `pre-packaged binary
+<https://root.cern/install/#download-a-pre-compiled-binary-distribution>`_ run::
+
+	wget https://root.cern/download/root_v6.24.02.Linux-ubuntu20-x86_64-gcc9.3.tar.gz
+	tar -xzvf root_v6.24.02.Linux-ubuntu20-x86_64-gcc9.3.tar.gz
+	source root/bin/thisroot.sh # also available: thisroot.{csh,fish,bat}
+
+
+""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
+2. `gcc <https://gcc.gnu.org/>`_
+""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
+
+.. To install *gcc* on a Mac, we can use `homebrew <https://brew.sh/>`_ ::
+
+..	brew install libconfig
+You almost certainly have a version of `gcc` already and *nrCascadeSim* is compatible with version
+4.4.7 or newer.
+
+On Linux machines, you can build `gcc` via the `apt-get` manager::
+
+	sudo apt-get install gcc
+
+
+^^^^^^^^^^^^^^^^
+Download & Build
+^^^^^^^^^^^^^^^^
+
+The `nrCascadeSim` source code can be downloaded by cloning this `git repository
+<https://github.com/villano-lab/nrCascadeSim>`_: ::
+
+   git clone https://github.com/villano-lab/nrCascadeSim.git
+   cd nrCascadeSim 
+
+The code is compiled and the executable and library is built by `Make <https://www.gnu.org/software/make/>`_. To build run the following commands from the repository's root folder.::
+
+	make	
+	sudo make install
+
+If everything worked well, the executable and library file are created as::
+
+	bin/realizeCascades
+	bin/lib/libncap.so	
+
+To clean the installation use::
+
+	make clean
+
+or::
+
+	sudo make clean
+
+if after a `sudo make install`.
+
+
+---------------------------------------
+Using *nrCascadeSim* command-line tools
+---------------------------------------
+
+*nrCascadeSim* contains pre-built executables built on the library functions that can be used as
+command-line tools. See Section 2 for list of the executables and inputs and outputs. 
+
+.. .. warning::
+
+..	The import of these parameters via libconfig is very case-sensitive. A float parameter has to be set to e.g. *1.0*, and **not** just *1*.
+
+..  .. raw:: html
+
+..	<details>
+.. 	<summary><a>The full configuration file</a></summary>
+ 
+ 
+.. .. raw:: html
+
+..	</details>
+
+----------------------------------
+Using *nrCascadeSim* as a library
+----------------------------------
+
+If we want to use *nrCascadeSim* functions in an external code, we can do so and import it as a library.
[…]
