+++ docs/build/html/_sources/References.rst.txt
@@ -1,0 +1,10 @@
+==================
+References
+==================
+
+.. .. [ref] author, *title*, `journal <>`_, `[arXiv:xxxx] <https://arxiv.org/abs/xxxx>`_
+.. https://journals.aps.org/pr/abstract/10.1103/PhysRev.83.1073 Weisskopf estimates
+.. .. [Weisskopf1951] V.F. Weisskopf, *Radiative transition probabilities in nuclei*, `Phys.Rev. 83 (1951) 1073 <https://journals.aps.org/pr/abstract/10.1103/PhysRev.83.1073>`_, `[arXiv:0203002] <https://arxiv.org/abs/0203002>`_.
+.. [Lindhard1963] J. Lindhard, V. Nielsen, M. Scharff, and P.V. Thomsen, *Integral equations governing radiation effects (notes on atomic collisions III)*, `Kgl. Danske Videnskab., Selskab. Mat. Fys. Medd. 10 (1963) <https://www.osti.gov/biblio/4701226-integral-equations-governing-radiation-effects-notes-atomic-collisions-iii>`_
+.. [ROOT1997] Rene Brun and Fons Rademakers , *ROOT  - An object oriented data analysis framework*, `Nucl. Inst. & Meth. in Phys. Res. A 389 (1997); Release v6.22/00 <https://doi.org/10.5281/zenodo.3895852>`_
+.. [Weisskopf1951] V.F. Weisskopf, *Radiative transition probabilities in nuclei*, `Phys.Rev. 83 (1951) 1073 <https://journals.aps.org/pr/abstract/10.1103/PhysRev.83.1073>`_
