### 2021-12-18 02322f9
[ANV] cleaning up random number generation
