+++ inc/edepmath.h
@@ -1,4 +1,3 @@
-// #include "MersenneTwister.h"
 #include <random>
 
 #ifndef EDEP_MATH
