+++ bin/realizeCascades.cpp
@@ -17,6 +17,7 @@
 #include <iomanip>
 #include <fstream>
 #include <cstdlib>
+#include <random>
 #include <map>
 #include <vector>
 #include <algorithm>
