### 2021-12-18 c6b9340
[ANV] tried to make random numbers doubles
