+++ src/cascadeProd.c
@@ -137,7 +137,7 @@ cri *geCascade(int n, int cid, double Sn, int nlev, double *Elev, double *taus, 
 double geDecay(double v, double M, double Egam, mt19937 *mtrand)
 {
   //random distribution
-  uniform_real_distribution<float> dist(0.,1.);
+  uniform_real_distribution<double> dist(0.,1.);
 
   //assume v in units of c, M in GeV, and Egam in MeV.
 
@@ -180,7 +180,7 @@ double geDecay(double v, double M, double Egam, mt19937 *mtrand)
 double *geStop(double E, double M, double tau, mt19937 *mtrand)
 {
   //random distribution
-  uniform_real_distribution<float> dist(0.,1.);
+  uniform_real_distribution<double> dist(0.,1.);
 
   //assume energy in eV, mass in GeV, tau in fs.
  
@@ -394,7 +394,7 @@ cri *siCascade(int n, int cid, double Sn, int nlev, double *Elev, double *taus, 
 double siDecay(double v, double M, double Egam, mt19937 *mtrand)
 {
   //random distribution
-  uniform_real_distribution<float> dist(0.,1.);
+  uniform_real_distribution<double> dist(0.,1.);
 
   //assume v in units of c, M in GeV, and Egam in MeV.
 
@@ -437,7 +437,7 @@ double siDecay(double v, double M, double Egam, mt19937 *mtrand)
 double *siStop(double E, double M, double tau, mt19937 *mtrand)
 {
   //random distribution
-  uniform_real_distribution<float> dist(0.,1.);
+  uniform_real_distribution<double> dist(0.,1.);
 
   //assume energy in eV, mass in GeV, tau in fs.
  
@@ -652,7 +652,7 @@ cri *arCascade(int n,int cid, double Sn, int nlev, double *Elev, double *taus, d
 double arDecay(double v, double M, double Egam, mt19937 *mtrand)
 {
   //random distribution
-  uniform_real_distribution<float> dist(0.,1.);
+  uniform_real_distribution<double> dist(0.,1.);
 
   //assume v in units of c, M in GeV, and Egam in MeV.
 
@@ -695,7 +695,7 @@ double arDecay(double v, double M, double Egam, mt19937 *mtrand)
 double *arStop(double E, double M, double tau, mt19937 *mtrand)
 {
   //random distribution
-  uniform_real_distribution<float> dist(0.,1.);
+  uniform_real_distribution<double> dist(0.,1.);
 
   //assume energy in eV, mass in GeV, tau in fs.
  
@@ -933,7 +933,7 @@ cri *neCascade(int n,int cid, double Sn, int nlev, double *Elev, double *taus, d
 double neDecay(double v, double M, double Egam, mt19937 *mtrand)
 {
   //random distribution
-  uniform_real_distribution<float> dist(0.,1.);
+  uniform_real_distribution<double> dist(0.,1.);
 
   //assume v in units of c, M in GeV, and Egam in MeV.
 
@@ -976,7 +976,7 @@ double neDecay(double v, double M, double Egam, mt19937 *mtrand)
 double *neStop(double E, double M, double tau, mt19937 *mtrand)
 {
   //random distribution
-  uniform_real_distribution<float> dist(0.,1.);
+  uniform_real_distribution<double> dist(0.,1.);
 
   //assume energy in eV, mass in GeV, tau in fs.
  
