+++ src/leSpectrumProd.c
@@ -125,7 +125,7 @@ double geDecay(double v, double M, double Egam, mt19937 *mtrand)
 
   //generate a random direction wrt the incoming direction in the CM
   //I think one angle will do, the azimuthal angle shouldn't matter
-  double costhet = (2*dist(mtrand))-1;
+  double costhet = (2*dist(*mtrand))-1;
   double sinthet = sqrt(1-pow(costhet,2.0));
 
   //calculate the recoiling energy in the CM frame
@@ -172,7 +172,7 @@ double *geStop(double E, double M, double tau, mt19937 *mtrand)
 
   //use the Mersenne Twister for a uniform rand number (tau in fs)
   //this is done by inversion method
-  double t = -log(1-dist(mtrand))*tau;
+  double t = -log(1-dist(*mtrand))*tau;
 
   //return the instantaneous velocity
   ret[0] = vgeS2(E,M,t);
@@ -382,7 +382,7 @@ double siDecay(double v, double M, double Egam, mt19937 *mtrand)
 
   //generate a random direction wrt the incoming direction in the CM
   //I think one angle will do, the azimuthal angle shouldn't matter
-  double costhet = (2*dist(mtrand))-1;
+  double costhet = (2*dist(*mtrand))-1;
   double sinthet = sqrt(1-pow(costhet,2.0));
 
   //calculate the recoiling energy in the CM frame
@@ -429,7 +429,7 @@ double *siStop(double E, double M, double tau, mt19937 *mtrand)
 
   //use the Mersenne Twister for a uniform rand number (tau in fs)
   //this is done by inversion method
-  double t = -log(1-dist(mtrand))*tau;
+  double t = -log(1-dist(*mtrand))*tau;
 
   //return the instantaneous velocity
   ret[0] = vsiS2(E,M,t);
@@ -640,7 +640,7 @@ double arDecay(double v, double M, double Egam, mt19937 *mtrand)
 
   //generate a random direction wrt the incoming direction in the CM
   //I think one angle will do, the azimuthal angle shouldn't matter
-  double costhet = (2*dist(mtrand))-1;
+  double costhet = (2*dist(*mtrand))-1;
   double sinthet = sqrt(1-pow(costhet,2.0));
 
   //calculate the recoiling energy in the CM frame
@@ -687,7 +687,7 @@ double *arStop(double E, double M, double tau, mt19937 *mtrand)
 
   //use the Mersenne Twister for a uniform rand number (tau in fs)
   //this is done by inversion method
-  double t = -log(1-dist(mtrand))*tau;
+  double t = -log(1-dist(*mtrand))*tau;
 
   //return the instantaneous velocity
   ret[0] = varS2(E,M,t);
@@ -921,7 +921,7 @@ double neDecay(double v, double M, double Egam, mt19937 *mtrand)
 
   //generate a random direction wrt the incoming direction in the CM
   //I think one angle will do, the azimuthal angle shouldn't matter
-  double costhet = (2*dist(mtrand))-1;
+  double costhet = (2*dist(*mtrand))-1;
   double sinthet = sqrt(1-pow(costhet,2.0));
 
   //calculate the recoiling energy in the CM frame
@@ -968,7 +968,7 @@ double *neStop(double E, double M, double tau, mt19937 *mtrand)
 
   //use the Mersenne Twister for a uniform rand number (tau in fs)
   //this is done by inversion method
-  double t = -log(1-dist(mtrand))*tau;
+  double t = -log(1-dist(*mtrand))*tau;
 
   //return the instantaneous velocity
   ret[0] = vneS2(E,M,t);
