+++ src/cascadeProd.c
@@ -143,7 +143,7 @@ double geDecay(double v, double M, double Egam, mt19937 *mtrand)
 
   //generate a random direction wrt the incoming direction in the CM
   //I think one angle will do, the azimuthal angle shouldn't matter
-  double costhet = (2*dist(mtrand))-1;
+  double costhet = (2*dist(*mtrand))-1;
   double sinthet = sqrt(1-pow(costhet,2.0));
 
   //calculate the recoiling energy in the CM frame
@@ -190,7 +190,7 @@ double *geStop(double E, double M, double tau, mt19937 *mtrand)
 
   //use the Mersenne Twister for a uniform rand number (tau in fs)
   //this is done by inversion method
-  double t = -log(1-dist(mtrand))*tau;
+  double t = -log(1-dist(*mtrand))*tau;
 
   //return the instantaneous velocity
   ret[0] = vgeS2(E,M,t);
@@ -400,7 +400,7 @@ double siDecay(double v, double M, double Egam, mt19937 *mtrand)
 
   //generate a random direction wrt the incoming direction in the CM
   //I think one angle will do, the azimuthal angle shouldn't matter
-  double costhet = (2*dist(mtrand))-1;
+  double costhet = (2*dist(*mtrand))-1;
   double sinthet = sqrt(1-pow(costhet,2.0));
 
   //calculate the recoiling energy in the CM frame
@@ -447,7 +447,7 @@ double *siStop(double E, double M, double tau, mt19937 *mtrand)
 
   //use the Mersenne Twister for a uniform rand number (tau in fs)
   //this is done by inversion method
-  double t = -log(1-dist(mtrand))*tau;
+  double t = -log(1-dist(*mtrand))*tau;
 
   //return the instantaneous velocity
   ret[0] = vsiS2(E,M,t);
@@ -658,7 +658,7 @@ double arDecay(double v, double M, double Egam, mt19937 *mtrand)
 
   //generate a random direction wrt the incoming direction in the CM
   //I think one angle will do, the azimuthal angle shouldn't matter
-  double costhet = (2*dist(mtrand))-1;
+  double costhet = (2*dist(*mtrand))-1;
   double sinthet = sqrt(1-pow(costhet,2.0));
 
   //calculate the recoiling energy in the CM frame
@@ -705,7 +705,7 @@ double *arStop(double E, double M, double tau, mt19937 *mtrand)
 
   //use the Mersenne Twister for a uniform rand number (tau in fs)
   //this is done by inversion method
-  double t = -log(1-dist(mtrand))*tau;
+  double t = -log(1-dist(*mtrand))*tau;
 
   //return the instantaneous velocity
   ret[0] = varS2(E,M,t);
@@ -939,7 +939,7 @@ double neDecay(double v, double M, double Egam, mt19937 *mtrand)
 
   //generate a random direction wrt the incoming direction in the CM
   //I think one angle will do, the azimuthal angle shouldn't matter
-  double costhet = (2*dist(mtrand))-1;
+  double costhet = (2*dist(*mtrand))-1;
   double sinthet = sqrt(1-pow(costhet,2.0));
 
   //calculate the recoiling energy in the CM frame
@@ -986,7 +986,7 @@ double *neStop(double E, double M, double tau, mt19937 *mtrand)
 
   //use the Mersenne Twister for a uniform rand number (tau in fs)
   //this is done by inversion method
-  double t = -log(1-dist(mtrand))*tau;
+  double t = -log(1-dist(*mtrand))*tau;
 
   //return the instantaneous velocity
   ret[0] = vneS2(E,M,t);
