+++ src/lindhard.c
@@ -2,7 +2,6 @@
 #include <string>
 #include <iostream>
 #include <cmath>
-// #include "MersenneTwister.h"
 #include "edepmath.h"
 
 using namespace std;
