### 2021-12-18 6e75735
[ANV] updated RELEASENOTES for MT pull req
