### 2021-12-19 villano-lab/nrCascadeSim#45 @villaa
I think because of the merging of #54 we should be OK with C++14 or C++17, but should now explicitly test C++17 in the CI. After upgrading that will consider this issue closed.
