+++ docs/source/03_Libraries.rst
@@ -102,59 +102,59 @@ constant-acceleration `S2` stopping. `S2` refers to the parameter from the Lindh
    :language: C
    :class: highlight
 
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-| prototypes for structs and functions                                                                             |   purpose                                         |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`cri *Cascade(int n,int cid,double Sn,int nlev,double *Elev, double *taus, double A, mt19937 *mtrand)`  |  func. for realizing cascades (Si,Ge only for now)|
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`cri *geCascade(int n,int cid,double Sn,int nlev,double *Elev, double *taus, double A, mt19937 *mtrand)`|  germanium func. for realizing cascades           |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
++-------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
+| prototypes for structs and functions                                                                              |   purpose                                         |
++-------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
+|  :C:`cri *Cascade(int n,int cid,double Sn,int nlev,double *Elev, double *taus, double A, mt19937 *mtrand)`        |  func. for realizing cascades (Si,Ge only for now)|
++-------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
+|  :C:`cri *geCascade(int n,int cid,double Sn,int nlev,double *Elev, double *taus, double A, mt19937 *mtrand)`      |  germanium func. for realizing cascades           |
++-------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
 |  :C:`double geDecay(double v, double M, double Egam, mt19937 *rand)`                                              |  germanium func. energy after mid-stop decay      |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
++-------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
 |  :C:`double *geStop(double E, double M, double tau, mt19937 *rand)`                                               |  germanium func. velocity at random stopping time |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`double rgeS2(double E, double M, double t)`                                                                 |  germanium func. returning distance after time t  |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`double vgeS2(double E, double M, double t)`                                                                 |  germanium func. returning velocity after time t  |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`double vgeS2func(double *x,double *par)`                                                                    |  germanium func. velocity as function of time (x) |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`cri *siCascade(int n,int cid,double Sn,int nlev,double *Elev, double *taus, double A, mt19937 *mtrand)`|  silicon func. for realizing cascades             |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`double siDecay(double v, double M, double Egam, mt19937 *rand)`                                        |  silicon func. energy after mid-stop decay        |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`double *siStop(double E, double M, double tau, mt19937 *rand)`                                         |  silicon func. velocity at random stopping time   |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`double rsiS2(double E, double M, double t)`                                                                 |  silicon func. returning distance after time t    |
[…]
