### 2021-12-19 146341e
[ANV] added static docs
