+++ docs/build/html/_sources/01_Getting_Started.rst.txt
@@ -41,12 +41,15 @@ On Linux machines, you can also install `ROOT` via a `pre-packaged binary
 .. To install *gcc* on a Mac, we can use `homebrew <https://brew.sh/>`_ ::
 
 ..	brew install libconfig
+
 You almost certainly have a version of `gcc` already and *nrCascadeSim* is compatible with version
 4.4.7 or newer.
 
 On Linux machines, you can build `gcc` via the `apt-get` manager::
 
 	sudo apt-get install gcc
+
+If, for some reason, you need to use a C++ standard older than 11, please use nrCascadeSim v1.2.2 or older.
 
 
 ^^^^^^^^^^^^^^^^
