+++ docs/build/html/_sources/03_Libraries.rst.txt
@@ -102,59 +102,59 @@ constant-acceleration `S2` stopping. `S2` refers to the parameter from the Lindh
    :language: C
    :class: highlight
 
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-| prototypes for structs and functions                                                                             |   purpose                                         |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`cri *Cascade(int n,int cid, double Sn, int nlev, double *Elev, double *taus, double A, MTRand *mtrand)`     |  func. for realizing cascades (Si,Ge only for now)|
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`cri *geCascade(int n,int cid, double Sn, int nlev, double *Elev, double *taus, double A, MTRand *mtrand)`   |  germanium func. for realizing cascades           |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`double geDecay(double v, double M, double Egam, MTRand *rand)`                                              |  germanium func. energy after mid-stop decay      |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`double *geStop(double E, double M, double tau, MTRand *rand)`                                               |  germanium func. velocity at random stopping time |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`double rgeS2(double E, double M, double t)`                                                                 |  germanium func. returning distance after time t  |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`double vgeS2(double E, double M, double t)`                                                                 |  germanium func. returning velocity after time t  |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`double vgeS2func(double *x,double *par)`                                                                    |  germanium func. velocity as function of time (x) |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`cri *siCascade(int n,int cid, double Sn, int nlev, double *Elev, double *taus, double A, MTRand *mtrand)`   |  silicon func. for realizing cascades             |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`double siDecay(double v, double M, double Egam, MTRand *rand)`                                              |  silicon func. energy after mid-stop decay        |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`double *siStop(double E, double M, double tau, MTRand *rand)`                                               |  silicon func. velocity at random stopping time   |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`double rsiS2(double E, double M, double t)`                                                                 |  silicon func. returning distance after time t    |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`double vsiS2(double E, double M, double t)`                                                                 |  silicon func. returning velocity after time t    |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`double vsiS2func(double *x,double *par)`                                                                    |  silicon func. velocity as function of time (x)   |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`cri *arCascade(int n,int cid, double Sn, int nlev, double *Elev, double *taus, double A, MTRand *mtrand)`   |  argon func. for realizing cascades               |
-+------------------------------------------------------------------------------------------------------------------+---------------------------------------------------+
-|  :C:`double arDecay(double v, double M, double Egam, MTRand *rand)`                                              |  argon func. energy after mid-stop decay          |
[…]
