### 2021-12-20 2fa7b00
[ANV] updated release notes
