+++ RELEASENOTES.md
@@ -2,6 +2,7 @@
 
 * PR #54 Move MersenneTwister to built-in C++ function.
 	* The software now requires C++ 11 or newer.
+* PR #55 fixed sphinx documentation warnings. 
 
 ## Release (v1.2.4) Date 21.12.15
 
