### 2021-12-20 eac6625
[ANV] cleaned up travis build file
