### 2021-12-20 69d936c
[ANV] shifted ROOT install to the script portion of travis file
