+++ .travis.yml
@@ -25,18 +25,24 @@ install:
 
   - sudo apt-get install make cmake g++ gcc binutils libx11-dev libxpm-dev libxft-dev libxext-dev openssl libssl-dev
 
-  #make repository code (my code)
-  - make
-  - sudo make install
 
 script:
   # install ROOT via binary
   - wget https://root.cern/download/root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
   - tar -xzvf root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
   - source root/bin/thisroot.sh
+
+  #make repository code (my code)
+  - make
+  - sudo make install
+
+  #do the tests
   - cd tests
   - ./test_realizeCascades
   - cd ../example-usecase
   - travis_wait 50 py.test
   - source activate piptest
   - travis_wait 50 py.test
+
+  #clean up
+  - sudo make clean
