### 2021-12-20 b54c141
Try testing if generated root file is empty.
