+++ tests/test_realizeCascades.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 
 #for some systems like OSX
 #alias md5sum='md5'
@@ -8,9 +8,4 @@ realizeCascades -n 100000 -o "output.root?reproducible=fixedname" -d 1 ../levelf
 md5sum output.root
 md5sum output.root | awk '{if ($1=="2029e8b402b89d3a05d745e48f546370"){system("rm output.root"); print"Checksum passed!"} else {system("rm output.root"); exit 1}}'
 # Check that file is uncorrupted and has data
-root  -b -l -q <<EOF
-TFile f("output.root")
-if f -> IsZombie():
-	exit(-1)
-fi
-EOF
+python ./isempty.py
