### 2021-12-20 #3993 @lucydot
@altavir @villaa @matthewfeickert I am setting my out of office for the next two week Christmas / New Year period. I will check back in next week to see progress, but a heads up that I may not respond to messages within the typical 3 day period.

RE: updating the software version number for review. I can manually update the version number in the review thread at the top, but first will double check with other members of the editorial team that this won't have have any unwanted side effects. The latest software release for review is `v1.2.4` with an update to `v.1.3.0` expected (as outlined by @villaa in [the post above](https://github.com/openjournals/joss-reviews/issues/3993#issuecomment-995414078)).
