### 2021-12-20 305490c
Have travis give itself permission to run script.
