+++ .travis.yml
@@ -44,6 +44,9 @@ install:
   - make
   - sudo make install
 
+  # give self permission to run script
+  - sudo chmod 777 test/test_realizeCascades.sh
+
 script:
 - cd tests
 - ./test_realizeCascades.sh
