### 2021-12-20 0978bea
[ANV] updated travis script
