+++ .travis.yml
@@ -1,9 +1,6 @@
 language: python
 python: 3.7.9
 dist: xenial
-env:
-  - ROOTDIST="https://root.cern/download/root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"
-  - ROOTDIST="https://root.cern/download/root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"
 install:
 
   #get apt-get and conda
@@ -28,27 +25,30 @@ install:
 
   - sudo apt-get install make cmake g++ gcc binutils libx11-dev libxpm-dev libxft-dev libxext-dev openssl libssl-dev
 
+env:
+  - ROOTDIST="https://root.cern/download/root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"
+  - ROOTDIST="https://root.cern/download/root_v6.24.06.Linux-ubuntu20-x86_64-gcc9.3.tar.gz"
 
 script:
   # install ROOT via binary
   - echo "installing..."
   - echo $ROOTDIST
-  - wget https://root.cern/download/root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
-  - tar -xzvf root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
-  - source root/bin/thisroot.sh
+  #- wget https://root.cern/download/root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
+  #- tar -xzvf root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
+  #- source root/bin/thisroot.sh
 
   #make repository code (my code)
-  - make
-  - sudo make install
+  #- make
+  #- sudo make install
 
   #do the tests
-  - cd tests
-  - ./test_realizeCascades
-  - cd ../example-usecase
-  - travis_wait 50 py.test
-  - source activate piptest
-  - travis_wait 50 py.test
+  #- cd tests
+  #- ./test_realizeCascades
+  #- cd ../example-usecase
+  #- travis_wait 50 py.test
+  #- source activate piptest
+  #- travis_wait 50 py.test
 
   #clean up
-  - cd ../
-  - sudo make clean
+  #- cd ../
+  #- sudo make clean
