+++ .travis.yml
@@ -33,22 +33,22 @@ script:
   # install ROOT via binary
   - echo "installing..."
   - echo $ROOTDIST
-  #- wget https://root.cern/download/root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
-  #- tar -xzvf root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
-  #- source root/bin/thisroot.sh
+  - wget https://root.cern/download/root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
+  - tar -xzf root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
+  - source root/bin/thisroot.sh
 
   #make repository code (my code)
-  #- make
-  #- sudo make install
+  - make
+  - sudo make install
 
   #do the tests
-  #- cd tests
-  #- ./test_realizeCascades
-  #- cd ../example-usecase
-  #- travis_wait 50 py.test
-  #- source activate piptest
-  #- travis_wait 50 py.test
+  - cd tests
+  - ./test_realizeCascades
+  - cd ../example-usecase
+  - travis_wait 50 py.test
+  - source activate piptest
+  - travis_wait 50 py.test
 
   #clean up
-  #- cd ../
-  #- sudo make clean
+  - cd ../
+  - sudo make clean
