+++ .travis.yml
@@ -1,6 +1,9 @@
 language: python
 python: 3.7.9
-dist: xenial
+dist: 
+  - xenial
+  - bionic
+  - focal 
 install:
 
   #get apt-get and conda
@@ -26,14 +29,17 @@ install:
   - sudo apt-get install make cmake g++ gcc binutils libx11-dev libxpm-dev libxft-dev libxext-dev openssl libssl-dev
 
 env:
-  - ROOTDIST="https://root.cern/download/root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"
+  - ROOTDIST="https://root.cern/download/root_v6.24.06.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"
+  - ROOTDIST="https://root.cern/download/root_v6.24.06.Linux-ubuntu18-x86_64-gcc7.5.tar.gz"
   - ROOTDIST="https://root.cern/download/root_v6.24.06.Linux-ubuntu20-x86_64-gcc9.3.tar.gz"
 
 script:
   # install ROOT via binary
   - echo "installing..."
   - echo $ROOTDIST
-  - wget https://root.cern/download/root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
+  - echo "wget command..." 
+  - echo "wget "$ROOTDIST 
+  - wget $ROOTDIST 
   - tar -xzf root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
   - source root/bin/thisroot.sh
 
