+++ .travis.yml
@@ -3,8 +3,11 @@ python: 3.7.9
 jobs:
    include:
     - dist: xenial
+    - env: ROOTDIST="https://root.cern/download/root_v6.24.06.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"
     - dist: bionic
+    - env:  ROOTDIST="https://root.cern/download/root_v6.24.06.Linux-ubuntu18-x86_64-gcc7.5.tar.gz"
     - dist: focal 
+    - env:  ROOTDIST="https://root.cern/download/root_v6.24.06.Linux-ubuntu20-x86_64-gcc9.3.tar.gz"
 install:
 
   #get apt-get and conda
@@ -28,11 +31,6 @@ install:
   - source activate test-environment
 
   - sudo apt-get install make cmake g++ gcc binutils libx11-dev libxpm-dev libxft-dev libxext-dev openssl libssl-dev
-
-env:
-  - ROOTDIST="https://root.cern/download/root_v6.24.06.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"
-  - ROOTDIST="https://root.cern/download/root_v6.24.06.Linux-ubuntu18-x86_64-gcc7.5.tar.gz"
-  - ROOTDIST="https://root.cern/download/root_v6.24.06.Linux-ubuntu20-x86_64-gcc9.3.tar.gz"
 
 script:
   # install ROOT via binary
