+++ .travis.yml
@@ -3,11 +3,11 @@ python: 3.7.9
 jobs:
    include:
     - dist: xenial
-      env: ROOTDIST="https://root.cern/download/root_v6.24.06.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"
+      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"
     - dist: bionic
-      env:  ROOTDIST="https://root.cern/download/root_v6.24.06.Linux-ubuntu18-x86_64-gcc7.5.tar.gz"
+      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu18-x86_64-gcc7.5.tar.gz"
     - dist: focal 
-      env:  ROOTDIST="https://root.cern/download/root_v6.24.06.Linux-ubuntu20-x86_64-gcc9.3.tar.gz"
+      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu20-x86_64-gcc9.3.tar.gz"
 install:
 
   #get apt-get and conda
@@ -37,9 +37,9 @@ script:
   - echo "installing..."
   - echo $ROOTDIST
   - echo "wget command..." 
-  - echo "wget "$ROOTDIST 
-  - wget $ROOTDIST 
-  - tar -xzf root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
+  - echo "wget https://root.cern/download/"$ROOTDIST 
+  - wget https://root.cern/download/$ROOTDIST 
+  - tar -xzvf $ROOTDIST 
   - source root/bin/thisroot.sh
 
   #make repository code (my code)
