### 2021-12-21 bb75da0
Fix ordering so that python script can run while file exists. Don't know why the file stopped existing but this seems to work.
