+++ tests/isempty.py
@@ -1,4 +1,4 @@
 import uproot
 file = uproot.open('output.root') #Change this to change tested file.
-if len(file['cascades'].keys()) == 0:
-	raise ValueError('No keys found under `cascades` - file appears to be empty!')
+if len(file['cascade'].keys()) == 0:
+	raise ValueError('No keys found under `cascade` - file appears to be empty!')
