### 2021-12-21 6a9e692
Fix checksum value. Was using outdated value from before hotfix.
