### 2021-12-21 1c8ff95
Specify that verbosity flags cannot currently handle spaces.
