### 2021-12-21 39b638f
[ANV] write out cflags for root compilation in testing
