### 2021-12-21 ebc6b68
Remove templates from CONTRIBUTING and use github's template management instead. Makes CONTRIBUTING more navigable.
