+++ CONTRIBUTING.md
@@ -116,31 +116,7 @@ are unable to implement the necessary CI, explain why and how it can be manually
 When a PR is accepted it will be staged for a release. We will make note of all the currently
 staged changes in the RELEASENOTES.md file. It is helpful, but not necessary to put a short
 description under the `Next Release` section briefly describing the work done in a PR.  
-
-## Template
-The following template is not required, but if you do not use it, please be sure to include all answers to all of the questions in some other way.
-
-**Does your pull request resolve or partially resolve an issue?** 
-Yes / No.
-
-**If Yes, which issue?** 
-
-**Does your pull request implement code improvements?**
-Yes / No.
-
-**Does your pull request implement any breaking changes?**
-Yes / No.
-
-**If breaking changes are implemented, please describe:**
-
-**Testing:**  
-This pull request:
-[ ] Alters the existing CI in some way.
-[ ] Adds a new step to the CI.
-[ ] Does not introduce any features that the CI could not already test.
-[ ] Is not accompanied by necessary CI changes due to some limitation described below. (Please also describe how new features can be manually tested.)
-
-See the `README.md` file in the `nrCascadeSim/tests` directory for more instructions on how to make tests. 
+A template is provided in `pull_request_template.md`.
 
 **Other information:**
 Anything else you want to say.
@@ -156,10 +132,11 @@ When submitting issues, please be specific. We can only resolve a bug if we know
 program isn't working, implement a feature if we know what aspect is being improved, and clarify
 documentation if we know what part is unclear.
 
-Below are outlines for determining what your issue qualifies as and how to report them. When
+Below are outlines for determining what your issue qualifies as. When
 submitting an issue, please specify which of these three categories you think it belongs in. We
 understand that the three categories can overlap, so don't worry too much if you aren't sure if
-the category you chose is appropriate. Each section also provides a template; these are just to
+the category you chose is appropriate. When creating an issue, you will be given the option to 
+select a template; these are just to
 help people know what to write, and their use is not strictly required (although it may help us
 address the issue faster).
 
@@ -183,30 +160,6 @@ An issue is not a bug report if:
 * The code is missing the ability to do something you think it should be able to do, but the documentation does not specify it is capable of. (Feature request)
 * The documentation is unclear but the code does not give results that directly contradict it. (Documentation issue)
 
-### Template
-
-I am submitting a **bug report**.
-
-**This bug occurs in:**
-make / realizeCascades / CI / specific file / etc.
-
-**Expected behavior:**
-____ should ____.
-
-**Current behavior:**
-____ instead does ____.
-
-**Steps to reproduce:**
-1. Do thing
-2. Do thing
-3. Result
-
-**Other Information:**
-Anything else you want to say.
-
-**Relevant Output:**
-Provide a log file, text from terminal, "No output", etc.
-
 ## Feature request
 
 An issue is a feature request if:
@@ -220,21 +173,6 @@ An issue is not a feature request if:
 * It is to fix unexpected behavior. (Bug report)
 * You are providing the feature you are requesting. (Pull request)
 
-### Template
-
-I am submitting a **feature request**.
-
-**The feature I am requesting is for:**
-make / realizeCascades / CI / specific file / Something new / etc.
-
-**I am requesting:**
-A completely new feature / An improvement on an existing feature / etc.
-
-**Ideas for implementation:**
-(Optional)
-
-**Other Information:**
-Anything else you want to say.
 
 ## Documentation
 
@@ -249,19 +187,4 @@ An issue is not a documentation issue if:
 * The code attempts to run but fails. (Bug report)
[…]
