### 2021-12-21 a005799
Also remove code of conduct from CONTRIBUTING as it's already in a separate file And github has a way of managing this already as well.
