+++ CONTRIBUTING.md
@@ -1,25 +1,3 @@
-# Code of Conduct
-
-The following code of conduct must be followed in all places associated with this repository
-itself. This includes issues, pull requests, and all files contained within the repository. The
-code of conduct applies to all contributors and community members.
-
-In addition to following GitHub's terms of service and any laws applicable to you, please:
-* Do your best to properly attribute any copied, vendored, or otherwise "borrowed" code. In particular, make sure to provide LICENSE files when necessary.
-* Be courteous to other contributors/community members.
-* Do not make major changes to LICENSE or CONTRIBUTING.md
-    * Clarifications and spelling/grammar fixes are fine; modifying the rules is not.
-* Avoid vulgar or offensive language.
-* Stay on-topic; all discussion, including issues and pull requests, must relate to this repository in some way, and should not focus on something other than the code and its modification.
-* No malicious code/malware of any kind, including but not limited to ransomware, adware, bloatware, and spyware.
-
-Failure to comply with this code of conduct could result in having your access to the community
-restricted, such as having offending posts removed or being barred from further submissions.
-
-You can report code of conduct violations to a maintainer (@gerudo7 or @villaa) or the
-@villano-lab group via direct message or email. If you see a violation of GitHub's ToS or of a
-local or federal law, please report it to the appropriate authorities first. Thanks!
-
 # Standards
 
 All features of `nrCascadeSim` should be tested with Travis-CI wherever possible.  This is done
