### 2021-12-21 413abfd
Update release notes Changes from this feature branch as well as the testifempty branch.
