+++ .travis.yml
@@ -39,7 +39,7 @@ script:
   - echo "wget command..." 
   - echo "wget https://root.cern/download/"$ROOTDIST 
   - wget https://root.cern/download/$ROOTDIST 
-  - tar -xzvf $ROOTDIST 
+  - tar -xzf $ROOTDIST 
   - source root/bin/thisroot.sh
   - echo "root libraries.."
   - echo `root-config --cflags -glibs`
@@ -49,7 +49,7 @@ script:
   - sudo make install
 
   # give self permission to run script
-  - sudo chmod 777 tests/test_realizeCascades.sh
+  #- sudo chmod 777 tests/test_realizeCascades.sh
 
   #do the tests
   - cd tests
