+++ .travis.yml
@@ -53,7 +53,7 @@ script:
 
   #do the tests
   - cd tests
-  - ./test_realizeCascades
+  - ./test_realizeCascades.sh
   - cd ../example-usecase
   - travis_wait 50 py.test
   - source activate piptest
