+++ .travis.yml
@@ -49,7 +49,7 @@ script:
   - sudo make install
 
   # give self permission to run script
-  #- sudo chmod 777 tests/test_realizeCascades.sh
+  - sudo chmod 777 tests/test_realizeCascades.sh
 
   #do the tests
   - cd tests
