+++ RELEASENOTES.md
@@ -1,14 +1,15 @@
-## Future Releases (v2.0.0) Date 21.12.14
+## Future Releases (v1.3.0) Date 21.12.21
 
-* PR #54 Move MersenneTwister to built-in C++ function.
+* PR #54 Move MersenneTwister to built-in C++ function. (Issue #27)
 	* The software now requires C++ 11 or newer.
-* PR #55 fixed sphinx documentation warnings. 
-* PR #56 Travis now checks if files are empty.
-* PR #? Improvements to documentation:
+* PR #55 fixed sphinx documentation warnings. (Issue #46)
+* PR #56 Travis now checks if files are empty. (Issue #50)
+* PR #58 Improvements to documentation:
 	* CONTRIBUTING is no longer cluttered with contents of other documents such as templates and the code of conduct.
-	* Syntax for verbosity flags was clarified.
-	* An error in the lifetime units in the documentation was fixed. (All code was already consistent in this regard.)
-	* CONTRIBUTING now outlines a workflow for changes to be made.
+	* Syntax for verbosity flags was clarified. (See comments in issue #48)
+	* An error in the lifetime units in the documentation was fixed. (All code was already consistent in this regard. Issue #53)
+	* CONTRIBUTING now outlines a workflow for changes to be made. (Issue #34)
+	* Added arXiv badge. (See issue #37)
 
 ## Release (v1.2.4) Date 21.12.15
 
