### 2021-12-21 6f2704c
Update notes about distributions tested Include root versions for all tested distros here as well.
