### 2021-12-21 e015582
[ANV] added checksum  values different for different distros
