+++ .travis.yml
@@ -3,11 +3,11 @@ python: 3.7.9
 jobs:
    include:
     - dist: xenial
-      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"
+      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"; CHECKSUM="e38c561bfed333a60b093ce655cdea73"
     - dist: bionic
-      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu18-x86_64-gcc7.5.tar.gz"
+      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu18-x86_64-gcc7.5.tar.gz"; CHECKSUM="e38c561bfed333a60b093ce655cdea73"
     - dist: focal 
-      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu20-x86_64-gcc9.3.tar.gz"
+      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu20-x86_64-gcc9.3.tar.gz"; CHECKSUM="6119a78ef8c534b6bcdd497d796b89b9"
 install:
 
   #get apt-get and conda
@@ -53,6 +53,7 @@ script:
 
   #do the tests
   - cd tests
+  - echo $CHECKSUM
   - ./test_realizeCascades.sh
   - cd ../example-usecase
   - travis_wait 50 py.test
