### 2021-12-21 45b0354
Fix formatting on help output.
