### 2021-12-21 2d51878
[ANV] updated cascadeProd.c function so control doesn't reach end without return
