### 2021-12-21 1cd0066
Remove `docs/build` directory from repository. And ignore future builds.
