+++ docs/build/html/_sources/04_Examples.rst.txt
@@ -1,163 +1,0 @@
-====================================
-4. Examples: Putting it all together
-====================================
-
-
----------------------------
-Reading a Level Input File
----------------------------
-
-This example simply reads a level input file called `inputfile.txt` of the correct format. It then
-prints out the contents of that file in a format showing the properties of each cascade to
-standard output. 
-
-
-.. code-block:: C 
-
-  //library commands
-  #include "cascadeProd.h"
-  #include "lindhard.h"
-  #include "weisskopf.h"
-  #include "isotope_info.h"
-  #include "MersenneTwister.h"
-  
-  //ROOT stuff
-  #include "rootUtil.h"
-
-  //get the file name
-  string filename="inputfile.txt";
-
-  //read the contents of a file into a cli object
-  int numc;
-  bool success=false;
-  cli *cascadeFile = readCascadeDistributionFile(numc,filenames[i],success);
-
-  //print the info that was read in
-  for(int i=0;i<numc;i++){
-      cout << "Cascade ID: " << i+1 << "/" << numc << endl;
-      cout << "Fraction of this cascade: " << cascadeFile[i].frac << endl;
-      cout << "Neutron separation: " << cascadeFile[i].Sn << endl;
-      cout << "Mass number: " << cascadeFile[i].A << endl;
-      cout << "Number of steps: " << cascadeFile[i].n << endl;
-      cout << endl;
-      cout << "Energy Levels (keV)\t|\ttau (fs)" << endl;
-      cout << "------------------------------------------------" << endl;
-      cout << setfill('0') << setw(5) << setprecision(5);
-      cout << "      *****       " << "\t \t" << " ***** " << endl;
-      for(int j=0;j<cascadeFile[i].n;j++){
-        cout << "      "<< cascadeFile[i].Elev[j] << "       " << "\t \t" << " "<< cascadeFile[i].taus[j] << " " << endl;
-   }
-
-   return 0;
-          
-
-----------------------------
-Printing Isotope Information
-----------------------------
-
-This example prints all of the information inside of the isotope database to standard output.
-
-.. code-block:: C 
-
-  //library commands
-  #include "cascadeProd.h"
-  #include "lindhard.h"
-  #include "weisskopf.h"
-  #include "isotope_info.h"
-  #include "MersenneTwister.h"
-  
-  //ROOT stuff
-  #include "rootUtil.h"
-
-  //print out all isotope information in database
-  listStuff();
-
-  return 0;
-
-----------------------------------
-Fetching an Ionization Yield Model
-----------------------------------
-
-This example computes the expected ionization based on the standard Lindhard model
-[Lindhard1963]_. The computation is done for a slowing germanium ion between 100 eV and 50 eV. The
-result is printed to standard output.
-
-.. code-block:: C 
-
-  //library commands
-  #include "cascadeProd.h"
-  #include "lindhard.h"
-  #include "weisskopf.h"
-  #include "isotope_info.h"
-  #include "MersenneTwister.h"
-  
-  //ROOT stuff
-  #include "rootUtil.h"
-
-  //seed an MT random number with 1
-  std::19937 *mtrand = new std::19937(1);
-
-  //get the ionization deposited between 100 and 50 eV
-  double *ionization;
-  double E0 = 100; //eV
-  double E1 = 50; //eV
-  ionization = geIonizationInRange_k(E0,E1,0.159,mtrand); //k-value for Germanium (accepted)
-
-  //print the ionization
-  cout << "Ionization Energy: " << ionization[0] << " eV " << endl;
-  cout << "Ionization Pairs: " << ionization[1] << " eV " << endl;
-
-  return 0;
-
[…]
