+++ docs/build/html/_sources/06_Citations.rst.txt
@@ -1,54 +1,0 @@
-=====================
-Citing *nrCascadeSim*
-=====================
-
------------
-How to cite
------------
-
-If you decide to use this code, or if you want to add a reference to it, please cite the latest archived version,
-
-    Villano, A.N., Harris, K., Brown, S., 2021, nrCascadeSim - A tool for generating nuclear recoil spectra resulting from neutron capture [Code, v1.2.4] [DOI:10.5281/zenodo.5579857]
-
-.. raw:: html
-
-	<details>
-	<summary><a>Bibtex entry.</a></summary>
- 
-.. code-block::
-
-    @software{nrcascadesim,
-    author = {Villano, A.N. and Harris, K. and Brown S.},
-    title = {{nrCascadeSim - A tool for generating nuclear recoil spectra resulting from neutron capture [Code, v1.2.4]}},
-    year         = {2021},
-    publisher    = {Zenodo},
-    version      = {v1.2.4},
-    doi          = {DOI:10.5281/zenodo.5579857},
-    url          = {https://doi.org/10.5281/zenodo.5579857},
-    howpublished={The code can be found under \url{https://github.com/villano-lab}.}
-    }
-
-.. raw:: html
-
-	</details>
-
-
----------------------------------------------------
-Research and research software using *nrCascadeSim*
----------------------------------------------------
-
-The library *nrCascadeSim* has been applied to obtain the scientific results of the following papers
-
-#. **First observation of isolated nuclear recoils following neutron capture**
-  
-  A.N. Villano, M. Fritts, N. Mast, S. Brown, P. Cushman, K. Harris, V. Mandic 
-
-  .. image:: https://img.shields.io/badge/arXiv-2110.02751-B31B1B.svg
-      :target: https://arxiv.org/abs/2110.02751
-      :alt: [arXiv:2110.02751]
-
-
-.. Here is a list of research software using *nrCascadeSim*:
-
-.. #. Emken, T., 2021, `Dark Matter Simulation Code for Underground Scatterings - Sun Edition (DaMaSCUS-SUN) <https://github.com/temken/DaMaSCUS-SUN>`_ Astrophysics Source Code Library, record `[ascl:2102.018] <https://ascl.net/2102.018>`_, `[DOI:10.5281/zenodo.4559874] <https://zenodo.org/record/4559874>`_
-

[…]
