### 2021-12-21 ddb61f4
Include python packages needed to build docs locally.
