### 2021-12-21 6ca8b5e
Forgot to update releasenotes when changing build-related stuff.
