### 2021-12-21 villano-lab/nrCascadeSim#45 @villaa
Closed with PR #57  where we work in all compilers ranging from c++11 to c++17 into Travis-CI
