### 2021-12-21 1c018b6
[ANV] first OSX try and updated release notes
