+++ RELEASENOTES.md
@@ -4,7 +4,7 @@
 	* The software now requires C++ 11 or newer.
 * PR #55 fixed sphinx documentation warnings. 
 * PR #56 tested for root empty files..
-* PR #57 adds testing for c++1--c++17 under ROOT 6.24.08
+* PR #57 adds testing for c++1--c++17 under ROOT 6.24.06
 
 ## Release (v1.2.4) Date 21.12.15
 
