+++ .travis.yml
@@ -1,22 +1,33 @@
-language: python
 jobs:
    include:
     - os: osx
       osx_image: xcode12 
       env:  ROOTDIST="root_v6.24.06.macos-10.15-x86_64-clang120.tar.gz"; CHECKSUM="e38c561bfed333a60b093ce655cdea73"
+      language: generic 
       python: 3.7
     - os: linux
       dist: xenial
       env:  ROOTDIST="root_v6.24.06.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"; CHECKSUM="e38c561bfed333a60b093ce655cdea73"
+      language: python
       python: 3.7
     - os: linux
       dist: bionic
       env:  ROOTDIST="root_v6.24.06.Linux-ubuntu18-x86_64-gcc7.5.tar.gz"; CHECKSUM="e38c561bfed333a60b093ce655cdea73"
+      language: python
       python: 3.7
     - os: linux
       dist: focal 
       env:  ROOTDIST="root_v6.24.06.Linux-ubuntu20-x86_64-gcc9.3.tar.gz"; CHECKSUM="6119a78ef8c534b6bcdd497d796b89b9"
+      language: python
       python: 3.7
+
+# osx deps
+before_install:
+  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew update; fi
+  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew list python &>/dev/null || brew install python; fi
+  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew list python3 &>/dev/null || brew install python3; fi
+  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew install pyenv-virtualenv; fi
+
 install:
 
   #get apt-get and conda
