### 2021-12-22 59089ba
Add version numbers to dependencies to troubleshoot.
