### 2021-12-22 5647fe6
[ANV] started install script for OSX or Linux
