+++ .travis.yml
@@ -58,7 +58,6 @@ install:
   - conda create -n piptest pip
   - source activate piptest
   - pip install -r example-usecase/requirements.txt
-  - source activate test-environment
 
   - sudo apt-get install make cmake g++ gcc binutils libx11-dev libxpm-dev libxft-dev libxext-dev openssl libssl-dev
 
@@ -85,6 +84,9 @@ script:
   - cd tests
   - echo $CHECKSUM
   - ./test_realizeCascades.sh
+
+  #activate python env and continue testing
+  - source activate test-environment
   - cd ../example-usecase
   - travis_wait 50 py.test
   - source activate piptest
