### 2021-12-22 59f826e
[ANV] ROOT 6.20.00 is the earliest version having reproducible files
