### 2021-12-22 a9659ec
[ANV] updated ROOT 6.20.00 checksums
