### 2021-12-22 95eab6e
[ANV] moving to an install script
