### 2021-12-22 be1f25a
[ANV] try install script a different way
