+++ .travis.yml
@@ -86,8 +86,6 @@ script:
   - echo $CHECKSUM
   - ./test_realizeCascades.sh
 
-  #activate python env and continue testing
-  - source activate test-environment
   - cd ../example-usecase
   - travis_wait 50 py.test
   - source activate piptest
