### 2021-12-22 a9652b4
[ANV] activate conda environment
