+++ .travis.yml
@@ -40,7 +40,7 @@ before_install:
 install:
    - sudo chmod 777 .travis/install.sh
    - ls -lh .travis/install.sh
-   - ./.travis/install.sh
+   - source .travis/install.sh
 #  #get apt-get and conda
 #  - sudo apt-get update || true
 #  - sleep 10
