### 2021-12-22 f058812
[ANV] added a baby install for testing
