+++ .travis.yml
@@ -38,62 +38,33 @@ before_install:
   - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew install pyenv-virtualenv; fi
 
 install:
-   - sudo chmod 777 .travis/install.sh
-   - ls -lh .travis/install.sh
-   - source .travis/install.sh
-#  #get apt-get and conda
-#  - sudo apt-get update || true
-#  - sleep 10
-#  - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
-#  - bash miniconda.sh -b -p $HOME/miniconda
-#  - export PATH="$HOME/miniconda/bin:$PATH"
-#  - hash -r
-#  - conda config --set always_yes yes --set changeps1 no
-#  - conda update -q conda || true
-#
-#  # Useful for debugging any issues with conda
-#  - conda info -a
-#
-#  #set up  conda environment
-#  - conda env create -n test-environment python=$TRAVIS_PYTHON_VERSION -f nrCascadeSim_env.yml 
-#  - conda create -n piptest pip
-#  - source activate piptest
-#  - pip install -r example-usecase/requirements.txt
-#
-#  - sudo apt-get install make cmake g++ gcc binutils libx11-dev libxpm-dev libxft-dev libxext-dev openssl libssl-dev
+   - sudo chmod 777 .travis/baby_install.sh
+   - ls -lh .travis/baby_install.sh
+   - source .travis/baby_install.sh
+
+    # Install some custom requirements on Linux
+    #get apt-get and conda
+    - sudo apt-get update || true
+    - sleep 10
+    - sudo apt-get install make cmake g++ gcc binutils libx11-dev libxpm-dev libxft-dev libxext-dev openssl libssl-dev
+    - sleep 10
+    - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
+    - bash miniconda.sh -b -p $HOME/miniconda
+    - export PATH="$HOME/miniconda/bin:$PATH"
+    - hash -r
+    
+    - conda config --set always_yes yes --set changeps1 no
+    - conda update -q conda || true
+      
+    # Useful for debugging any issues with conda
+    - conda info -a
+      
+    #set up  conda environment
+    - conda env create -n test-environment python=$TRAVIS_PYTHON_VERSION -f nrCascadeSim_env.yml 
+    - source activate test-environment
+    - pip install -r example-usecase/requirements.txt
+    - conda list papermill
 
 script:
-  # install ROOT via binary
-  - echo "installing..."
-  - echo $ROOTDIST
-  - echo "wget command..." 
-  - echo "wget https://root.cern/download/"$ROOTDIST 
-  - wget https://root.cern/download/$ROOTDIST 
-  - tar -xzf $ROOTDIST 
-  - source root/bin/thisroot.sh
-  - echo "root libraries.."
-  - echo `root-config --cflags --glibs`
-
-  #make repository code (my code)
-  - make
-  - sudo make install
-
-  # give self permission to run script
-  - sudo chmod 777 tests/test_realizeCascades.sh
-
-  #do the tests
-  - cd tests
-  - echo $CHECKSUM
-  - ./test_realizeCascades.sh
-
-  #need conda environment
   - source activate test-environment
   - conda list papermill
-  - cd ../example-usecase
-  - travis_wait 50 py.test
-  - source activate piptest
-  - travis_wait 50 py.test
-
-  #clean up
-  - cd ../
-  - sudo make clean
