### 2021-12-22 000c1ce
[ANV] added a baby install for testing
