### 2021-12-23 2a51694
[ANV] using simpler conda env for testing
