+++ .travis/install.sh
@@ -37,10 +37,10 @@ else
     conda info -a
     
     #set up  conda environment
-    conda env create -n test-environment python=$TRAVIS_PYTHON_VERSION -f nrCascadeSim_env.yml 
-    source activate test-environment
-    #conda create -n piptest pip
-    #source activate piptest
+    #conda env create -n test-environment python=$TRAVIS_PYTHON_VERSION -f nrCascadeSim_env.yml 
+    #source activate test-environment
+    conda create -n piptest pip
+    source activate piptest
     pip install -r example-usecase/requirements.txt
     conda list papermill
     #activate python env and continue testing
