### 2021-12-23 a917bd9
[ANV] did some conda stuff on Mac environment
