### 2021-12-23 82264a2
[ANV] trimmed down Makefile
