### 2021-12-23 67ff5e7
[ANV] cleaned up makefile more
