### 2021-12-23 3fef7c2
[ANV] actually everything needs the root-config --cflags to be sure the compiler standard is the same
