+++ Makefile
@@ -32,7 +32,7 @@ RUN_SCRIPT := $(shell mkdir -p 'bin/lib')
 all: $(BUILDDIR)/realizeCascades $(LIBDIROUT)/rootUtil.o $(LIBDIROUT)/edepmath.o $(LIBDIROUT)/cascadeProd.o $(LIBDIROUT)/isotope_info.o $(LIBDIROUT)/weisskopf.o $(LIBDIROUT)/lindhard.o $(LIBDIROUT)/libncap.so
 
 $(LIBDIROUT)/isotope_info.o: $(SRCDIR)/isotope_info.c $(INCDIR)/isotope_info.h 
-	$(CXX) -fPIC -c $(CFLAGS) $(INCFLAG) $(SRCDIR)/isotope_info.c  $(LIBFLAG) 
+	$(CXX) -fPIC -c $(CFLAGS) $(INCFLAG) $(SRCDIR)/isotope_info.c  `root-config --cflags` $(LIBFLAG) 
 	mv isotope_info.o $(LIBDIROUT)/
 
 $(LIBDIROUT)/rootUtil.o: $(SRCDIR)/rootUtil.c $(INCDIR)/rootUtil.h 
@@ -40,19 +40,19 @@ $(LIBDIROUT)/rootUtil.o: $(SRCDIR)/rootUtil.c $(INCDIR)/rootUtil.h
 	mv rootUtil.o $(LIBDIROUT)/
 
 $(LIBDIROUT)/edepmath.o: $(SRCDIR)/edepmath.c $(INCDIR)/edepmath.h 
-	$(CXX) -fPIC -c $(CFLAGS) $(INCFLAG) $(SRCDIR)/edepmath.c  $(LIBFLAG) 
+	$(CXX) -fPIC -c $(CFLAGS) $(INCFLAG) $(SRCDIR)/edepmath.c `root-config --cflags` $(LIBFLAG) 
 	mv edepmath.o $(LIBDIROUT)/
 
 $(LIBDIROUT)/weisskopf.o: $(SRCDIR)/weisskopf.c $(INCDIR)/weisskopf.h 
-	$(CXX) -fPIC -c $(CFLAGS) $(INCFLAG) $(SRCDIR)/weisskopf.c  $(LIBFLAG) 
+	$(CXX) -fPIC -c $(CFLAGS) $(INCFLAG) $(SRCDIR)/weisskopf.c  `root-config --cflags` $(LIBFLAG) 
 	mv weisskopf.o $(LIBDIROUT)/
 
 $(LIBDIROUT)/lindhard.o: $(SRCDIR)/lindhard.c $(INCDIR)/lindhard.h 
-	$(CXX) -fPIC -c $(CFLAGS) $(INCFLAG) $(SRCDIR)/lindhard.c  $(LIBFLAG) 
+	$(CXX) -fPIC -c $(CFLAGS) $(INCFLAG) $(SRCDIR)/lindhard.c  `root-config --cflags` $(LIBFLAG) 
 	mv lindhard.o $(LIBDIROUT)/
 
 $(LIBDIROUT)/cascadeProd.o: $(SRCDIR)/cascadeProd.c $(INCDIR)/cascadeProd.h 
-	$(CXX) -fPIC -c $(CFLAGS) $(INCFLAG) $(SRCDIR)/cascadeProd.c  $(LIBFLAG) 
+	$(CXX) -fPIC -c $(CFLAGS) $(INCFLAG) $(SRCDIR)/cascadeProd.c  `root-config --cflags` $(LIBFLAG) 
 	mv cascadeProd.o $(LIBDIROUT)/
 
[…]
