### 2021-12-23 376c0a4
THIS TEST SHOULD FAIL. Test altered to fail because of python script when the file is missing. If this test fails I will re-include the necessary commands for it to succeed.
