+++ .travis.yml
@@ -2,7 +2,7 @@ jobs:
    include:
     - os: osx
       osx_image: xcode12 
-      env:  ROOTDIST="root_v6.24.06.macos-10.15-x86_64-clang120.tar.gz"; CHECKSUM="e38c561bfed333a60b093ce655cdea73"
+      env:  ROOTDIST="root_v6.24.06.macos-10.15-x86_64-clang120.tar.gz"; CHECKSUM="8cc0046b8fbbc9c3cdde2f5ee493ff40"
       language: generic 
       python: 3.7
     - os: linux
