### 2021-12-23 01d2ed0
[ANV] added verbosity to see if its always an n=0 cascade that gives OSX segfault
