+++ tests/test_realizeCascades.sh
@@ -3,7 +3,7 @@
 #for some systems like OSX
 #alias md5sum='md5'
 
-realizeCascades -n 100000 -o "output.root?reproducible=fixedname" -d 1 ../levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt 
+realizeCascades -n 100000 -o "output.root?reproducible=fixedname" -d 1 -v2 ../levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt 
 # Check that file is uncorrupted and has data
 python3 ./isempty.py
 # Check md5sum
