### 2021-12-23 8c4148d
Switch back to old env for 'default' but include new one also Users shouldn't really need to build docs so this will provide a faster env build for testing while keeping one for the docs build.
