+++ nrCascadeSim_env.yml
@@ -3,18 +3,16 @@ channels:
 #  - conda-forge
   - default
 dependencies:
-  - python==3.8.5
+  - python>=3.6
   - anaconda
 #  - ipykernel
 #  - nb_conda_kernels
-  - anaconda::pytest==6.1.1
-  - sphinx_rtd_theme==0.4.3
-  - conda-forge::myst-parser==0.16.1
-  - pip==21.1.1
+  - anaconda::pytest
+  - pip
   - pip:
     - uproot==3.12.0
-    - papermill==2.3.3
-    - numpy==1.19.2
-    - scipy==1.6.2
-    - matplotlib==3.3.2
+    - papermill
+    - numpy
+    - scipy
+    - matplotlib
 #  - root
