### 2021-12-24 d219982
[ANV] debugging the regex stuff for OSX
