### 2021-12-24 98ea954
[ANV] debugging the regex stuff for OSX
