+++ src/cascadeProd.c
@@ -1158,13 +1158,16 @@ cli *readCascadeDistributionFile(int &n, string file,bool &success)
 
   //do some regex matching to parse the cascade info 
   regex_t regex;
+  //string matchfile="^(0+)$";
   string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)\\s+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)\\s+([0-9][0-9]?[0-9]?)\\s+(\\[.*\\])\\s+(\\[.*\\])$";
   //string matchfile="^([0-9]+\\.?[0-9]+e?[+-]?[0-9]+?)\\s+([0-9]+\\.?[0-9]+?|[0-9]+[A-Z][a-z])\\s+([0-9]+)\\s+(\\[.*?\\])\\s+(\\[.*?\\])$";
   cout << matchfile << endl;
   int reti = regcomp(&regex,matchfile.c_str(),REG_EXTENDED);
     char       buffer[100];                                                     
     regerror(reti, &regex, buffer, 100);                                        
-    printf("regexec() failed with '%s'\n", buffer); 
+    printf("regcomp() failed with '%s'\n", buffer); 
+    cout << "DOES IT COMPILE " << reti << endl;
+   
 
   int count=0;
   while(!getline(in,line).eof()){
@@ -1172,6 +1175,8 @@ cli *readCascadeDistributionFile(int &n, string file,bool &success)
     regmatch_t matchptr[6];
     reti = regexec(&regex,line.c_str(),6,matchptr,0);
     cout << "DOES IT MATCH ANYTHIN " << reti << endl;
+    regerror(reti, &regex, buffer, 100);                                        
+    printf("regexec() failed with '%s'\n", buffer); 
     if(!reti){
       
       //print out the match	    
