### 2021-12-26 a8c697b
[ANV] updated declarations  in regexPlayground.cpp
