+++ bin/regexPlayground.cpp
@@ -130,8 +130,9 @@ int main(int argc, char** argv) {
   regex_t regex;
 
   //*****regex match and input file line*******
-  string iline = "0.625846        29Si    29      [4934.39        0]      [0.84   100000000000000.0]";
-  string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)\\s+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)\\s+([0-9][0-9]?[0-9]?)\\s+(\\[.*\\])\\s+(\\[.*\\])$";
+  string iline = "0.038974	29Si	29	[4934.39	1273.37	0]	[0.84	291.0	100000000000000.0]";
+  //string iline = "0.625846        29Si    29      [4934.39        0]      [0.84   100000000000000.0]";
+  string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)[ \\t\\r\\n\\v\\f]+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)[ \\t\\r\\n\\v\\f]+([0-9][0-9]?[0-9]?)[ \\t\\r\\n\\v\\f]+(\\[.*\\])[ \\t\\r\\n\\v\\f]+(\\[.*\\])$";
 
  
   //compile the regex 
