### 2021-12-26 bbf670e
[ANV] trying C++  version of regex
