+++ bin/regexPlayground.cpp
@@ -21,6 +21,9 @@
 #include <vector>
 #include <algorithm>
 
+#include <iterator>
+#include <regex>
+
 //command line options
 #include <getopt.h>
 #include <stdio.h>
@@ -130,8 +133,8 @@ int main(int argc, char** argv) {
   regex_t regex;
 
   //*****regex match and input file line*******
-  //string iline = "0.038974	29Si	29	[4934.39	1273.37	0]	[0.84	291.0	100000000000000.0]";
-  string iline = "0.625846        29Si    29      [4934.39        0]      [0.84   100000000000000.0]";
+  string iline = "0.038974	29Si	29	[4934.39	1273.37	0]	[0.84	291.0	100000000000000.0]";
+  //string iline = "0.625846        29Si    29      [4934.39        0]      [0.84   100000000000000.0]";
   //string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)[ \\t\\r\\n\\v\\f]+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)[ \\t\\r\\n\\v\\f]+([0-9][0-9]?[0-9]?)[ \\t\\r\\n\\v\\f]+(\\[.*\\])[ \\t\\r\\n\\v\\f]+(\\[.*\\])$";
   string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)[ \\t]+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)[ \\t]+([0-9][0-9]?[0-9]?)[ \\t]+(\\[.*\\])[ \\t]+(\\[.*\\])$";
   //string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)\\s+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)\\s+([0-9][0-9]?[0-9]?)\\s+(\\[.*\\])\\s+(\\[.*\\])$";
@@ -160,6 +163,40 @@ int main(int argc, char** argv) {
     printf("regexec() failed with '%s'\n", buffer); 
 
   //******************************************
+  
 
+
+  //*******Try the C++ way****************************
+  // https://en.cppreference.com/w/cpp/regex
+
+ 
+ 
+    std::regex word_regex("^(0+\\.[0-9]?e?[+-]?[0-9]+)\\s+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)\\s+([0-9][0-9]?[0-9]?)\\s+(\\[.*\\])\\s+(\\[.*\\])$");
+    auto iline_begin = 
+        std::sregex_iterator(iline.begin(), iline.end(), word_regex);
+    auto iline_end = std::sregex_iterator();
+ 
+    std::cout << "Found "
+              << std::distance(iline_begin, iline_end)
+              << " matches\n";
+ 
+    for (std::sregex_iterator i = iline_begin; i != iline_end; ++i) {
+        std::smatch match = *i;
+        std::string match_str = match.str();
+        std::cout << "  " << match_str << '\n';
+    }
+
+    std::smatch matches;
+    if(std::regex_search(iline, matches, word_regex)) {
+    std::cout << "Match found\n";
+
+      for (size_t i = 0; i < matches.size(); ++i) {
+          std::cout << i << ": '" << matches[i].str() << "'\n";
+       }
+    } else {
+        std::cout << "Match not found\n";
+    }
+ 
+  //**************************************************
   return 0;
 }
