+++ src/cascadeProd.c
@@ -1158,26 +1158,33 @@ cli *readCascadeDistributionFile(int &n, string file,bool &success)
 
   //do some regex matching to parse the cascade info 
   regex_t regex;
-  //string matchfile="^(0+)$";
-  string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)[ \\t]+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)[ \\t]+([0-9][0-9]?[0-9]?)[ \\t]+(\\[.*\\])[ \\t]+(\\[.*\\])$";
+  string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)[[:space:]]+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)[[:space:]]+([0-9][0-9]?[0-9]?)[[:space:]]+(\\[.*\\])[[:space:]]+(\\[.*\\])$";
+  //string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)[ \\t]+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)[ \\t]+([0-9][0-9]?[0-9]?)[ \\t]+(\\[.*\\])[ \\t]+(\\[.*\\])$";
   //string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)[ \\t\\r\\n\\v\\f]+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)[ \\t\\r\\n\\v\\f]+([0-9][0-9]?[0-9]?)[ \\t\\r\\n\\v\\f]+(\\[.*\\])[ \\t\\r\\n\\v\\f]+(\\[.*\\])$";
   //string matchfile="^([0-9]+\\.?[0-9]+e?[+-]?[0-9]+?)\\s+([0-9]+\\.?[0-9]+?|[0-9]+[A-Z][a-z])\\s+([0-9]+)\\s+(\\[.*?\\])\\s+(\\[.*?\\])$";
-  cout << matchfile << endl;
   int reti = regcomp(&regex,matchfile.c_str(),REG_EXTENDED);
-    char       buffer[100];                                                     
-    regerror(reti, &regex, buffer, 100);                                        
-    printf("regcomp() failed with '%s'\n", buffer); 
-    cout << "DOES IT COMPILE " << reti << endl;
+
+  //for diagnostics
+  //char       buffer[100];                                                     
+  //regerror(reti, &regex, buffer, 100);  
+  //if(reti==0)
+  //  printf("regcomp() success!\n"); 
+  //else
+  //  printf("regcomp() failed with '%s'\n", buffer); 
    
 
   int count=0;
   while(!getline(in,line).eof()){
-    cout << line << endl;
     regmatch_t matchptr[6];
     reti = regexec(&regex,line.c_str(),6,matchptr,0);
-    cout << "DOES IT MATCH ANYTHIN " << reti << endl;
-    regerror(reti, &regex, buffer, 100);                                        
-    printf("regexec() failed with '%s'\n", buffer); 
+    
+    //for diagnostics
+    //regerror(reti, &regex, buffer, 100);  
+    //if(reti==0)
+    //  printf("regexec() success!\n"); 
+    //else
+    //  printf("regexec() failed with '%s'\n", buffer); 
+
     if(!reti){
       
       //print out the match	    
