+++ .travis.yml
@@ -22,12 +22,12 @@ jobs:
       python: 3.7
     - os: linux
       dist: bionic
-      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu18-x86_64-gcc7.5.tar.gz"; CHECKSUM="e38c561bfed333a60b093ce655cdea73"
+      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu18-x86_64-gcc7.5.tar.gz"; CHECKSUM="ca4acd2e09d37c8f9583c1ebd4260dd5"
       language: python
       python: 3.7
     - os: linux
       dist: focal 
-      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu20-x86_64-gcc9.3.tar.gz"; CHECKSUM="6119a78ef8c534b6bcdd497d796b89b9"
+      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu20-x86_64-gcc9.3.tar.gz"; CHECKSUM="e7afa5796a8e09b0ce6fb517ebd2b147"
       language: python
       python: 3.7
 
