### 2021-12-27 d87f00b
Slightly edit Makefile so that user can supply CFLAGS entries.
