+++ src/cascadeProd.c
@@ -1158,7 +1158,8 @@ cli *readCascadeDistributionFile(int &n, string file,bool &success)
 
   //do some regex matching to parse the cascade info 
   regex_t regex;
-  string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)[[:space:]]+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)[[:space:]]+([0-9][0-9]?[0-9]?)[[:space:]]+(\\[.*\\])[[:space:]]+(\\[.*\\])$";
+  string matchfile="^(-?[0-9]+([.][0-9]+)?(e[+-]?[0-9]+)?)[[:space:]]+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)[[:space:]]+([0-9][0-9]?[0-9]?)[[:space:]]+(\\[.*\\])[[:space:]]+(\\[.*\\])$";
+  //string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)[[:space:]]+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)[[:space:]]+([0-9][0-9]?[0-9]?)[[:space:]]+(\\[.*\\])[[:space:]]+(\\[.*\\])$";
   //string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)[ \\t]+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)[ \\t]+([0-9][0-9]?[0-9]?)[ \\t]+(\\[.*\\])[ \\t]+(\\[.*\\])$";
   //string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)[ \\t\\r\\n\\v\\f]+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)[ \\t\\r\\n\\v\\f]+([0-9][0-9]?[0-9]?)[ \\t\\r\\n\\v\\f]+(\\[.*\\])[ \\t\\r\\n\\v\\f]+(\\[.*\\])$";
   //string matchfile="^([0-9]+\\.?[0-9]+e?[+-]?[0-9]+?)\\s+([0-9]+\\.?[0-9]+?|[0-9]+[A-Z][a-z])\\s+([0-9]+)\\s+(\\[.*?\\])\\s+(\\[.*?\\])$";
@@ -1175,8 +1176,8 @@ cli *readCascadeDistributionFile(int &n, string file,bool &success)
 
   int count=0;
   while(!getline(in,line).eof()){
-    regmatch_t matchptr[6];
-    reti = regexec(&regex,line.c_str(),6,matchptr,0);
+    regmatch_t matchptr[8];
+    reti = regexec(&regex,line.c_str(),8,matchptr,0);
     
     //for diagnostics
     //regerror(reti, &regex, buffer, 100);  
@@ -1196,19 +1197,27 @@ cli *readCascadeDistributionFile(int &n, string file,bool &success)
 
       //cout << "For this match: " << endl;
       int num;
-      for(int i=1;i<6;i++){
-        //Ecout << line.substr(matchptr[i].rm_so,matchptr[i].rm_eo-matchptr[i].rm_so) << endl;
-        string component = line.substr(matchptr[i].rm_so,matchptr[i].rm_eo-matchptr[i].rm_so);
+      for(int i=1;i<8;i++){
+        //cout << matchptr[i].rm_so << endl;
+        string component;
+        if(i!=2&&i!=3){ //matches for 2 & 3 are optional, they  are only  valid pointers if scientific notation
[…]
