+++ bin/regexPlayground.cpp
@@ -134,11 +134,12 @@ int main(int argc, char** argv) {
   regex_t regex;
 
   //*****regex match and input file line*******
-  string iline = "0.038974	29Si	29	[4934.39	1273.37	0]	[0.84	291.0	100000000000000.0]";
+  //string iline = "0.038974	29Si	29	[4934.39	1273.37	0]	[0.84	291.0	100000000000000.0]";
+  string iline = "9.038974e-5	29Si	29	[4934.39	1273.37	0]	[0.84	291.0	100000000000000.0]";
   //string iline = "0.625846        29Si    29      [4934.39        0]      [0.84   100000000000000.0]";
   //string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)[ \\t\\r\\n\\v\\f]+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)[ \\t\\r\\n\\v\\f]+([0-9][0-9]?[0-9]?)[ \\t\\r\\n\\v\\f]+(\\[.*\\])[ \\t\\r\\n\\v\\f]+(\\[.*\\])$";
   //string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)[ \\t]+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)[ \\t]+([0-9][0-9]?[0-9]?)[ \\t]+(\\[.*\\])[ \\t]+(\\[.*\\])$";
-  string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)[[:space:]]+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)[[:space:]]+([0-9][0-9]?[0-9]?)[[:space:]]+(\\[.*\\])[[:space:]]+(\\[.*\\])$";
+  string matchfile="^(-?[0-9]+([.][0-9]+)?(e[+-]?[0-9]+)?)[[:space:]]+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)[[:space:]]+([0-9][0-9]?[0-9]?)[[:space:]]+(\\[.*\\])[[:space:]]+(\\[.*\\])$";
   //string matchfile="^(0+\\.[0-9]?e?[+-]?[0-9]+)\\s+([0-9]+\\.?[0-9]*|[0-9][0-9]?[0-9]?[A-Z][a-z]?)\\s+([0-9][0-9]?[0-9]?)\\s+(\\[.*\\])\\s+(\\[.*\\])$";
 
  
@@ -154,8 +155,8 @@ int main(int argc, char** argv) {
 
 
   //match the input line
-  regmatch_t input_matchptr[6];
-  reti = regexec(&regex,iline.c_str(),6,input_matchptr,0);
+  regmatch_t input_matchptr[8];
+  reti = regexec(&regex,iline.c_str(),8,input_matchptr,0);
   
   regerror(reti, &regex, buffer, 100);  
                     
