### 2021-12-27 65f29f8
Add the `-shared` flag so that program can be compiled with lcov.
