+++ .travis.yml
@@ -3,37 +3,37 @@ jobs:
     # see mac images: https://docs.travis-ci.com/user/reference/osx/
     - os: osx
       osx_image: xcode11.4 
-      env:  ROOTDIST="root_v6.20.00.macosx64-10.15-clang110.tar.gz"; CHECKSUM="8cc0046b8fbbc9c3cdde2f5ee493ff40"
+      env:  ROOTDIST="root_v6.20.00.macosx64-10.15-clang110.tar.gz"; CHECKSUM="5ec00b6c19662ecb44a430afc2326838"
       language: generic 
       python: 3.7
     - os: osx
       osx_image: xcode12 
-      env:  ROOTDIST="root_v6.24.06.macos-10.15-x86_64-clang120.tar.gz"; CHECKSUM="8cc0046b8fbbc9c3cdde2f5ee493ff40"
+      env:  ROOTDIST="root_v6.24.06.macos-10.15-x86_64-clang120.tar.gz"; CHECKSUM="a668732758574cbc90e143b66f22b860"
       language: generic 
       python: 3.7
     - os: linux
       dist: xenial
-      env:  ROOTDIST="root_v6.20.00.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"; CHECKSUM="68857ef79d40213f71afdc33b3012d05"
+      env:  ROOTDIST="root_v6.20.00.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"; CHECKSUM="3a5e00bcb6ebc0e8c76f6877249dfc85"
       language: python
       python: 3.7
     - os: linux
       dist: bionic
-      env:  ROOTDIST="root_v6.20.00.Linux-ubuntu18-x86_64-gcc7.4.tar.gz"; CHECKSUM="68857ef79d40213f71afdc33b3012d05"
+      env:  ROOTDIST="root_v6.20.00.Linux-ubuntu18-x86_64-gcc7.4.tar.gz"; CHECKSUM="3a5e00bcb6ebc0e8c76f6877249dfc85"
       language: python
       python: 3.7
     - os: linux
       dist: xenial
-      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"; CHECKSUM="ca4acd2e09d37c8f9583c1ebd4260dd5"
+      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"; CHECKSUM="e38c561bfed333a60b093ce655cdea73"
       language: python
       python: 3.7
     - os: linux
       dist: bionic
-      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu18-x86_64-gcc7.5.tar.gz"; CHECKSUM="ca4acd2e09d37c8f9583c1ebd4260dd5"
+      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu18-x86_64-gcc7.5.tar.gz"; CHECKSUM="e38c561bfed333a60b093ce655cdea73"
       language: python
       python: 3.7
     - os: linux
       dist: focal 
-      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu20-x86_64-gcc9.3.tar.gz"; CHECKSUM="e7afa5796a8e09b0ce6fb517ebd2b147"
+      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu20-x86_64-gcc9.3.tar.gz"; CHECKSUM="6119a78ef8c534b6bcdd497d796b89b9"
       language: python
       python: 3.7
 
