+++ docs/source/07_Release_History.rst
@@ -2,6 +2,7 @@
 Release history
 ===============
 
+* 27.12.2021: Release of `v1.3.0 <https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.3.0>`_
 * 15.12.2021: Release of `v1.2.4 <https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.2.4>`_
 * 19.11.2021: Release of `v1.2.3 <https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.2.3>`_
 * 13.11.2021: Release of `v1.2.0 <https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.2.0>`_
