### 2021-12-27 347641e
[ANV] updated formatting of RELEASENOTES.md
