+++ RELEASENOTES.md
@@ -1,4 +1,6 @@
-## Future Releases (v1.3.0) Date 21.12.21
+## Future Releases (v1.3.1) Date XX.XX.XX 
+
+## Release (v1.3.0) Date 21.12.27
 
 * PR #54 Move MersenneTwister to built-in C++ function. (Issue #27)
 	* The software now requires C++ 11 or newer.
