### 2021-12-27 d55a0fc
Fix issue in Makefile and try adding lcov in.
