### 2021-12-27 #3993 @villaa
@altavir @matthewfeickert @lucydot @gerudo7  -- the version `v1.3.0` has now been released. This is the best version for your further reviews and comments. There are some issues left from the pre-review but they were both explicitly said to be not required for the review (adding code coverage and using doxygen to auto-generate docs).

We have upgraded so that OSX 10.14, 10.15 are now working on Travis-CI.

We will work on the remaining issues concurrently with the other issues from the review and update here for any changes. I invite you to please give feedback for your review and thank you for agreeing to participate.
