### 2021-12-28 41f0462
Remove CFLAGS from command line as they're now already included.
