+++ .travis.yml
@@ -81,7 +81,7 @@ script:
   - echo `root-config --cflags --glibs`
 
   #make repository code (my code)
-  - make CFLAGS="--coverage"
+  - make
   - sudo make install
 
   # give self permission to run script
