### 2021-12-28 e500621
First attempt at getting travis to send data to codecov.
