### 2021-12-28 a35f9f8
Install lcov to use it.
