### 2021-12-28 0ec4a24
And switch out of lcov directory when done `make`ing
