### 2021-12-28 cbbc03e
Try saving coverage info before make clean Getting nothing for some reason, but it works locally for me?
