### 2021-12-28 cb5563b
Think I have the make command working again now. Hypothetically should also produce coverage info
