### 2021-12-28 82f6984
Switch back to main Makefile now that it's fixed. Also added coverage output to .gitignore.
