### 2021-12-28 7d8ed34
Try using coverage for OSX compatibility apparently --coverage is necessary for OSX.
