+++ docs/source/03a_Doxy_Libraries.rst
@@ -1,0 +1,14 @@
+========================================
+3. The libraries of *nrCascadeSim*
+========================================
+
+The functionality of the executables of *nrCascadeSim* such as `realizeCascades` is constructed
+from a set of library functions whose code is stored in the `libncap.so` file and whose prototypes
+are stored in several header files. Below we list the different header files with the internal
+functions, data structures, and their uses. 
+
+Docs
+====
+
+.. doxygenstruct:: cat
+   :members:
