+++ .readthedocs.yaml
@@ -18,7 +18,7 @@ build:
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
    configuration: docs/source/conf.py
-   fail_on_warning: true 
+   fail_on_warning: false 
 
 
 # If using Sphinx, optionally build your docs in additional formats such as PDF
