### 2021-12-29 30408f2
Working on making examples compile. All examples are present in some form and are in makefile. Updated compiling examples in the docs.
