### 2021-12-29 villano-lab/nrCascadeSim#64 @villaa
Migrate to CMake
I am submitting a **feature request**.

**The feature I am requesting is for:**
make/CI tests

**I am requesting:**
Migrating to CMake--a more modern build system; this will also allow testing to be simplified.

**Ideas for implementation:**
see the "obscura" code by temken on github.
