### 2021-12-29 villano-lab/nrCascadeSim#64 @gerudo7
Link to the aforementioned Obscura code: https://github.com/temken/obscura/blob/master/.github/workflows/main.yml
Link to sample results for Obscura's testing through GitHub: https://github.com/temken/obscura/runs/4345220111?check_suite_focus=true
Link to the Obscura repository main page: https://github.com/temken/obscura
