### 2021-12-29 villano-lab/nrCascadeSim#64 @villaa
Solving this issue will obviate issue #26 because CMake is a more modern replacement for automake tools.
