### 2021-12-29 ad6e738
Finish getting examples compiling and running. Ignore binaries and other common outputs. Include tests for new binaries.
