+++ tests/README.md
@@ -35,3 +35,5 @@ C++ tests are supplied in the `bin` subdirectory
 to test the coverage offered by examples provided on readthedocs.
 They can be built from the top-level directory using the command `make tests`;
 they are not normal usecases, so they are not made by default.
+They should be runnable from this (the `tests`) directory 
+(you will need to supply the path to the binaries).
