### 2021-12-29 05be758
Test that regexPlayground runs. Nothing fancy but it's a start.
