### 2021-12-29 026e36d
More tests!Test multiple elements and test help functions for main binaries.
