+++ .travis.yml
@@ -14,7 +14,7 @@ jobs:
       python: 3.7
     - os: osx
       osx_image: xcode13.1 
-      env:  ROOTDIST="https://root.cern/download/root_v6.24.06.macos-11.5-x86_64-clang120.tar.gz"; CHECKSUM="a668732758574cbc90e143b66f22b860"
+      env:  ROOTDIST="root_v6.24.06.macos-11.5-x86_64-clang120.tar.gz"; CHECKSUM="a668732758574cbc90e143b66f22b860"
       language: generic 
       python: 3.7
     - os: linux
