### 2021-12-29 70cd39d
[ANV] added some doxygen comments to cascadeProd.h
