### 2021-12-29 932252f
[ANV] added some doxygen comments to cascadeProd.h
