+++ inc/cascadeProd.h
@@ -115,59 +115,115 @@ void freecriarray(int n,cri *cascade_data);
   A function to simulate n cascade realizations for a germanium isotope 
 */
 cri *geCascade(int n,int cid, double Sn, int nlev, double *Elev, double *taus, double A, mt19937 *mtrand);
-//return the Energy after the mid-stop kick. 
+/**
+  germanium func. to return the Energy after the mid-stop decay 
+*/
 double geDecay(double v, double M, double Egam, mt19937 *rand);
-//return the velocity at a random stopping time
+/**
+  germanium func. to return the velocity at a random stopping time
+*/
 double *geStop(double E, double M, double tau, mt19937 *rand);
-//constant stopping power stopping. return distance after time t for initial energy E and mass M
+/**
+  germanium func. return distance after time t for initial energy E and mass M
+*/
 double rgeS2(double E, double M, double t);
-//constant stopping power stopping. return velocity after time t for initial energy E and mass M
+/**
+  germanium func. constant stopping power stopping. return velocity after time t for initial energy E and mass M
+*/
 double vgeS2(double E, double M, double t);
+/**
+  germanium func. constant stopping power stopping. return velocity as function of time (x) for initial energy E and mass M (par)
+*/
 double vgeS2func(double *x,double *par);
 
 //******silicon*****************
 //do a generalized multi-step cascade (for now just print a table and do one event)
 //eventually: can do n events, put in a yield model function, generalize to other elements 
+/**
+  A function to simulate n cascade realizations for a silicon isotope 
+*/
 cri *siCascade(int n,int cid, double Sn, int nlev, double *Elev, double *taus, double A, mt19937 *mtrand);
-//return the Energy after the mid-stop kick. 
+/**
+  silicon func. to return the Energy after the mid-stop decay 
+*/
 double siDecay(double v, double M, double Egam, mt19937 *rand);
-//return the velocity at a random stopping time
+/**
+  silicon func. to return the velocity at a random stopping time
+*/
 double *siStop(double E, double M, double tau, mt19937 *rand);
-//constant stopping power stopping. return distance after time t for initial energy E and mass M
+/**
+  silicon func. return distance after time t for initial energy E and mass M
+*/
 double rsiS2(double E, double M, double t);
-//constant stopping power stopping. return velocity after time t for initial energy E and mass M
+/**
+  silicon func. constant stopping power stopping. return velocity after time t for initial energy E and mass M
+*/
 double vsiS2(double E, double M, double t);
+/**
+  silicon func. constant stopping power stopping. return velocity as function of time (x) for initial energy E and mass M (par)
+*/
 double vsiS2func(double *x,double *par);
 
 //******argon*****************
 //do a generalized multi-step cascade (for now just print a table and do one event)
 //eventually: can do n events, put in a yield model function, generalize to other elements 
+/**
+  A function to simulate n cascade realizations for a argon isotope 
+*/
 cri *arCascade(int n,int cid, double Sn, int nlev, double *Elev, double *taus, double A, mt19937 *mtrand);
-//return the Energy after the mid-stop kick. 
+/**
+  argon func. to return the Energy after the mid-stop decay 
+*/
 double arDecay(double v, double M, double Egam, mt19937 *rand);
-//return the velocity at a random stopping time
+/**
+  argon func. to return the velocity at a random stopping time
+*/
 double *arStop(double E, double M, double tau, mt19937 *rand);
-//constant stopping power stopping. return distance after time t for initial energy E and mass M
+/**
+  argon func. return distance after time t for initial energy E and mass M
+*/
 double rarS2(double E, double M, double t);
-//constant stopping power stopping. return velocity after time t for initial energy E and mass M
+/**
[…]
