+++ .travis.yml
@@ -119,6 +119,9 @@ script:
   #- source activate test-environment
   #- conda list papermill
   - cd ../example-usecase
+  - realizeCascades -n 10000 -o data/Gefile.root ../levelfiles/allge_ngam_WSlow.txt
+  - realizeCascades -n 10000 -o data/Nefile.root ../levelfiles/20ne_ngam.txt
+  - realizeCascades -n 10000 -o data/Arfile.root ../levelfiles/40ar_ngam.txt
   - travis_wait 50 py.test
 
   #Save coverage info
