### 2021-12-30 97dabdb
Update codecov upload to new method And only upload from one build, not 8 times
