### 2021-12-30 677fa1f
Try 3. Also ignoring the files the script downloads.
