### 2021-12-30 8bdba3f
Change over the badge target branch to master in preparation for release.
