### 2021-12-30 #3993 @villaa
@lucydot @matthewfeickert @altavir --  one last update. We have released  [`v1.3.1` ](https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.3.1)and all issues from the previous review and pre-review are  addressed. We have test coverage of around 80%.

If you have not started the review this is the best version to use. If you started on `v1.3.0` the only differences are upgraded documentation and code test coverage.
