### 2022-01-01 #3993 @lucydot
Thanks for the update @villaa and Happy New Year ✨

I'll update the version number now. Usually we update the version number at the end of the review process - in this case, I'll update now to avoid any confusion as to which version is under review.
