### 2022-01-02 a55ddf9
Working on translation of Makefile to CMake Currently trying to figure out how to do libncap
