+++ CMakeLists.txt
@@ -1,0 +1,115 @@
+###############################################################
+#
+#
+# Makefile for nrCascadeSim
+###############################################################
+cmake_minimum_required(VERSION 3.10.2) #Just setting this to my version since I know that works.
+
+# everybody here needs root
+SET(rootFlags `root-config --cflags`)
+#trick for getting the git version in the code
+SET(gitVersion "$(shell sh -c 'git describe --abrev=4 --always')")
+# other variables
+#SET(topDir $(shell pwd))    #obsolete - we will mark this using CMake
+#SET(buildDir $(topDir)/bin) #obsolete - we will mark this using CMake
+SET(incDirOut include)
+SET(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR}/lib)
+SET(testBin tests/bin)
+SET(incDir inc/)
+SET(srcDir src/)
+SET(libFLAG "-L. -L$(CMAKE_LIBRARY_OUTPUT_DIRECTORY)")
+SET(cFlags "-D__GIT_VERSION=\'$(GIT_VERSION)'")
+# Make necessary bin/lib directory
+# SET(runScript "$(shell mkdir -p 'bin/lib')") #hopefully don't need this anymore
+# Send files to the right places
+
+
+add_library(isotope_info.o ${srcDir}/isotope_info.c ${incDir}/isotope_info.h)
+add_library(rootUtil.o ${srcDir}/rootUtil.c ${incDir}/rootUtil.h)
+add_library(edepmath.o ${srcDir}/edepmath.c ${incDir}/edepmath.h)
+add_library(weisskopf.o ${srcDir}/weisskopf.c ${incDir}/weisskopf.h)
+add_library(lindhard.o ${srcDir}/lindhard.c ${incDir}/lindhard.h)
+add_library(cascadeProd.o ${srcDir}/cascadeProd.c ${incDir}/cascadeProd.h)
+
+add_executable(regexPlayground regexPlayground.cpp)
+
+link_libraries(${CMAKE_LIBRARY_OUTPUT_DIRECTORY}/weisskopf.o ${CMAKE_LIBRARY_OUTPUT_DIRECTORY}/lindhard.o $(CMAKE_LIBRARY_OUTPUT_DIRECTORY)/cascadeProd.o $(CMAKE_LIBRARY_OUTPUT_DIRECTORY)/edepmath.o $(CMAKE_LIBRARY_OUTPUT_DIRECTORY)/rootUtil.o)
+add_library(libncap)
+
+link_libraries(${CMAKE_LIBRARY_OUTPUT_DIRECTORY}/libncap.so) #link libncap to all later executables.
+add_executable(realizeCascades realizeCascades.cpp)
+add_executable(fetchYieldModel ${testBin}/fetchYieldModel.cpp)
+add_executable(printIsotopeInfo ${testBin}/printIsotopeInfo.cpp)
+add_executable(readLevelfile ${testBin}/readLevelfile.cpp)
+add_executable(realizeAndSave ${testBin}/realizeAndSave.cpp)
+
[…]
