### 2022-01-03 villano-lab/nrCascadeSim#75 @altavir
[JOSS Review]: No comparison with other algorithms and no data-tests
The provided paper includes some considerations about how the computations should be made, but there is no comparison with other existing tools. Also test suite do not include data-based tests. That means that subsequent versions could also change the result of the simulation and it could not be tracked.

Changing of physics results in different code versions is a major problem of simulation software in particle physics so I think there should be an effort to mitigate that. As well as a chapter in the paper about the comparison.
