### 2022-01-03 #3993 @altavir
Happy new year to everyone. I've finally found some time to review the code and the paper and open [some issues](https://github.com/villano-lab/nrCascadeSim/issues/created_by/altavir). Not all of them are critical though. What I think is the most important is [this one](https://github.com/villano-lab/nrCascadeSim/issues/75). The lack of comparison of results with other tools as well as regression tests could be a real problem in the future.

I also failed to build the project on WSL even with CERN ROOT installed ( I have no clue why it is needed anyway), so I can't comment on the program working (I will try later to run it again), but even from looking into the code there seems to some things that could be improved. I am not a C++ expert though.
