### 2022-01-03 #3993 @villaa
@altavir Happy New Year! Thank you very much for your review,  I see the issues on our issue tracker and the team will address them and comment here when they're finished.

Do you know what style of Linux WSL uses? We haven't tested on Windows but could try to match the system.

@gerudo7 do you run WSL?
