### 2022-01-03 #3993 @gerudo7
> @gerudo7 do you run WSL?
Yes! I've been doing all my work with nrCascadeSim in WSL. I know for certain that it works for me with root version 6.22/00 installed via conda, but I just tried it with the latest precompiled binary for version 6.24 and nrCascadeSim won't make. I'm guessing that the problem is with using the precompiled binaries in WSL, but I'm still testing that theory. (I will work on adding whatever I find to the docs as well.)
