### 2022-01-03 #3993 @altavir
I am using Root 6.24.0 installed via nix-env. I think it is better to continue the discussion in your repository, You can open an issue and we can discuss it there.
