### 2022-01-03 villano-lab/nrCascadeSim#75 @villaa
@gerudo7  I think if we solve our previous issues about checking all input files and checking the resulting distributions in the Python testing, that will mostly satisfy this issue.

Also for testing the Python we can probably just make output data files to check distributions instead of checksums on the .pngs.
