### 2022-01-03 9c45675
First attempt at running checksum off of `checksums` text file.
