+++ tests/test_realizeCascades.sh
@@ -12,6 +12,10 @@ then
         exit $error_code
 fi
 # Check md5sum
-echo $CHECKSUM
+export SUMLINE=$BUILDLINE
+declare -i SUMLINE
+SUMLINE+=30
+CHECKSUM=$(sed -n "$SUMLINE"p ../.travis/checksums | sed 's/^ *//g') #There may be an elegant solution here, but I don't know it.
+echo Checksum from file: $CHECKSUM, found on line $SUMLINE
 md5sum output.root
 md5sum output.root | awk -v cs=$CHECKSUM '{if ($1==cs){system("rm output.root"); print"Checksum passed!"} else {system("rm output.root"); exit 1}}'
