+++ .travis/.travis.yml
@@ -4,76 +4,50 @@ jobs:
     # more recent build envs:  https://changelog.travis-ci.com/
     - os: osx
       osx_image: xcode11.4 
-      env:  ROOTDIST="root_v6.20.00.macosx64-10.15-clang110.tar.gz"; CHECKSUM="5ec00b6c19662ecb44a430afc2326838"
+      env:  ROOTDIST="root_v6.20.00.macosx64-10.15-clang110.tar.gz"; BUILDLINE=2
       language: generic 
       python: 3.7
     - os: osx
       osx_image: xcode12 
-      env:  ROOTDIST="root_v6.24.06.macos-10.15-x86_64-clang120.tar.gz"; CHECKSUM="a668732758574cbc90e143b66f22b860"
+      env:  ROOTDIST="root_v6.24.06.macos-10.15-x86_64-clang120.tar.gz"; BUILDLINE=461
       language: generic 
       python: 3.7
     - os: osx
       osx_image: xcode13.1 
-      env:  ROOTDIST="root_v6.24.06.macos-11.5-x86_64-clang120.tar.gz"; CHECKSUM="a668732758574cbc90e143b66f22b860"
+      env:  ROOTDIST="root_v6.24.06.macos-11.5-x86_64-clang120.tar.gz"; BUILDLINE=527
       language: generic 
       python: 3.7
     - os: linux
       dist: xenial
-      env:  ROOTDIST="root_v6.20.00.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"; CHECKSUM="3a5e00bcb6ebc0e8c76f6877249dfc85"
+      env:  ROOTDIST="root_v6.20.00.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"; BUILDLINE=198
       language: python
       python: 3.7
     - os: linux
       dist: bionic
-      env:  ROOTDIST="root_v6.20.00.Linux-ubuntu18-x86_64-gcc7.4.tar.gz"; CHECKSUM="3a5e00bcb6ebc0e8c76f6877249dfc85"
+      env:  ROOTDIST="root_v6.20.00.Linux-ubuntu18-x86_64-gcc7.4.tar.gz"; BUILDLINE=264
       language: python
       python: 3.7
     - os: linux
       dist: xenial
-      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"; CHECKSUM="e38c561bfed333a60b093ce655cdea73"; UPLOAD_COVERAGE="true"
+      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"; BUILDLINE=593
       language: python
       python: 3.7
     - os: linux
       dist: bionic
-      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu18-x86_64-gcc7.5.tar.gz"; CHECKSUM="e38c561bfed333a60b093ce655cdea73"
+      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu18-x86_64-gcc7.5.tar.gz"; BUILDLINE=659
       language: python
       python: 3.7
     - os: linux
       dist: focal 
-      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu20-x86_64-gcc9.3.tar.gz"; CHECKSUM="6119a78ef8c534b6bcdd497d796b89b9"
+      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu20-x86_64-gcc9.3.tar.gz"; BUILDLINE=725
       language: python
       python: 3.7
-
-# osx deps
-#before_install:
[…]
