### 2022-01-03 68a4477
Run script for checksum retrieval on travis.
