### 2022-01-03 494e05c
Have travis give self permission to run checksum script.
