### 2022-01-03 3442cba
Fix line spacing reference in the test and pretty up some formatting.
