+++ levelfiles/v3_cross_sections_and_abundance.txt
@@ -1,9 +1,9 @@
 #Isotope	thermalxn(b)	natabundance	xn*ab		xn*ab(total)
-28si		0.169		0.92223		0.155857	0.161432	
-29si		0.119		0.04685		0.00557515	0.161432	
-30si		0.107		0.03092		0.00330844	0.161432	
-70ge		3.7		0.2057		0.76109		3.11411
-72ge		1.2		0.2745		0.3294		3.11411
-73ge		22.0		0.0775		1.705		3.11411
-74ge		0.82		0.365		0.2993		3.11411
-76ge		0.25		0.0773		0.019325	3.11411
+28si		0.169		    0.92223		    0.155857	0.161432	
+29si		0.119		    0.04685		    0.00557515	0.161432	
+30si		0.107		    0.03092		    0.00330844	0.161432	
+70ge		3.7		        0.2057		    0.76109		3.11411
+72ge		1.2		        0.2745		    0.3294		3.11411
+73ge		22.0		    0.0775		    1.705		3.11411
+74ge		0.82		    0.365		    0.2993		3.11411
+76ge		0.25		    0.0773		    0.019325	3.11411
