### 2022-01-04 villano-lab/nrCascadeSim#75 @villaa
@gerudo7 for comparison with other algorithms I can try to set up a comparison with Geant4 which is basically the only other code I know of that implements something like the NR recoils after capture.
