### 2022-01-04 a872f70
First try at running checksum verification on multiple files.
