### 2022-01-04 4d776f5
Fix variable calling for travis built-in vars.
