+++ inc/cascadeProd.h
@@ -29,7 +29,7 @@ struct cli {
     */
     double frac;
     /**
-      neutron separation energy  
+      neutron separation energy in **MeV** 
     */
     double Sn;
     /**
@@ -112,27 +112,43 @@ void freecriarray(int n,cri *cascade_data);
 //do a generalized multi-step cascade (for now just print a table and do one event)
 //eventually: can do n events, put in a yield model function, generalize to other elements 
 /**
-  A function to simulate n cascade realizations for a germanium isotope 
+  A function to simulate n cascade realizations for a germanium isotope. `cid` is the cascade id;
+`Sn` is the neutron separation energy in **MeV**; `nlev` is the number of levels of the cascade;
+`Elev` is an array of level energys of size `nlev` in **keV**; `taus` is an array of level
+lifetimes of size `nlev` in **fs**; `A` is the mass number; `mtrand` is the random number
+generator object--in this case `std::mt19937`. The function returns a `cri` struct. 
 */
 cri *geCascade(int n,int cid, double Sn, int nlev, double *Elev, double *taus, double A, mt19937 *mtrand);
 /**
-  germanium func. to return the Energy after the mid-stop decay 
+  germanium func. to return the Energy after the mid-stop decay. `v` is the velocity in units of
+the speed of light (c); `M` is the mass of the slowing object in **GeV**, `Egam` is the energy of
+the emitted gamma ray in **MeV**; and `mtrand` is the random number generator object--in this case
+`std::mt19937`. The function returns the energy in **eV** after the mid-step decay kick. 
 */
 double geDecay(double v, double M, double Egam, mt19937 *rand);
 /**
-  germanium func. to return the velocity at a random stopping time
+  germanium func. to return the velocity at a random stopping time. `E` is the initial atom energy
+in **eV**; `M` is the mass of the slowing object in **GeV**; `tau` is the (random) decay time of the current
+intermediate nuclear level; `mtrand` is the random generator object--in this case `std::mt19937`.
+The function returns the velocity at the random decay time in units of the speed of light (c).
 */
 double *geStop(double E, double M, double tau, mt19937 *rand);
 /**
-  germanium func. return distance after time t for initial energy E and mass M
+  germanium func. constant stopping-power stopping. `E` is the initial energy of the stopping
+atom in **eV**; `M` is the mass of the slowing object in **GeV**; `t` is the time (in **fs**) at
+which to return the distance traveled. Function returns distance in **nm** 
 */
 double rgeS2(double E, double M, double t);
 /**
-  germanium func. constant stopping power stopping. return velocity after time t for initial energy E and mass M
+  germanium func. constant stopping-power stopping. `E` is the initial energy of the stopping
+atom in **eV**; `M` is the mass of the slowing object in **GeV**; `t` is the time (in **fs**) at
+which to return the atoms velocity. Function returns velocity in units of the speed of light (c).
 */
 double vgeS2(double E, double M, double t);
 /**
-  germanium func. constant stopping power stopping. return velocity as function of time (x) for initial energy E and mass M (par)
+  germanium func. constant stopping-power stopping. `x[0]` is the time (in **fs**) at
+which to return the atoms velocity; `par[0]` is the atoms initial energy in **eV**; `par[1] is the
+atoms mass in **GeV**. Function returns velocity in units of the speed of light (c).
 */
 double vgeS2func(double *x,double *par);
 
@@ -140,27 +156,43 @@ double vgeS2func(double *x,double *par);
 //do a generalized multi-step cascade (for now just print a table and do one event)
 //eventually: can do n events, put in a yield model function, generalize to other elements 
 /**
-  A function to simulate n cascade realizations for a silicon isotope 
[…]
