### 2022-01-05 fcaa875
Update instructions, update docs env, and try testing instructions in WSL.
