+++ .travis.yml
@@ -4,71 +4,82 @@ jobs:
     # more recent build envs:  https://changelog.travis-ci.com/
     - os: osx
       osx_image: xcode11.4 
-      env:  ROOTDIST="root_v6.20.00.macosx64-10.15-clang110.tar.gz"; BUILDLINE=2
+      env:  ROOTDIST="root_v6.20.00.macosx64-10.15-clang110.tar.gz"
       language: generic 
       python: 3.7
     - os: osx
       osx_image: xcode12 
-      env:  ROOTDIST="root_v6.24.06.macos-10.15-x86_64-clang120.tar.gz"; BUILDLINE=472
+      env:  ROOTDIST="root_v6.24.06.macos-10.15-x86_64-clang120.tar.gz"
       language: generic 
       python: 3.7
     - os: osx
       osx_image: xcode13.1 
-      env:  ROOTDIST="root_v6.24.06.macos-11.5-x86_64-clang120.tar.gz"; BUILDLINE=539
+      env:  ROOTDIST="root_v6.24.06.macos-11.5-x86_64-clang120.tar.gz"
       language: generic 
       python: 3.7
     - os: linux
       dist: xenial
-      env:  ROOTDIST="root_v6.20.00.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"; BUILDLINE=203
+      env:  ROOTDIST="root_v6.20.00.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"
       language: python
       python: 3.7
     - os: linux
       dist: bionic
-      env:  ROOTDIST="root_v6.20.00.Linux-ubuntu18-x86_64-gcc7.4.tar.gz"; BUILDLINE=270
+      env:  ROOTDIST="root_v6.20.00.Linux-ubuntu18-x86_64-gcc7.4.tar.gz"
       language: python
       python: 3.7
     - os: linux
       dist: xenial
-      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"; BUILDLINE=606
+      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu16-x86_64-gcc5.4.tar.gz"
       language: python
       python: 3.7
     - os: linux
       dist: bionic
-      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu18-x86_64-gcc7.5.tar.gz"; BUILDLINE=673
+      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu18-x86_64-gcc7.5.tar.gz"
       language: python
       python: 3.7
     - os: linux
       dist: focal 
-      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu20-x86_64-gcc9.3.tar.gz"; BUILDLINE=740
+      env:  ROOTDIST="root_v6.24.06.Linux-ubuntu20-x86_64-gcc9.3.tar.gz"
+      language: python
+      python: 3.7
+    - os: windows
       language: python
       python: 3.7
 
 install:
-   - sudo chmod 777 .travis/install.sh
-   - ls -lh .travis/install.sh
-   - source .travis/install.sh
+  - if [ "$TRAVIS_OS_NAME" = "windows" ]; then curl -L -o ubuntu-1804.zip https://aka.ms/wsl-ubuntu-1804; fi
+  - if [ "$TRAVIS_OS_NAME" = "windows" ]; then powershell.exe -NoProfile -Command "Expand-Archive -Path ubuntu-1804.zip -DestinationPath Ubuntu"; fi
+  - if [ "$TRAVIS_OS_NAME" = "windows" ]; then Ubuntu/ubuntu1804.exe install --root; fi
[…]
