### 2022-01-05 2f5f057
Revert .travis.yml as Windows support is not where it needs to be for this test yet.
