### 2022-01-05 4c365bc
[ANV] updated release notes
