### 2022-01-05 0b8e1a1
Use seeds for resolutions, premade root files, and hashlib to ensure consistent output.
