+++ nrCascadeSim_env.yml
@@ -15,4 +15,5 @@ dependencies:
     - numpy
     - scipy
     - matplotlib
+    - hashlib
 #  - root
