+++ nrCascadeSim_withDocs.yml
@@ -17,4 +17,5 @@ dependencies:
     - numpy==1.19.2
     - scipy==1.6.2
     - matplotlib==3.3.2
+    - hashlib
 #  - root
