+++ example-usecase/requirements.txt
@@ -7,4 +7,4 @@ uproot==3.12.0
 papermill
 #Only really needed for testing - user can comment out but doesn't hurt anything
 nbconvert
-hashlib
+#hashlib
