### 2022-01-07 d0baeba
Try testing via txt files instead. Should be universal across OS If not maybe there is an issue in the seeding method?
