+++ .travis/example-usecase_output/Ar_small_Sor.txt
@@ -1,0 +1,9162 @@
+34.418286604006624
+27.381839044304975
+55.59214098563249
+28.504891821234345
+66.2496795595543
+22.328026842175863
+27.594371537065896
+25.31309430170871
+5.722087401459072
+52.694297972144746
+16.51263060496065
+25.56232123998902
+72.61098610774512
+72.19838540824725
+58.2372142140276
+46.01729808176062
+33.779974891187045
+62.91582690199662
+72.05952072493882
+49.356724580083124
+8.503679661486247
+42.693173062351526
+38.382024817450414
+2.8531938510312123
+9.229398532889434
+12.644357580709993
+29.16767314653235
+36.390277321200394
+49.69452451125414
+1.9749419283346237
+58.84013238604683
+24.47810296396187
+26.48966174517189
+67.63459565727283
+34.581342287580355
+66.00700587536743
+36.6551554115844
+2.4641970636259254
+17.26881866406277
+79.43228615143087
+54.12009065333861
+69.58408284792755
+20.18065323455734
+75.548303349375
+2.63640744184049
+5.790018656794151
+29.61767661333299
+44.33845412940972
+20.49919954057371
+77.9672865861052
+14.77640846126495
+7.065242839848283
+21.860756006636976
+52.98485360279438
+30.634357607995863
+80.04247059586181
+71.26705037204424
+11.232486602008569
+40.22987402244978
+23.336678259118656
+29.885663382924072
+32.52014181509287
+39.47570916124977
+30.530582902418814
+58.75278101904509
+45.48279254533964
+37.11881829108765
+17.547714071158595
+40.31764786561375
+61.486059251722715
+60.10938054287858
+72.63488119877702
+9.063280579557844
+42.326297833953014
+33.73746536188992
+27.411790255243233
+14.296765879116798
+8.216491505278585
+15.295718170726008
+0.005874355028772593
+45.28876898401621
+45.49555931740804
+22.802446348548564
+23.39228549521597
+7.691740716868804
+46.17939066687016
+67.94308940168271
+33.22874201846494
+67.57964982969119
+25.62242303077999
+60.19829547296921
+45.559647071008584
+58.80731876139739
+29.885823005501145
+9.268442536763002
+10.688848758194112
+59.2747725435204
+51.027755777775
+56.57883726447251
+35.396969958614996
+43.96215826511779
+5.4447329572897205
+38.411260540528936
+13.830582826182845
+74.3512053334026
+47.84242497287555
+35.9668579589865
+59.1357367039305
+66.35790096451979
+51.23686236192695
+38.458549537008544
+49.935786750779315
+18.08311006835794
+11.282290246726856
+65.37335457365701
+26.80600649188031
+51.41081552680851
+58.9570141043287
+58.17695535634853
+27.03495339913466
+26.426558822584955
+41.22170143368824
+62.344501701593025
+63.30514001860364
+41.85218188769268
+49.91190084761401
+47.31410841396444
+14.703257789047058
+52.677060989987574
+7.072457971303413
+1.3298004396412033
+76.51375138238997
+10.70106754858528
+76.24037388212437
+67.27694597938532
+1.750160636644973
+12.458741623442997
+61.968276611381945
+18.018208112509072
+60.374596092766055
+19.064681199746563
+11.0261361728453
+31.422420063051604
+8.053715705489106
+28.25441102452681
+38.57969038599508
+22.930875114691226
+5.595011966457491
+23.475799160477564
+55.73615502252842
+41.80665298982205
+16.055367179548455
+27.799074918465763
+57.49867861244159
+70.75328089005832
+14.422901595013148
+63.641183500463534
+5.927447729563369
+31.642820432608843
[…]
