+++ .travis/example-usecase_output/Ge_small_Lind.txt
@@ -1,0 +1,5617 @@
+75.79641702469337
+68.3826772721413
+68.65980013529206
+66.85353968133258
+73.28017786784338
+62.78024619613099
+76.19582037551109
+67.88717790374582
+71.46870660310682
+69.58415915647598
+75.25852287726474
+63.580595211279544
+69.34197421569631
+69.13761777777363
+74.16992477532774
+66.76427832182036
+69.83925899799263
+67.50042270425206
+70.5508990517023
+72.34324975185828
+66.76186495467918
+74.20624337642174
+73.40014200580269
+72.07694789180456
+73.39770588976292
+68.14405773695496
+70.00350094367525
+67.3084201718587
+69.52276416955482
+72.1693207672643
+68.11775642927094
+69.09551398865874
+68.13263645259472
+67.60868217126375
+68.18544059640045
+70.36895126349671
+66.7065257915221
+71.1881398337884
+75.91397320662348
+72.87116935957961
+69.77491442961123
+67.46802870178261
+67.93375571085197
+76.02223150491436
+70.57939227139285
+68.29899722398069
+71.04391596046392
+4.431538337970358
+0.25353538973623757
+1.3022985614826548
+0.6333593751067961
+1.2377749333944372
+1.7702550032752056
+1.9646253915505063
+0.6025892562864861
+1.8676478566112364
+2.643551807542238
+1.0822819210819987
+1.0307732891941188
+2.387738007522511
+3.20681348265084
+4.611564107784972
+0.33767821653492236
+1.848716423555235
+0.6659897386797872
+1.7470264959012898
+0.48549989161884904
+1.6078435904583726
+0.39364453498527996
+0.865208975609421
+0.4184124094674842
+0.25110824265373816
+0.7966595891371401
+0.25704288179193197
+2.383211045694151
+2.5297167265921803
+0.39068004089112157
+0.8935730870752784
+0.1631875442154604
+0.09198936792026105
+1.4728475918923822
+2.583711257139603
+0.8513673641785882
+1.2524506840302114
+0.3573969337537897
+1.56257411181314
+0.06881669352298657
+0.6649863172092357
+1.7853989664044208
+0.7396520149812666
+2.366191364377251
+0.8627801350467068
+2.687625012635019
+4.150584690927491
+2.608306104255451
+3.434343089144193
+0.713204679772144
+1.8216586200178857
+1.1616341165098145
+1.6728224870898238
+1.0984160712248126
+1.6919267772485376
+60.43826751456008
+59.69909698410431
+59.96803459496801
+63.0454627840098
+62.66983452568098
+61.969315277466805
+60.72721127483089
+60.53735904766113
+62.25122369096642
+61.027757430354036
+62.45468090659962
+59.30738620817768
+52.57087219326307
+63.58425951096797
+67.2245316158558
+61.689987864240635
+59.97307778116442
+59.858020416965296
+59.91315787998123
+60.345821242161676
+56.73321815300203
+58.65115507104226
+57.129510041310056
+61.07945183247801
+59.350111211187425
+61.86071158097515
+59.7367205589229
+63.41783969638103
+60.96764189619255
+67.23632628659617
+54.2781348426301
+58.23954842811331
+63.14882473449212
+68.30678313319858
+59.50231736771516
+60.42907832031104
+59.573085529216684
+64.51206274653231
+59.379702781903696
+62.446812997892195
+59.27668007678563
+56.255924897648214
+61.28473976326547
+61.885991852462666
+64.39084908214254
+59.94044723391062
+58.33325256306479
+62.074880872904004
+61.05389071548476
+61.18048400983062
+60.058817217083956
+63.969524471859614
+61.46211564023791
+66.329421637543
+63.8132743011022
[…]

[…]
