### 2022-01-11 5ba552a
Remove all malloc references in cascadeProd Still need to change other things over to match. Currently doesn't make.
