+++ inc/cascadeProd.h
@@ -39,11 +39,11 @@ struct cli {
     /**
       array of energy levels in **keV**   
     */
-    double *Elev;
+    vector<double> Elev;
     /**
       half-lives of energy level in **fs**  
     */
-    double *taus;
+    vector<double> taus;
 };
 
 //define a structure for the returned cascade information
@@ -63,27 +63,27 @@ struct cri {
     /**
       recoil energy at beginning of recoil step **eV** 
     */
-    double *E;
+    vector<double> E;
     /**
       energy deposited in step **eV** 
     */
-    double *delE;
+    vector<double> delE;
     /**
       ionization created in each step in n e/h pairs 
     */
-    int *I;
+    vector<int> I;
     /**
       electron-equivalent ionization energy **eV** 
     */
-    double *Ei;
+    vector<double> Ei;
     /**
       time since capture **fs**  
     */
-    double *time;
+    vector<double> time;
     /**
       gamma energy emitted in each step **MeV** 
     */
-    double *Eg;
+    vector<double> Eg;
 };
 
 typedef struct cli cli;
@@ -118,7 +118,7 @@ void freecriarray(int n,cri *cascade_data);
 lifetimes of size `nlev` in **fs**; `A` is the mass number; `mtrand` is the random number
 generator object--in this case `std::mt19937`. The function returns a `cri` struct. 
 */
-cri *geCascade(int n,int cid, double Sn, int nlev, double *Elev, double *taus, double A, mt19937 *mtrand);
+cri *geCascade(int n,int cid, double Sn, int nlev, vector<double> Elev, vector<double> &taus, double A, mt19937 *mtrand);
 /**
   germanium func. to return the Energy after the mid-stop decay. `v` is the velocity in units of
 the speed of light (c); `M` is the mass of the slowing object in **GeV**, `Egam` is the energy of
@@ -162,7 +162,7 @@ double vgeS2func(double *x,double *par);
 lifetimes of size `nlev` in **fs**; `A` is the mass number; `mtrand` is the random number
 generator object--in this case `std::mt19937`. The function returns a `cri` struct. 
 */
-cri *siCascade(int n,int cid, double Sn, int nlev, double *Elev, double *taus, double A, mt19937 *mtrand);
+cri *siCascade(int n,int cid, double Sn, int nlev, vector<double> Elev, vector<double> &taus, double A, mt19937 *mtrand);
 /**
   silicon func. to return the Energy after the mid-stop decay. `v` is the velocity in units of
 the speed of light (c); `M` is the mass of the slowing object in **GeV**, `Egam` is the energy of
@@ -206,7 +206,7 @@ double vsiS2func(double *x,double *par);
 lifetimes of size `nlev` in **fs**; `A` is the mass number; `mtrand` is the random number
 generator object--in this case `std::mt19937`. The function returns a `cri` struct. 
 */
-cri *arCascade(int n,int cid, double Sn, int nlev, double *Elev, double *taus, double A, mt19937 *mtrand);
+cri *arCascade(int n,int cid, double Sn, int nlev, vector<double> Elev, vector<double> &taus, double A, mt19937 *mtrand);
 /**
   argon func. to return the Energy after the mid-stop decay. `v` is the velocity in units of
 the speed of light (c); `M` is the mass of the slowing object in **GeV**, `Egam` is the energy of
@@ -250,7 +250,7 @@ double varS2func(double *x,double *par);
 lifetimes of size `nlev` in **fs**; `A` is the mass number; `mtrand` is the random number
 generator object--in this case `std::mt19937`. The function returns a `cri` struct. 
 */
-cri *neCascade(int n,int cid, double Sn, int nlev, double *Elev, double *taus, double A, mt19937 *mtrand);
[…]
