+++ src/cascadeProd.c
@@ -15,39 +15,17 @@
 
 void freecliarray(int n,cli *cascade_levels)
 {
-  for(int i=0;i<n;i++){
-    freecli((cascade_levels+i));
-  }
   free(cascade_levels);
-  return;
-}
-void freecli(cli *cascade_levels)
-{
-  free(cascade_levels->Elev);
-  free(cascade_levels->taus);
   return;
 }
 void freecriarray(int n,cri *cascade_data)
 {
-  for(int i=0;i<n;i++){
-    freecri((cascade_data+i));
-  }
   free(cascade_data);
-  return;
-}
-void freecri(cri *cascade_data)
-{
-  free(cascade_data->E);
-  free(cascade_data->delE);
-  free(cascade_data->I);
-  free(cascade_data->Ei);
-  free(cascade_data->time);
-  free(cascade_data->Eg);
   return;
 }
 //do a generalized multi-step cascade (for now just print a table and do one event)
 //eventually: can do n events, put in a yield model function, generalize to other elements 
-cri *geCascade(int n, int cid, double Sn, int nlev, double *Elev, double *taus, double A, mt19937 *mtrand)
+vector<cri> &geCascade(int n, int cid, double Sn, int nlev, vector<double> &Elev, vector<double> &taus, double A, mt19937 *mtrand) // vector<double> &Elev ---> vector<double> &Elev
 {
   //input:
   //the neutron separation Sn in MeV
@@ -58,22 +36,22 @@ cri *geCascade(int n, int cid, double Sn, int nlev, double *Elev, double *taus, 
   //a previously seeded random number
 
   //make an output struct
-  cri *outinfo;
-  outinfo = (cri*) malloc(n*sizeof(cri));
+  vector<cri> outinfo;
+  outinfo = vector<cri>(n);
 
   //string of isotope name
   char isoname[5]; 
   sprintf(isoname,"%dGe",(int)A);
 
   for(int ev=0;ev<n;ev++){
-    outinfo[ev].n = nlev;
+    outinfo[ev].n = nlev; 
     outinfo[ev].cid = cid;
-    outinfo[ev].E = (double*) malloc(outinfo[ev].n*sizeof(double));
-    outinfo[ev].delE = (double*) malloc(outinfo[ev].n*sizeof(double));
-    outinfo[ev].I = (int*) malloc(outinfo[ev].n*sizeof(int));
-    outinfo[ev].Ei = (double*) malloc(outinfo[ev].n*sizeof(double));
-    outinfo[ev].time = (double*) malloc(outinfo[ev].n*sizeof(double));
-    outinfo[ev].Eg = (double*) malloc(outinfo[ev].n*sizeof(double));
+    outinfo[ev].E = vector<double>(outinfo[ev].n); // will become outinfo[ev].E = vector<double>(outinfo[ev].n,0.0)
+    outinfo[ev].delE = vector<double>(outinfo[ev].n);
+    outinfo[ev].I = vector<int>(outinfo[ev].n);
+    outinfo[ev].Ei = vector<double>(outinfo[ev].n);
+    outinfo[ev].time = vector<double>(outinfo[ev].n);
+    outinfo[ev].Eg = vector<double>(outinfo[ev].n);
 
     //keep track of some running variables like energy (eV), deposited energy (eV), Ionization (npairs), time (fs)
     double E,delE,I,Ei,time,Eg;
@@ -179,7 +157,7 @@ double geDecay(double v, double M, double Egam, mt19937 *mtrand)
   return El;
 }
 //return the velocity at a random stopping time
-double *geStop(double E, double M, double tau, mt19937 *mtrand)
+vector<double> &geStop(double E, double M, double tau, mt19937 *mtrand)
 {
   //random distribution
   uniform_real_distribution<double> dist(0.,1.);
[…]
