### 2022-01-13 5e1f1f7
Finish switching over to vectors Still need to figure out the warnings and errors.
