+++ src/cascadeProd.c
@@ -78,11 +78,10 @@ vector<cri> &geCascade(int n, int cid, double Sn, int nlev, vector<double> &Elev
         egam = ((Elev[i]/1000.0)-(Elev[i+1]/1000.0)); // in MeV
 
       //compute deposited energy and decay time
-      double *stopinfo;
+      vector<double> stopinfo;
       stopinfo = geStop(E,mass,taus[i],mtrand);
       double vdecay = stopinfo[0]; 
       double t = stopinfo[1];
-      free(stopinfo);
       double Eleft = ((mass*1e9)/2.0)*(pow(vdecay,2.0));
       double Erest = geDecay(vdecay,mass,egam,mtrand);
 
@@ -282,7 +281,7 @@ double vgeS2func(double *x,double *par)
 }
 //do a generalized multi-step cascade (for now just print a table and do one event)
 //eventually: can do n events, put in a yield model function, generalize to other elements 
-cri *siCascade(int n, int cid, double Sn, int nlev, vector<double> &Elev, vector<double> &taus, double A, mt19937 *mtrand)
+vector<cri> siCascade(int n, int cid, double Sn, int nlev, vector<double> &Elev, vector<double> &taus, double A, mt19937 *mtrand)
 {
   //input:
   //the neutron separation Sn in MeV
@@ -293,8 +292,8 @@ cri *siCascade(int n, int cid, double Sn, int nlev, vector<double> &Elev, vector
   //a previously seeded random number
 
   //make an output struct
-  cri *outinfo;
-  outinfo = cri (n);
+  vector<cri> outinfo;
+  outinfo = vector<cri>(n);
 
   //string of isotope name
   char isoname[5]; 
@@ -335,11 +334,10 @@ cri *siCascade(int n, int cid, double Sn, int nlev, vector<double> &Elev, vector
         egam = ((Elev[i]/1000.0)-(Elev[i+1]/1000.0)); // in MeV
 
       //compute deposited energy and decay time
-      double *stopinfo;
+      vector<double> stopinfo;
       stopinfo = siStop(E,mass,taus[i],mtrand);
       double vdecay = stopinfo[0]; 
       double t = stopinfo[1];
-      free(stopinfo);
       double Eleft = ((mass*1e9)/2.0)*(pow(vdecay,2.0));
       double Erest = siDecay(vdecay,mass,egam,mtrand);
 
@@ -414,7 +412,7 @@ double siDecay(double v, double M, double Egam, mt19937 *mtrand)
   return El;
 }
 //return the velocity at a random stopping time
-double *siStop(double E, double M, double tau, mt19937 *mtrand)
+vector<double> &siStop(double E, double M, double tau, mt19937 *mtrand)
 {
   //random distribution
   uniform_real_distribution<double> dist(0.,1.);
@@ -422,8 +420,8 @@ double *siStop(double E, double M, double tau, mt19937 *mtrand)
   //assume energy in eV, mass in GeV, tau in fs.
  
   //return both the energy and stopping time
-  double *ret;
-  ret = (double*) malloc(2*sizeof(double));
+  vector<double> ret;
+  ret = vector<double>(2);
 
   //use the Mersenne Twister for a uniform rand number (tau in fs)
   //this is done by inversion method
@@ -539,7 +537,7 @@ double vsiS2func(double *x,double *par)
 }
 //do a generalized multi-step cascade (for now just print a table and do one event)
 //eventually: can do n events, put in a yield model function, generalize to other elements 
[…]
