+++ inc/cascadeProd.h
@@ -118,7 +118,7 @@ void freecriarray(int n,cri *cascade_data);
 lifetimes of size `nlev` in **fs**; `A` is the mass number; `mtrand` is the random number
 generator object--in this case `std::mt19937`. The function returns a `cri` struct. 
 */
-cri *geCascade(int n,int cid, double Sn, int nlev, vector<double> Elev, vector<double> &taus, double A, mt19937 *mtrand);
+vector<cri> &geCascade(int n,int cid, double Sn, int nlev, vector<double> Elev, vector<double> &taus, double A, mt19937 *mtrand);
 /**
   germanium func. to return the Energy after the mid-stop decay. `v` is the velocity in units of
 the speed of light (c); `M` is the mass of the slowing object in **GeV**, `Egam` is the energy of
@@ -132,7 +132,7 @@ in **eV**; `M` is the mass of the slowing object in **GeV**; `tau` is the (rando
 intermediate nuclear level; `mtrand` is the random generator object--in this case `std::mt19937`.
 The function returns the velocity at the random decay time in units of the speed of light (c).
 */
-double *geStop(double E, double M, double tau, mt19937 *rand);
+vector<double> &geStop(double E, double M, double tau, mt19937 *rand);
 /**
   germanium func. constant stopping-power stopping. `E` is the initial energy of the stopping
 atom in **eV**; `M` is the mass of the slowing object in **GeV**; `t` is the time (in **fs**) at
@@ -162,7 +162,7 @@ double vgeS2func(double *x,double *par);
 lifetimes of size `nlev` in **fs**; `A` is the mass number; `mtrand` is the random number
 generator object--in this case `std::mt19937`. The function returns a `cri` struct. 
 */
-cri *siCascade(int n,int cid, double Sn, int nlev, vector<double> Elev, vector<double> &taus, double A, mt19937 *mtrand);
+vector<cri> &siCascade(int n,int cid, double Sn, int nlev, vector<double> Elev, vector<double> &taus, double A, mt19937 *mtrand);
 /**
   silicon func. to return the Energy after the mid-stop decay. `v` is the velocity in units of
 the speed of light (c); `M` is the mass of the slowing object in **GeV**, `Egam` is the energy of
@@ -176,7 +176,7 @@ in **eV**; `M` is the mass of the slowing object in **GeV**; `tau` is the (rando
 intermediate nuclear level; `mtrand` is the random generator object--in this case `std::mt19937`.
 The function returns the velocity at the random decay time in units of the speed of light (c).
 */
-double *siStop(double E, double M, double tau, mt19937 *rand);
+vector<double> &siStop(double E, double M, double tau, mt19937 *rand);
 /**
   silicon func. constant stopping-power stopping. `E` is the initial energy of the stopping
 atom in **eV**; `M` is the mass of the slowing object in **GeV**; `t` is the time (in **fs**) at
@@ -206,7 +206,7 @@ double vsiS2func(double *x,double *par);
 lifetimes of size `nlev` in **fs**; `A` is the mass number; `mtrand` is the random number
 generator object--in this case `std::mt19937`. The function returns a `cri` struct. 
 */
-cri *arCascade(int n,int cid, double Sn, int nlev, vector<double> Elev, vector<double> &taus, double A, mt19937 *mtrand);
+vector<cri> &arCascade(int n,int cid, double Sn, int nlev, vector<double> Elev, vector<double> &taus, double A, mt19937 *mtrand);
 /**
   argon func. to return the Energy after the mid-stop decay. `v` is the velocity in units of
 the speed of light (c); `M` is the mass of the slowing object in **GeV**, `Egam` is the energy of
@@ -220,7 +220,7 @@ in **eV**; `M` is the mass of the slowing object in **GeV**; `tau` is the (rando
[…]
