### 2022-01-13 8cdfced
More work hunting down existing errors.                                                                                (base)                                                                    Why does it think there is no Elev or taus member? both are part of the struct.
