+++ src/rootUtil.c
@@ -7,7 +7,7 @@ using namespace std;
 
 #include "rootUtil.h"
 
-bool addToNRTTree(TTree *t,int nr,cri *recoil_lists,cli cascade_spec)
+bool addToNRTTree(TTree *t,int nr,vector<cri> &recoil_lists,vector<cli> &cascade_spec)
 {
   //set up some variables
   Long64_t n,cid; //number of steps and cascade ID
