### 2022-01-13 89d388a
Successful conversion of arrays to vectors
