+++ inc/cascadeProd.h
@@ -91,23 +91,6 @@ typedef struct cri cri;
 
 //data functions
 
-/**
-  function to free memory in cli struct 
-*/
-void freecli(cli *cascade_levels);
-/**
-  function to free memory in array of cli structs 
-*/
-void freecliarray(int n,cli *cascade_levels);
-/**
-  function to free memory in cri struct 
-*/
-void freecri(cri *cascade_data);
-/**
-  function to free memory in array of cri structs 
-*/
-void freecriarray(int n,cri *cascade_data);
-
 //******germanium*****************
 //do a generalized multi-step cascade (for now just print a table and do one event)
 //eventually: can do n events, put in a yield model function, generalize to other elements 
@@ -118,7 +101,7 @@ void freecriarray(int n,cri *cascade_data);
 lifetimes of size `nlev` in **fs**; `A` is the mass number; `mtrand` is the random number
 generator object--in this case `std::mt19937`. The function returns a `cri` struct. 
 */
-vector<cri> &geCascade(int n,int cid, double Sn, int nlev, vector<double> &Elev, vector<double> &taus, double A, mt19937 *mtrand);
+vector<cri> geCascade(int n,int cid, double Sn, int nlev, vector<double> &Elev, vector<double> &taus, double A, mt19937 *mtrand);
 /**
   germanium func. to return the Energy after the mid-stop decay. `v` is the velocity in units of
 the speed of light (c); `M` is the mass of the slowing object in **GeV**, `Egam` is the energy of
@@ -132,7 +115,7 @@ in **eV**; `M` is the mass of the slowing object in **GeV**; `tau` is the (rando
 intermediate nuclear level; `mtrand` is the random generator object--in this case `std::mt19937`.
 The function returns the velocity at the random decay time in units of the speed of light (c).
 */
-vector<double> &geStop(double E, double M, double tau, mt19937 *rand);
+vector<double> geStop(double E, double M, double tau, mt19937 *rand);
 /**
   germanium func. constant stopping-power stopping. `E` is the initial energy of the stopping
 atom in **eV**; `M` is the mass of the slowing object in **GeV**; `t` is the time (in **fs**) at
@@ -150,7 +133,7 @@ double vgeS2(double E, double M, double t);
 which to return the atoms velocity; `par[0]` is the atoms initial energy in **eV**; `par[1] is the
 atoms mass in **GeV**. Function returns velocity in units of the speed of light (c).
 */
-double vgeS2func(double *x,double *par);
+double vgeS2func(vector<double> &x,vector<double> &par);
 
 //******silicon*****************
 //do a generalized multi-step cascade (for now just print a table and do one event)
@@ -162,7 +145,7 @@ double vgeS2func(double *x,double *par);
 lifetimes of size `nlev` in **fs**; `A` is the mass number; `mtrand` is the random number
 generator object--in this case `std::mt19937`. The function returns a `cri` struct. 
 */
-vector<cri> &siCascade(int n,int cid, double Sn, int nlev, vector<double> &Elev, vector<double> &taus, double A, mt19937 *mtrand);
+vector<cri> siCascade(int n,int cid, double Sn, int nlev, vector<double> &Elev, vector<double> &taus, double A, mt19937 *mtrand);
 /**
   silicon func. to return the Energy after the mid-stop decay. `v` is the velocity in units of
 the speed of light (c); `M` is the mass of the slowing object in **GeV**, `Egam` is the energy of
@@ -176,7 +159,7 @@ in **eV**; `M` is the mass of the slowing object in **GeV**; `tau` is the (rando
 intermediate nuclear level; `mtrand` is the random generator object--in this case `std::mt19937`.
 The function returns the velocity at the random decay time in units of the speed of light (c).
 */
[…]
