### 2022-01-13 8e5bc2f
Temporarily disable fail_on_warning to help troubleshoot.
