### 2022-01-17 #3993 @altavir
The primary critical point that should be fixed in my opinion is https://github.com/villano-lab/nrCascadeSim/issues/75. The part with comparison. Data tests would be welcome for future development, but not so critical.
