### 2022-01-18 #3993 @villaa
> The primary critical point that should be fixed in my opinion is [villano-lab/nrCascadeSim#75](https://github.com/villano-lab/nrCascadeSim/issues/75). The part with comparison. Data tests would be welcome for future development, but not so critical.

@altavir thanks, we are working on using Geant4 model for our comparison. It is taking a bit of time b/c we want to get the most recent version and I haven't run the module in a while. Incidentally, [here](https://github.com/villano-lab/G4PhysicsTesting) is the code we're using for that.
