### 2022-01-18 #3993 @matthewfeickert
Thanks for your submission and software @villaa. The following is my review of it, which goes into detail (and so isn't just noted in the first post of this Issue) only to try to help make requested revisions more actionable and clear. As always, the review is following the [JOSS paper review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html). Also, please feel free to ask for clarification on anything as I'd like to make sure that this information is helpful and actionable to you. :+1:

For consistency and reproducibility of all my comments, I will be using the public Docker image `atlasamglab/stats-base:root6.24.06` (where ROOT is built from source and installed from the Git tag `6.24.06`) to conduct my review which has the following software specs

$ docker run --rm -ti -v $PWD:$PWD -w $PWD atlasamglab/stats-base:root6.24.06
# gcc --version | head -n 1
gcc (Debian 10.2.1-6) 10.2.1 20210110
# root --version
ROOT Version: 6.24/06
Built for linuxx8664gcc on Jan 12 2022, 09:39:00
From tags/v6-24-06@v6-24-06
# root-config --cflags
-pthread -std=c++17 -m64 -I/usr/local/venv/include
```

## Review checklist

This review is for [tag `v1.3.1`](https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.3.1) of the project.

   - The repo has an OSI approved license (MIT), but vendors the Msersenne Twister prime algorithm without any LICENSE, which is a problem. All vendored software must be properly licensed. It is also unclear why this is even needed as the [Mersenne Twister 19937 generator is in the C++ Standard Library](https://www.cplusplus.com/reference/random/mt19937/).

   - Yes. Out of the number of commits in this project related to the software itself the submitting author @villaa has made the majority of them, and team member @gerudo7 has made nearly all the rest.

   ```
   $ git log --oneline ":(exclude)paper" | wc -l
   507
   $ git log --author="Anthony Villano" --oneline ":(exclude)paper" | wc -l
   296
   $ git log --author="Kitty" --oneline ":(exclude)paper" | wc -l
   195
   ```

JOSS asks for a minimum of 3 months of work equivalent effort for the software to be considered. The project definitely fulfills this requirement.

   - Yes. :+1:

<summary>Install example:</summary>

# git clone --depth 1 https://github.com/villano-lab/nrCascadeSim.git --branch v1.3.1 --single-branch
# cd nrCascadeSim/
# make
g++ -fPIC -c -D__GIT_VERSION=\"v1.3.1\" -IMersenne -Iinc src//isotope_info.c  `root-config --cflags` -L. -L/home/data/nrCascadeSim/bin/lib -fprofile-arcs -ftest-coverage
mv isotope_info.o /home/data/nrCascadeSim/bin/lib/
g++ -fPIC -c -D__GIT_VERSION=\"v1.3.1\" -IMersenne -Iinc src//weisskopf.c  `root-config --cflags` -L. -L/home/data/nrCascadeSim/bin/lib -fprofile-arcs -ftest-coverage
mv weisskopf.o /home/data/nrCascadeSim/bin/lib/
g++ -fPIC -c -D__GIT_VERSION=\"v1.3.1\" -IMersenne -Iinc src//lindhard.c  `root-config --cflags` -L. -L/home/data/nrCascadeSim/bin/lib -fprofile-arcs -ftest-coverage
mv lindhard.o /home/data/nrCascadeSim/bin/lib/
g++ -fPIC -c -D__GIT_VERSION=\"v1.3.1\" -IMersenne -Iinc src//cascadeProd.c  `root-config --cflags` -L. -L/home/data/nrCascadeSim/bin/lib -fprofile-arcs -ftest-coverage
mv cascadeProd.o /home/data/nrCascadeSim/bin/lib/
g++ -fPIC -c -D__GIT_VERSION=\"v1.3.1\" -IMersenne -Iinc src//edepmath.c `root-config --cflags` -L. -L/home/data/nrCascadeSim/bin/lib -fprofile-arcs -ftest-coverage
mv edepmath.o /home/data/nrCascadeSim/bin/lib/
g++ -fPIC -c -D__GIT_VERSION=\"v1.3.1\" -IMersenne -Iinc src//rootUtil.c `root-config --cflags`  -L. -L/home/data/nrCascadeSim/bin/lib -fprofile-arcs -ftest-coverage
mv rootUtil.o /home/data/nrCascadeSim/bin/lib/
g++ -fPIC -shared /home/data/nrCascadeSim/bin/lib/lindhard.o /home/data/nrCascadeSim/bin/lib/weisskopf.o /home/data/nrCascadeSim/bin/lib/isotope_info.o /home/data/nrCascadeSim/bin/lib/cascadeProd.o /home/data/nrCascadeSim/bin/lib/edepmath.o /home/data/nrCascadeSim/bin/lib/rootUtil.o `root-config --cflags` -L`root-config --libdir` -lCore -lRIO -lTree  -o /home/data/nrCascadeSim/bin/lib/libncap.so -fprofile-arcs -ftest-coverage
g++ -fPIC -Wl,-rpath,`root-config --libdir`,-rpath /home/data/nrCascadeSim/bin/lib -D__GIT_VERSION=\"v1.3.1\" -IMersenne -Iinc -L. -L/home/data/nrCascadeSim/bin/lib /home/data/nrCascadeSim/bin/realizeCascades.cpp `root-config --cflags` -L`root-config --libdir` -lCore -lRIO -lTree -lncap -o /home/data/nrCascadeSim/bin/realizeCascades -coverage -fprofile-arcs -ftest-coverage
g++ -fPIC -Wl,-rpath /home/data/nrCascadeSim/bin/lib -D__GIT_VERSION=\"v1.3.1\" -IMersenne -Iinc -L. -L/home/data/nrCascadeSim/bin/lib /home/data/nrCascadeSim/bin/regexPlayground.cpp -o /home/data/nrCascadeSim/bin/regexPlayground -coverage -fprofile-arcs -ftest-coverage
root@4b4531b84c8f:~/data/nrCascadeSim# tree bin  # I install tree for example purposes
bin
├── lib
│   ├── cascadeProd.o
│   ├── edepmath.o
│   ├── isotope_info.o
│   ├── libncap.so
│   ├── lindhard.o
│   ├── rootUtil.o
│   └── weisskopf.o
├── realizeCascades
├── realizeCascades.cpp
├── regexPlayground
└── regexPlayground.cpp

1 directory, 11 files
root@4b4531b84c8f:~/data/nrCascadeSim# command -v realizeCascades
root@4b4531b84c8f:~/data/nrCascadeSim# make install
cp /home/data/nrCascadeSim/bin/realizeCascades /usr/local/bin/
cp /home/data/nrCascadeSim/bin/regexPlayground /usr/local/bin/
root@4b4531b84c8f:~/data/nrCascadeSim# command -v realizeCascades
/usr/local/bin/realizeCascades

```

The [`install` Make target](https://github.com/villano-lab/nrCascadeSim/blob/182e8a56406a8ae4e5cfc8df489d9779d25d93da/Makefile#L80-L82) hardcodes the install target directory to `/usr/local/bin`. Install locations can have defaults, but should allow for user input &mdash; there are many situations on remotes (example: CERN's LXPLUS) where users do not have access to `/usr/local/` and the users would need to manually edit the `Makefile`.

Also, the `clean` Make target should for the same reasons _not_ [try to `rm -f` anything under `/usr/local`](https://github.com/villano-lab/nrCascadeSim/blob/182e8a56406a8ae4e5cfc8df489d9779d25d93da/Makefile#L95-L96). This is also probably never what a user wants to have happen when `make clean` is run &mdash; `clean` targets are not uninstallers. Similarly, advocating directly for the use of `sudo` for installation of local software should be avoided. The `.gcno` generated during the build should also be cleaned in the `clean` target (also check for `.gcda` files to be safe).

A better alternative default is to assume the user is NOT root, and to attempt to install in the user's `~/bin` where they will be assured write access.

Yes. The software produces reasonable results as far as I can tell and is able to pass the test suite (which includes the claims for support of Neon, Argon, Silicon, and Germanium cascades).

No performance claims are made.

Yes. The docs include [the statement](https://github.com/villano-lab/nrCascadeSim/blob/182e8a56406a8ae4e5cfc8df489d9779d25d93da/docs/source/index.rst?plain=1#L34-L37)

> The purpose of this code is to simulate energy deposits due to cascading of energy levels following neutron capture. This code was written for use in nuclear recoil calibration for dark matter detectors, but may be useful in other particle physics applications as well, including coherent elastic neutrino nucleus scattering (CEνNS).

The [docs on ReadTheDocs](https://nrcascadesim.readthedocs.io/) are not versioned beyond `latest` and `stable`. It would be great to see releases have versioned docs on ReadTheDocs too.

   - Yes. The software only depends on the `ROOT` C++ library and the `gcc` compiler. It is not possible to install `ROOT` with a OS package manager (and probably never will be given the size of the software) but the team nicely mentions that there are pre-package binaries from the CERN ROOT team as well as a conda-forge distribution (Conda is a package/environment manager, so this counts in my mind).

Yes. I have tested the examples and found that they work. They also pertain to dark matter detection physics (the area of research expertise of the authors).

<summary>The main example from the top level of the repository:</summary>

# time realizeCascades --verbose -n 100000 -o output.root levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
Seed used: 498136696
MT Result: 0.424727

real	0m1.495s
user	0m1.345s
sys	0m0.098s
# root -l output.root
root [0]
Attaching file output.root as _file0...
(TFile *) 0x55d12fd98a40
root [1] _file0->ls()
TFile**		output.root
 TFile*		output.root
  KEY: TTree	cascade;1	cascade
root [2] cascade->Print()
******************************************************************************
*Tree    :cascade   : cascade                                                *
*Entries :    99968 : Total =        19053073 bytes  File  Size =    7391802 *
*        :          : Tree compression factor =   2.58                       *
******************************************************************************
*Br    0 :n         : n/L                                                    *
*Entries :    99968 : Total  Size=     802407 bytes  File Size  =       7376 *
*Baskets :       26 : Basket Size=      32000 bytes  Compression= 108.68     *
*............................................................................*
*Br    1 :cid       : cid/L                                                  *
*Entries :    99968 : Total  Size=     802467 bytes  File Size  =       6890 *
*Baskets :       26 : Basket Size=      32000 bytes  Compression= 116.35     *
*............................................................................*
*Br    2 :Elev      : Elev[n]/D                                              *
*Entries :    99968 : Total  Size=    2181158 bytes  File Size  =     142555 *
*Baskets :       81 : Basket Size=      32000 bytes  Compression=  15.29     *
*............................................................................*
*Br    3 :taus      : taus[n]/D                                              *
*Entries :    99968 : Total  Size=    2181158 bytes  File Size  =     142237 *
*Baskets :       81 : Basket Size=      32000 bytes  Compression=  15.32     *
*............................................................................*
*Br    4 :E         : E[n]/D                                                 *
*Entries :    99968 : Total  Size=    2180903 bytes  File Size  =    1141056 *
*Baskets :       81 : Basket Size=      32000 bytes  Compression=   1.91     *
*............................................................................*
*Br    5 :delE      : delE[n]/D                                              *
*Entries :    99968 : Total  Size=    2181158 bytes  File Size  =    1751576 *
*Baskets :       81 : Basket Size=      32000 bytes  Compression=   1.24     *
*............................................................................*
*Br    6 :I         : I[n]/D                                                 *
*Entries :    99968 : Total  Size=    2180903 bytes  File Size  =     421026 *
*Baskets :       81 : Basket Size=      32000 bytes  Compression=   5.18     *
*............................................................................*
*Br    7 :Ei        : Ei[n]/D                                                *
*Entries :    99968 : Total  Size=    2180988 bytes  File Size  =    1749733 *
*Baskets :       81 : Basket Size=      32000 bytes  Compression=   1.25     *
*............................................................................*
*Br    8 :time      : time[n]/D                                              *
*Entries :    99968 : Total  Size=    2181158 bytes  File Size  =    1880987 *
*Baskets :       81 : Basket Size=      32000 bytes  Compression=   1.16     *
*............................................................................*
*Br    9 :Eg        : Eg[n]/D                                                *
*Entries :    99968 : Total  Size=    2180988 bytes  File Size  =     142468 *
*Baskets :       81 : Basket Size=      32000 bytes  Compression=  15.29     *
*............................................................................*

```

The individual examples that are in the [`tests/bin` directory](https://github.com/villano-lab/nrCascadeSim/tree/182e8a56406a8ae4e5cfc8df489d9779d25d93da/bin) build and run, and from the documentation seem to be doing what they each describe.

The examples in [`example-usecase/Yields_and_Resolutions.ipynb`](https://github.com/villano-lab/nrCascadeSim/blob/182e8a56406a8ae4e5cfc8df489d9779d25d93da/example-usecase/Yields_and_Resolutions.ipynb) are nice and it would be great to see these get added into the docs using some of the Jupyter rendering Sphinx extensions.

Yes, the docs exist and while there are some Sphinx errors in terms of syntax, these are easy to fix.

There is an outline of the CLI API for the `realizeCascades` executable [that is hardcoded](https://github.com/villano-lab/nrCascadeSim/blob/182e8a56406a8ae4e5cfc8df489d9779d25d93da/docs/source/02_Executables.rst) and the [`nrCascadeSim` library docs are generated with Doxygen](https://github.com/villano-lab/nrCascadeSim/blob/182e8a56406a8ae4e5cfc8df489d9779d25d93da/docs/source/03_Doxy_Libraries.rst).

The `realizeCascades` text input file format and ROOT binary output file format are also documented (I might suggest promoting I/O to its own docs section).

The use of Doxygen for the library is nice (kudos so the authors for taking this pre-review comment under suggestion) and the only advice I have about hardcoding of CLI API is to make sure that all examples in the docs for that section are rigorously tested in CI.

Yes. The author have a CI system using Travis CI that tests the software against ROOT `v6.20.00` and `v6.24.06` across `c++11`, `c++14`, and `c++17` on various Ubuntu Linux distributions and a subset of those configurations on various MacOS releases.

The code coverage at the time of the review is 82%. While low, this is acceptable for the review (though I'd encourage the authors to keep up their nice work of improving the coverage).

As already mentioned in https://github.com/villano-lab/nrCascadeSim/issues/75, the code is lacking regressions tests. It would be good if there could be at least one in the test suite and if the authors could plan to add more in the future.

While changing CI providers is not something that is relevant for this review, it should be noted that Travis CI now (2020 onwards) [offers a open source projects a one time non-renewable 10,000 build credit](https://blog.travis-ci.com/2020-11-02-travis-ci-new-billing).
While it does not impact this review, the authors will need to consider how they plan to continue testing their software once this on time limit has been reached.

Yes. There is a [`CONTRIBUTING.md`](https://github.com/villano-lab/nrCascadeSim/blob/182e8a56406a8ae4e5cfc8df489d9779d25d93da/CONTRIBUTING.md) and [`CODE_OF_CONDUCT.md`](https://github.com/villano-lab/nrCascadeSim/blob/182e8a56406a8ae4e5cfc8df489d9779d25d93da/CODE_OF_CONDUCT.md) and [the `README` mentions](https://github.com/villano-lab/nrCascadeSim/blob/182e8a56406a8ae4e5cfc8df489d9779d25d93da/README.md?plain=1#L51)

> For questions, support, bug reports, or other suggestions, please open an [issue](https://github.com/villano-lab/nrCascadeSim/issues).

Additionally, there is a [Contact & Support](https://nrcascadesim.readthedocs.io/en/stable/09_Contact.html) section of the docs and

# realizeCascades --help | tail -n 3
Report bugs to: villaa-at-gmail-dot-com
realizeCascades (nrCascadeSim) home page: <https://github.com/villano-lab/nrCascadeSim>
General help nrCascadeSim Software: <https://nrcascadesim.readthedocs.io/en/latest/>
```

Yes. The introduction does a good job of this.

Yes.

Yes. They compare `nrCascadeSim` to `Geant4` and `FIFRELIN` in a fair manner.

The authors don't mention this, but it is worth noting that [`Geant4` is a considerably larger project](https://github.com/Geant4/geant4) and given the large number of use cases it is meant to handle can be much more cumbersome to use. While `Geant4` is fully open source, it is not clear from searches if there are open source implementations of the `FIFRELIN` Monte Carlo simulator. If not, `nrCascadeSim` provides an advantage by being open source.

The paper is of good quality and of the proper scope following JOSS guidelines.

Yes.

## Suggested revisions before JOSS publication

@lucydot, I think the paper and the project are in overall good shape. My recommendation is that once the following revisions are addressed (I'll open up GitHub Issues for them) the project be accepted for publication pending your approval. Though we also still need @altavir's written review as well.

* Change the `Makefile` to remove targeting `/usr/local` as a default. In many situations this is an unwritable location for users.
   - If https://github.com/villano-lab/nrCascadeSim/issues/64 is enacted this can be easily achieved with setting the [`CMAKE_INSTALL_PREFIX` option](https://cmake.org/cmake/help/latest/variable/CMAKE_INSTALL_PREFIX.html) during `cmake`.
* Add removal of the `.gcno` and `.gcda` files generated during the build to the `clean` Make target.
* Remove recommendations of use of `sudo` to install software. Anyone who needs to use `sudo` does not need to be told to use `sudo` and all software should also be able to be installed and used locally.
* Add at least one regression test to the test suite and plan to add more in the future (compliment to https://github.com/villano-lab/nrCascadeSim/issues/75).

#### Paper
* In the State of the Field section it would be beneficial to mention that `Genat4` is open source software and to clarify if `FIFRELIN` is open source or not.

## Suggested revisions outside of review

These suggestions should not be considered inside the scope of the review and editors should not view their completion as a dependency for approval.

* Fix all Sphinx errors and typos.
* Strive to increase code coverage. Code coverage is not a magic bullet, and even with 100% coverage there can still be integration errors. However, in my experience once you start getting code coverage going over 90% you can develop much more aggressively as you know that your test suites are doing the hard work for you. :rocket:
* Migrate off of Travis CI. Travis CI has decided it will no longer offer free testing to the open source community and the 10,000 credits it gives users on the free trial plan should be viewed as an off ramp to ensure that there is robust CI support. As mentioned in https://github.com/villano-lab/nrCascadeSim/issues/10, alternatives that are easy to switch to are GitHub Actions or Azure Pipelines.
* The [docs on ReadTheDocs](https://nrcascadesim.readthedocs.io/) are not versioned beyond `latest` and `stable`. It would be great to see releases have versioned docs on ReadTheDocs too.
* The examples in [`example-usecase/Yields_and_Resolutions.ipynb`](https://github.com/villano-lab/nrCascadeSim/blob/182e8a56406a8ae4e5cfc8df489d9779d25d93da/example-usecase/Yields_and_Resolutions.ipynb) are nice and it would be great to see these get added into the docs using some of the Jupyter rendering Sphinx extensions.
