### 2022-01-18 3bc21dc
Update RELEASE_CHECKLIST to make sure that readthedocs version are updated with each release.
