### 2022-01-18 27cc7db
Making progress on getting CMakeLists working. Remove makefile as it will be replaced by CMakeLists.
