+++ Makefile
@@ -1,99 +1,0 @@
-###############################################################
-#
-#
-# Makefile for nrCascadeSim
-###############################################################
-
-#everybody here needs root
-ROOTFLAGS=`root-config --cflags`
-
-TOPDIR=$(shell pwd)
-BUILDDIR=$(TOPDIR)/bin
-INCDIROUT=$(BUILDDIR)/include
-LIBDIROUT=$(BUILDDIR)/lib
-TESTBIN=$(TOPDIR)/tests/bin
-INCDIR=inc/
-SRCDIR=src/
-INCFLAG= -IMersenne -Iinc
-LIBFLAG= -L. -L$(LIBDIROUT)
-
-
-#need to use an older version of gcc b/c of Mersenne Twister using the deprecated `register` keyword
-#https://github.com/danini/graph-cut-ransac/issues/23
-#CXX=clang++
-#CXXFLAGS=-g -std=c++11 -Wall -pedantic
-
-#trick for getting the git version in the code
-GIT_VERSION = $(shell sh -c 'git describe --abbrev=4 --always')
-
-override CFLAGS += -D__GIT_VERSION=\"$(GIT_VERSION)\"
-
-RUN_SCRIPT := $(shell mkdir -p 'bin/lib')
-
-all: $(BUILDDIR)/realizeCascades $(BUILDDIR)/regexPlayground $(LIBDIROUT)/rootUtil.o $(LIBDIROUT)/edepmath.o $(LIBDIROUT)/cascadeProd.o $(LIBDIROUT)/isotope_info.o $(LIBDIROUT)/weisskopf.o $(LIBDIROUT)/lindhard.o $(LIBDIROUT)/libncap.so
-
-$(LIBDIROUT)/isotope_info.o: $(SRCDIR)/isotope_info.cpp $(INCDIR)/isotope_info.h 
-	$(CXX) -fPIC -c $(CFLAGS) $(INCFLAG) $(SRCDIR)/isotope_info.cpp  `root-config --cflags` $(LIBFLAG) -fprofile-arcs -ftest-coverage
-	mv isotope_info.o $(LIBDIROUT)/
-
-$(LIBDIROUT)/rootUtil.o: $(SRCDIR)/rootUtil.cpp $(INCDIR)/rootUtil.h 
-	$(CXX) -fPIC -c $(CFLAGS) $(INCFLAG) $(SRCDIR)/rootUtil.cpp `root-config --cflags`  $(LIBFLAG) -fprofile-arcs -ftest-coverage
-	mv rootUtil.o $(LIBDIROUT)/
-
-$(LIBDIROUT)/edepmath.o: $(SRCDIR)/edepmath.cpp $(INCDIR)/edepmath.h 
-	$(CXX) -fPIC -c $(CFLAGS) $(INCFLAG) $(SRCDIR)/edepmath.cpp `root-config --cflags` $(LIBFLAG) -fprofile-arcs -ftest-coverage
-	mv edepmath.o $(LIBDIROUT)/
-
-$(LIBDIROUT)/weisskopf.o: $(SRCDIR)/weisskopf.cpp $(INCDIR)/weisskopf.h 
-	$(CXX) -fPIC -c $(CFLAGS) $(INCFLAG) $(SRCDIR)/weisskopf.cpp  `root-config --cflags` $(LIBFLAG) -fprofile-arcs -ftest-coverage
-	mv weisskopf.o $(LIBDIROUT)/
-
-$(LIBDIROUT)/lindhard.o: $(SRCDIR)/lindhard.cpp $(INCDIR)/lindhard.h 
[…]
