+++ CMakeLists.txt
@@ -4,6 +4,7 @@
 # Makefile for nrCascadeSim
 ###############################################################
 cmake_minimum_required(VERSION 3.10.2) #Just setting this to my version since I know that works.
+project(nrCascadeSim)
 
 # everybody here needs root
 SET(rootFlags `root-config --cflags`)
@@ -14,9 +15,10 @@ SET(gitVersion "$(shell sh -c 'git describe --abrev=4 --always')")
 #SET(buildDir $(topDir)/bin) #obsolete - we will mark this using CMake
 SET(incDirOut include)
 SET(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR}/lib)
-SET(testBin tests/bin)
-SET(incDir inc/)
-SET(srcDir src/)
+SET(testBin ${PROJECT_SOURCE_DIR}/tests/bin)
+SET(incDir ${PROJECT_SOURCE_DIR}/inc/)
+SET(srcDir ${PROJECT_SOURCE_DIR}/src/)
+SET(binDir ${PROJECT_SOURCE_DIR}/bin/)
 SET(libFLAG "-L. -L$(CMAKE_LIBRARY_OUTPUT_DIRECTORY)")
 SET(cFlags "-D__GIT_VERSION=\'$(GIT_VERSION)'")
 # Make necessary bin/lib directory
@@ -24,50 +26,51 @@ SET(cFlags "-D__GIT_VERSION=\'$(GIT_VERSION)'")
 # Send files to the right places
 
 
-add_library(isotope_info.o ${srcDir}/isotope_info.c ${incDir}/isotope_info.h)
-add_library(rootUtil.o ${srcDir}/rootUtil.c ${incDir}/rootUtil.h)
-add_library(edepmath.o ${srcDir}/edepmath.c ${incDir}/edepmath.h)
-add_library(weisskopf.o ${srcDir}/weisskopf.c ${incDir}/weisskopf.h)
-add_library(lindhard.o ${srcDir}/lindhard.c ${incDir}/lindhard.h)
-add_library(cascadeProd.o ${srcDir}/cascadeProd.c ${incDir}/cascadeProd.h)
+add_library(isotope_info.o ${srcDir}/isotope_info.cpp ${incDir}/isotope_info.h)
+add_library(rootUtil.o ${srcDir}/rootUtil.cpp ${incDir}/rootUtil.h)
+add_library(edepmath.o ${srcDir}/edepmath.cpp ${incDir}/edepmath.h)
+add_library(weisskopf.o ${srcDir}/weisskopf.cpp ${incDir}/weisskopf.h)
+add_library(lindhard.o ${srcDir}/lindhard.cpp ${incDir}/lindhard.h)
+add_library(cascadeProd.o ${srcDir}/cascadeProd.cpp ${incDir}/cascadeProd.h)
 
-add_executable(regexPlayground regexPlayground.cpp)
+add_executable(regexPlayground ${binDir}/regexPlayground.cpp)
 
[…]
