### 2022-01-19 d07a2d7
try testing CMake progress
