### 2022-01-19 737b412
Set up coverage and detect c++ standard from ROOT installation.
