### 2022-01-19 ed6107e
Fix default install location.
