### 2022-01-19 c36437d
Move test executables to appropriate directory on make install.
