+++ CMakeLists.txt
@@ -17,12 +17,14 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -D__GIT_VERSION=\\\"${GIT_VERSION}\\\"")
 #SET(topDir $(shell pwd))    #obsolete - we will mark this using CMake
 #SET(buildDir $(topDir)/bin) #obsolete - we will mark this using CMake
 SET(incDirOut include)
-SET(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR}/lib)
+SET(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR}/lib/)
 SET(testBin ${PROJECT_SOURCE_DIR}/tests/bin)
 SET(incDir ${PROJECT_SOURCE_DIR}/inc/)
 SET(srcDir ${PROJECT_SOURCE_DIR}/src/)
 SET(binDir ${PROJECT_SOURCE_DIR}/bin/)
 SET(CMAKE_STATIC_LIBRARY_SUFFIX ".o")
+SET(CMAKE_STATIC_LIBRARY_PREFIX "")
+SET(CMAKE_SHARED_LIBRARY_PREFIX "")
 SET(libFLAG "-L. -L$(CMAKE_LIBRARY_OUTPUT_DIRECTORY)")
 SET(INSTALL_DIR /usr/local/bin)
 # Make necessary bin/lib directory
@@ -96,7 +98,7 @@ add_executable(regexPlayground ${binDir}/regexPlayground.cpp)
 link_libraries(weisskopf lindhard cascadeProd edepmath rootUtil)
 add_library(libncap SHARED ${srcDir}/isotope_info.cpp ${srcDir}/rootUtil.cpp ${srcDir}/edepmath.cpp ${srcDir}/weisskopf.cpp ${srcDir}/lindhard.cpp ${srcDir}/cascadeProd.cpp ${incDir}/isotope_info.h ${incDir}/rootUtil.h ${incDir}/edepmath.h ${incDir}/weisskopf.h ${incDir}/lindhard.h ${incDir}/cascadeProd.h)
 
-link_libraries(libncap weisskopf lindhard cascadeProd edepmath rootUtil ${ROOT_LIBRARIES})
+link_libraries(libncap ${ROOT_LIBRARIES})
 
 add_executable(realizeCascades ${binDir}/realizeCascades.cpp)
 add_executable(fetchYieldModel ${testBin}/fetchYieldModel.cpp)
@@ -104,78 +106,7 @@ add_executable(printIsotopeInfo ${testBin}/printIsotopeInfo.cpp)
 add_executable(readLevelfile ${testBin}/readLevelfile.cpp)
 add_executable(realizeAndSave ${testBin}/realizeAndSave.cpp ${incDir}/cascadeProd.h)
 
-install(TARGETS realizeCascades regexPlayground RUNTIME DESTINATION ${INSTALL_DIR})
-install(CODE "EXECUTE_PROCESS(COMMAND mv fetchYieldModel ${testBin})")
-install(CODE "EXECUTE_PROCESS(COMMAND mv printIsotopeInfo ${testBin})")
-install(CODE "EXECUTE_PROCESS(COMMAND mv readLevelfile ${testBin})")
-install(CODE "EXECUTE_PROCESS(COMMAND mv realizeAndSave ${testBin})")
[…]
