### 2022-01-19 a187538
Get tests target working. Still having weird behavior for install and clean
