### 2022-01-20 4a6329e
Got linker working for make install Need to fix up make clean and coverage next
