### 2022-01-20 villano-lab/nrCascadeSim#89 @matthewfeickert
[JOSS Review] Enable versioning of docs on ReadTheDocs
(This is a suggested improvement, but not a change that needs to be made to pass review).

From https://github.com/openjournals/joss-reviews/issues/3993#issuecomment-1015075303:

> * The [docs on ReadTheDocs](https://nrcascadesim.readthedocs.io/) are not versioned beyond `latest` and `stable`. It would be great to see releases have versioned docs on ReadTheDocs too.

ReadTheDocs allows for easily providing versioned docs that match releases. This can be enabled on the project page. c.f. RTD's [Versioned Documentation page](https://docs.readthedocs.io/en/stable/versions.html) for more.
