### 2022-01-20 villano-lab/nrCascadeSim#90 @matthewfeickert
[JOSS Review] Fix Sphinx bugs and typos
(This is a suggested improvement, but not a change that needs to be made to pass review).

From openjournals/joss-reviews#3993 (comment):

> Fix all Sphinx errors and typos.

Sphinx is great, but is also touchy and can require some messing around with to get things formatted properly (Typesetting is hard).

---

**The file(s) in question is/are:**

I'll link to them below using the commit hash to get a render and permalink.

**The problem is in the following category/categories:**
- [x] Clarity
- [ ] Examples
- [ ] Broken links and images
- [x] Typos, spelling, and grammar
- [ ] Undocumented Information
- [ ] Out-of-date
- [ ] Other

**Description of the problem:**
Describe whatever is wrong with the documentation or could otherwise be improved.

- [ ] In general, when typing out something inline in the text that would be considered a command it is probably best to enter the `:code:` formatting mode, which has a shortcut of double backticks. Single backticks like

https://github.com/villano-lab/nrCascadeSim/blob/bf4795d83f8be6c9c1d737f5f51741def32cceb5/docs/source/02_Executables.rst#L13

end up just making things italicized, which probably isn't really the style you were going for and can obscure things like the number of dashes. This applies basically everywhere, but in the above try checking what happens when you use

```
``realizeCascades --help``
```

Same sort of thing for

https://github.com/villano-lab/nrCascadeSim/blob/bf4795d83f8be6c9c1d737f5f51741def32cceb5/docs/source/02_Executables.rst#L79-L80

and a bunch of other places in the docs.

- [ ] For

https://github.com/villano-lab/nrCascadeSim/blob/bf4795d83f8be6c9c1d737f5f51741def32cceb5/docs/source/02_Executables.rst#L46-L49

I think you need a return between the text and the list and you might also need indents (I can't remember off the top of my head). Otherwise it will all just blend together and be difficult to read.

- [ ] I like all of

https://github.com/villano-lab/nrCascadeSim/blob/bf4795d83f8be6c9c1d737f5f51741def32cceb5/docs/source/06_Citations.rst#L5-L33

a lot. You might be interested in using [`sphinx-copybutton`](https://sphinx-copybutton.readthedocs.io/en/latest/) as well to make it even nicer to copy both code blocks and citation format cells. :+1:

(There are a few other things, but it is getting a bit late tonight so I'll drop them in here later on. These are just suggestions, so not something to sprint on. :+1:)
