### 2022-01-20 #3993 @altavir
@lucydot I've marked all but the State of the field box. I think that a comparison of the results is required for it.
