### 2022-01-20 villano-lab/nrCascadeSim#89 @gerudo7
Think I got to this just before you posted it here -

> There are now doc versions for each minor version since we started using readthedocs (v1.2 and v1.3). Creating doc versions has also been added to the release checklist (3bc21dc).

_Originally posted by `@gerudo7` in #83_
