### 2022-01-20 bad7183
Adjust CMakeLists.txt defaults to match recommendations by matthewfeickert. And some other cleaning up.
