### 2022-01-20 bdbe2d6
Fix TEST_OUTPUT_DIR variable to be used correctly and set if empty Still need to figure out how to get make clean to include these files.
