### 2022-01-20 b94ba1e
Specify install location in cmake command.
