### 2022-01-21 7429fab
All CMake targets working locally! Currently only working without coverage.
