### 2022-01-21 fabf00b
Have .travis.yml use build dir instead of building directly in source dir.
