### 2022-01-21 c40c1df
Add debug line to retrieve location of repository in travis-ci. Where are you
