### 2022-01-21 9747fee
Another correction to clone directory for travis.
