### 2022-01-21 0b4aa7b
Move a troubleshooting line in travis to after the correct make install command.
