### 2022-01-21 355556a
Use built-in RPATH variables only instead of having special var for lib locations.
