+++ .travis.yml
@@ -80,7 +80,6 @@ script:
   - cmake -DCMAKE_INSTALL_PREFIX=/usr/local/bin -DUSE_GCOV=ON ~/build/villano-lab/nrCascadeSim
   - make
   - sudo make install
-  - ls /usr/local/lib | grep libncap
   - make tests
   # Send tests to directory they were written to run from.
   - mv readLevelfile ~/build/villano-lab/nrCascadeSim/tests/bin
