### 2022-01-21 3233660
Make C++ standard detection more flexible in CMakeLists.txt
