### 2022-01-22 3031c1c
Reinclude UPLOAD_COVERAGE var for xenial build with root v6.24/06
