### 2022-01-25 #3993 @lucydot
Hello @altavir - regarding state of the field: we do not usually require that there is a quantative comparison made with other  software. This section can include a qualitative comparison of functionality against other existing software (as is the case here), however if performance claims are made (e.g. that it runs faster) this must be backed up. As performance claims are not made for this code (from what I can see - please let me know if I am wrong), my view is that a comparison of results to other software is not needed.

However, as mentioned by Matthew above, and yourself in https://github.com/villano-lab/nrCascadeSim/issues/75, regression tests are needed to compare results from different versions of this software.
