### 2022-01-25 #3993 @altavir
@lucydot the problem in this particular case is the only aim of this piece of software is a simulation of a specific process. Without a comparison, it is not possible to decide if the aim of the development is achieved. In simulations it is quite easy to get completely wrong result even with correct theory. Therefore, I believe that at least some comparison should be made. Eithere with an experiment data (it is better) or with other software.

If the results are consistent with GEANT, but it worls much faster, it is OK. But if the results are significantly different, additional research is required to understand, which one is wrong (and GEANT could and is wrong on many occasions).
