### 2022-01-25 #3993 @villaa
@lucydot @altavir we are committed to putting in the comparison with Geant4 at this point in any case. We also already have regression tests in the queue to be released. So, perhaps those things will be enough to satisfy all.
