### 2022-01-25 villano-lab/nrCascadeSim#75 @gerudo7
Noting here that #77 added tests for all of the inputs and #80 added tests for resulting distributions, so I believe the only thing left is the Geant4 comparison testing, which is in progress.
