### 2022-01-25 f80c90e
[ANV] made small tweak to CMakeLists.txt file to add ROOT libs to libncap.so linking
