### 2022-01-25 c9e27ec
Get travis to switch to correct directory before trying coverage upload.
