+++ .travis.yml
@@ -114,4 +114,5 @@ script:
 
 # Send coverage info to codecov
 after_success: 
+  - cd ~/build/villano-lab/nrCascadeSim
   - if [ "$UPLOAD_COVERAGE" == "true" ]; then source .travis/uploadtocodecov.sh; fi
