### 2022-01-25 8987595
[ANV] might have fixed the git version variable in CMakeLists.txt
