### 2022-01-26 d1018d7
Try more general form of the lib location in rpath variables.
