+++ CMakeLists.txt
@@ -29,6 +29,9 @@ SET(CMAKE_SHARED_LIBRARY_PREFIX "")
 if(NOT CMAKE_INSTALL_PREFIX)
     SET(CMAKE_INSTALL_PREFIX ~/bin)
 endif(NOT CMAKE_INSTALL_PREFIX)
+if(NOT CMAKE_LIB_PREFIX)
+    SET(CMAKE_LIB_PREFIX ${CMAKE_INSTALL_PREFIX}/../lib)
+endif(NOT CMAKE_LIB_PREFIX)
 message("Install prefix: ${CMAKE_INSTALL_PREFIX}")
 #SET(LIB_INSTALL_DIR ${CMAKE_INSTALL_PREFIX}/../lib)
 # Make necessary bin/lib directory
@@ -38,11 +41,11 @@ include_directories(${incDir})
 
 #Get linker to work for make install
 ##[[
-list( APPEND CMAKE_INSTALL_RPATH ${CMAKE_INSTALL_PREFIX}/../lib )
+list( APPEND CMAKE_INSTALL_RPATH ${CMAKE_LIB_PREFIX} )
 set(CMAKE_INSTALL_RPATH_USE_LINK_PATH TRUE)
-list(FIND CMAKE_PLATFORM_IMPLICIT_LINK_DIRECTORIES "${CMAKE_INSTALL_PREFIX}/../lib" isSystemDir)
+list(FIND CMAKE_PLATFORM_IMPLICIT_LINK_DIRECTORIES "${CMAKE_LIB_PREFIX}" isSystemDir)
 if("${isSystemDir}" STREQUAL "-1")
-    set(CMAKE_INSTALL_RPATH "${CMAKE_INSTALL_PREFIX}/../lib")
+    list(APPEND CMAKE_INSTALL_RPATH "${CMAKE_LIB_PREFIX}")
 endif("${isSystemDir}" STREQUAL "-1")
 #]]
 
@@ -134,7 +137,7 @@ add_executable(readLevelfile EXCLUDE_FROM_ALL ${testBin}/readLevelfile.cpp)
 add_executable(realizeAndSave EXCLUDE_FROM_ALL ${testBin}/realizeAndSave.cpp ${incDir}/cascadeProd.h)
 
 install(TARGETS realizeCascades regexPlayground RUNTIME DESTINATION ${CMAKE_INSTALL_PREFIX})
-install(TARGETS libncap LIBRARY DESTINATION ${CMAKE_INSTALL_RPATH})
+install(TARGETS libncap LIBRARY DESTINATION ${CMAKE_LIB_PREFIX})
 #install(TARGETS fetchYieldModel printIsotopeInfo readLevelfile realizeAndSave DESTINATION ${testBin})
 
 add_custom_target(tests 
@@ -144,5 +147,5 @@ add_custom_target(tests
 add_custom_target(uninstall
                 COMMAND rm ${CMAKE_INSTALL_PREFIX}/realizeCascades
                 COMMAND rm ${CMAKE_INSTALL_PREFIX}/regexPlayground
-                COMMAND rm ${CMAKE_INSTALL_PREFIX}/../libncap.so
+                COMMAND rm ${CMAKE_LIB_PREFIX}/libncap.so
 )
