+++ CMakeLists.txt
@@ -44,11 +44,11 @@ include_directories(${incDir})
 
 #Get linker to work for make install
 ##[[
-list( APPEND CMAKE_INSTALL_RPATH ${CMAKE_LIB_PREFIX} )
+list( APPEND CMAKE_INSTALL_RPATH ${CMAKE_LIB_PREFIX} ${ROOT_INCLUDE_DIR})
 set(CMAKE_INSTALL_RPATH_USE_LINK_PATH TRUE)
-list(FIND CMAKE_PLATFORM_IMPLICIT_LINK_DIRECTORIES "${CMAKE_LIB_PREFIX}" isSystemDir)
+list(FIND CMAKE_PLATFORM_IMPLICIT_LINK_DIRECTORIES "${CMAKE_LIB_PREFIX} ${ROOT_INCLUDE_DIR}" isSystemDir)
 if("${isSystemDir}" STREQUAL "-1")
-    list(APPEND CMAKE_INSTALL_RPATH "${CMAKE_LIB_PREFIX}")
+    list(APPEND CMAKE_INSTALL_RPATH "${CMAKE_LIB_PREFIX} ${ROOT_INCLUDE_DIR}")
 endif("${isSystemDir}" STREQUAL "-1")
 #]]
 
