### 2022-01-26 eebf40a
[ANV] turn on make verbosity to diagnose CI problems
