### 2022-01-26 cc3e760
Check that root location listed by `root-config` has correct lib location and contents.
