+++ .travis.yml
@@ -58,6 +58,7 @@ install:
    - source root/bin/thisroot.sh
    - echo "root libraries.."
    - echo `root-config --cflags --glibs`
+   - ls ~/build/villano-lab/nrCascadeSim/root/lib
 
    # install and set up lcov
    - wget http://downloads.sourceforge.net/ltp/lcov-1.14.tar.gz
@@ -77,7 +78,7 @@ install:
 
 script:
   #make repository code (my code)
-  - cmake -DCMAKE_INSTALL_PREFIX=/usr/local/bin -DUSE_GCOV=ON --enable-rpath ~/build/villano-lab/nrCascadeSim
+  - cmake -DCMAKE_INSTALL_PREFIX=/usr/local/bin -DUSE_GCOV=ON ~/build/villano-lab/nrCascadeSim
   - make VERBOSE=1
   - sudo make install
   - make tests
