### 2022-01-28 bc1531c
[ANV] fixed the uninstall target and revised CMakeLists.txt
