+++ CMakeLists.txt
@@ -15,7 +15,6 @@ execute_process (
 )
 message("git version: ${GIT_VERSION}")
 SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -D__GIT_VERSION=\"\\\"${GIT_VERSION}\\\"\"")
-SET(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -D__GIT_VERSION=\"${GIT_VERSION}\"")
 # other variables
 #SET(topDir $(shell pwd))    #obsolete - we will mark this using CMake
 #SET(buildDir $(topDir)/bin) #obsolete - we will mark this using CMake
@@ -30,10 +29,10 @@ SET(CMAKE_STATIC_LIBRARY_PREFIX "")
 SET(CMAKE_SHARED_LIBRARY_PREFIX "")
 #SET(libFLAG "-L. -L$(CMAKE_LIBRARY_OUTPUT_DIRECTORY)")
 if(NOT CMAKE_INSTALL_PREFIX)
-    SET(CMAKE_INSTALL_PREFIX ~/bin)
+    SET(CMAKE_INSTALL_PREFIX ~/)
 endif(NOT CMAKE_INSTALL_PREFIX)
 if(NOT CMAKE_LIB_PREFIX)
-    SET(CMAKE_LIB_PREFIX ${CMAKE_INSTALL_PREFIX}/../lib)
+    SET(CMAKE_LIB_PREFIX ${CMAKE_INSTALL_PREFIX})
 endif(NOT CMAKE_LIB_PREFIX)
 message("Install prefix: ${CMAKE_INSTALL_PREFIX}")
 #SET(LIB_INSTALL_DIR ${CMAKE_INSTALL_PREFIX}/../lib)
@@ -41,39 +40,6 @@ message("Install prefix: ${CMAKE_INSTALL_PREFIX}")
 # SET(runScript "$(shell mkdir -p 'bin/lib')") #hopefully don't need this anymore
 # Send files to the right places
 include_directories(${incDir})
-
-#Get linker to work for make install
-list( APPEND CMAKE_INSTALL_RPATH ${CMAKE_LIB_PREFIX} ${ROOT_INCLUDE_DIR})
-set(CMAKE_INSTALL_RPATH_USE_LINK_PATH TRUE)
-#list(FIND CMAKE_PLATFORM_IMPLICIT_LINK_DIRECTORIES "${CMAKE_LIB_PREFIX} ${ROOT_INCLUDE_DIR}" isSystemDir)
-#if("${isSystemDir}" STREQUAL "-1")
-#    list(APPEND CMAKE_INSTALL_RPATH "${CMAKE_LIB_PREFIX} ${ROOT_INCLUDE_DIR}")
-#endif("${isSystemDir}" STREQUAL "-1")
-
-
-#[[ no longer used
-if(NOT TEST_OUTPUT_DIR)
-    SET(TEST_OUTPUT_DIR ${testBin})
-endif(NOT TEST_OUTPUT_DIR) #]]
-message("Tests output directory: ${TEST_OUTPUT_DIR} (Default: ${testBin})")
-
-file(GLOB sources ${srcDir}/*.cpp)
-file(GLOB headers ${incDir}/*.h)
-
-#Coverage
-if(NOT USE_GCOV)
-    SET(USE_GCOV OFF)
-endif(NOT USE_GCOV)
-message("Use GCOV? ${USE_GCOV}")
-if(USE_GCOV)
-	SET(GCC_COVERAGE_COMPILE_FLAGS "-fprofile-arcs -ftest-coverage")
[…]
