### 2022-01-28 e8aef8b
[ANV] updated diagnostic prints
