+++ CMakeLists.txt
@@ -86,12 +86,33 @@ message("C++ Standard: ${CMAKE_CXX_STANDARD}. (Detected from ROOT.)")
 
 
 #Get linker to work for make install
-list( APPEND CMAKE_INSTALL_RPATH ${CMAKE_LIB_PREFIX}/lib)
+#list( APPEND CMAKE_INSTALL_RPATH ${CMAKE_LIB_PREFIX}/lib)
+#set(CMAKE_INSTALL_RPATH_USE_LINK_PATH TRUE)
+#list(FIND CMAKE_PLATFORM_IMPLICIT_LINK_DIRECTORIES "${CMAKE_LIB_PREFIX}/lib" isSystemDir)
+#if("${isSystemDir}" STREQUAL "-1")
+#    list(APPEND CMAKE_INSTALL_RPATH "${CMAKE_LIB_PREFIX}/lib")
+#endif("${isSystemDir}" STREQUAL "-1")
+
+# use, i.e. don't skip the full RPATH for the build tree
+set(CMAKE_SKIP_BUILD_RPATH FALSE)
+
+# when building, don't use the install RPATH already
+# (but later on when installing)
+set(CMAKE_BUILD_WITH_INSTALL_RPATH FALSE)
+
+set(CMAKE_INSTALL_RPATH "${CMAKE_INSTALL_PREFIX}/lib")
+
+# add the automatically determined parts of the RPATH
+# which point to directories outside the build tree to the install RPATH
 set(CMAKE_INSTALL_RPATH_USE_LINK_PATH TRUE)
-list(FIND CMAKE_PLATFORM_IMPLICIT_LINK_DIRECTORIES "${CMAKE_LIB_PREFIX}/lib" isSystemDir)
+
+# the RPATH to be used when installing, but only if it's not a system directory
+list(FIND CMAKE_PLATFORM_IMPLICIT_LINK_DIRECTORIES "${CMAKE_INSTALL_PREFIX}/lib" isSystemDir)
 if("${isSystemDir}" STREQUAL "-1")
-    list(APPEND CMAKE_INSTALL_RPATH "${CMAKE_LIB_PREFIX}/lib")
+    set(CMAKE_INSTALL_RPATH "${CMAKE_INSTALL_PREFIX}/lib")
 endif("${isSystemDir}" STREQUAL "-1")
+message("RPATH for Install: ${CMAKE_INSTALL RPATH}.")
+
 
 
 #[[ no longer used
