+++ CMakeLists.txt
@@ -51,6 +51,7 @@ message(STATUS "CMAKE_PREFIX_PATH: " ${CMAKE_PREFIX_PATH})
 message(STATUS "CMAKE_MODULE_PATH: " ${CMAKE_PREFIX_PATH})
 #get_directory_property(_vars_before VARIABLES)
 find_package(ROOT REQUIRED COMPONENTS RIO Net)
+include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include)
 SET(ROOT_INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/include)
 SET(ROOT_LIB_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/lib)
 include_directories(${ROOT_INCLUDE_DIRS})
