### 2022-01-29 8a8110e
[ANV] add another diagnostic print
