### 2022-01-29 villano-lab/nrCascadeSim#64 @villaa
One of the most confounding things is that the failure happens only for Mac OSX and  only on Travis (local builds seem  to work fine).

Here is the compile line for `realizeCascades` on Travis:

`/Applications/Xcode-11.4.1.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++   -D__GIT_VERSION="\"8a81\"" -fprofile-arcs -ftest-coverage -isysroot /Applications/Xcode-11.4.1.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX10.15.sdk -Wl,-search_paths_first -Wl,-headerpad_max_install_names  -coverage CMakeFiles/realizeCascades.dir/bin/realizeCascades.cpp.o  -o realizeCascades  -Wl,-rpath,/Users/travis/build/nrCascadeSim_build/lib weisskopf.o lindhard.o cascadeProd.o edepmath.o rootUtil.o /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libCore.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libImt.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libRIO.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libNet.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libHist.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libGraf.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libGraf3d.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libGpad.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libROOTDataFrame.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libTree.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libTreePlayer.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libRint.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libPostscript.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libMatrix.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libPhysics.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libMathCore.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libThread.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libMultiProc.so lib/libncap.dylib /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libCore.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libImt.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libRIO.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libNet.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libHist.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libGraf.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libGraf3d.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libGpad.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libROOTDataFrame.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libTree.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libTreePlayer.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libRint.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libPostscript.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libMatrix.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libPhysics.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libMathCore.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libThread.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libMultiProc.so weisskopf.o lindhard.o cascadeProd.o edepmath.o rootUtil.o`

And then here  is the error we get upon running `realizeCascades`

```
$ ./test_realizeCascades.sh
dyld: Library not loaded: @rpath/libCore.6.20.so
  Referenced from: /usr/local/bin/realizeCascades
```
