### 2022-01-29 villano-lab/nrCascadeSim#64 @villaa
Incidentally this is the compile line that works on my Mac:

`/usr/local/Cellar/cmake/3.22.1/bin/cmake -E cmake_link_script CMakeFiles/realizeCascades.dir/link.txt --verbose=1
/Library/Developer/CommandLineTools/usr/bin/c++  -D__GIT_VERSION="\"v1.3.2-69-g4a1f\"" -fprofile-arcs -ftest-coverage -isysroot /Library/Developer/CommandLineTools/SDKs/MacOSX10.14.sdk -Wl,-search_paths_first -Wl,-headerpad_max_install_names  -coverage CMakeFiles/realizeCascades.dir/bin/realizeCascades.cpp.o -o realizeCascades  -Wl,-rpath,/Users/villaa/nrCascadeSim_build/lib weisskopf.o lindhard.o cascadeProd.o edepmath.o rootUtil.o /usr/local/lib/root/libCore.so /usr/local/lib/root/libImt.so /usr/local/lib/root/libRIO.so /usr/local/lib/root/libNet.so /usr/local/lib/root/libHist.so /usr/local/lib/root/libGraf.so /usr/local/lib/root/libGraf3d.so /usr/local/lib/root/libGpad.so /usr/local/lib/root/libROOTDataFrame.so /usr/local/lib/root/libTree.so /usr/local/lib/root/libTreePlayer.so /usr/local/lib/root/libRint.so /usr/local/lib/root/libPostscript.so /usr/local/lib/root/libMatrix.so /usr/local/lib/root/libPhysics.so /usr/local/lib/root/libMathCore.so /usr/local/lib/root/libThread.so /usr/local/lib/root/libMultiProc.so /usr/local/lib/root/libROOTVecOps.so lib/libncap.dylib /usr/local/lib/root/libCore.so /usr/local/lib/root/libImt.so /usr/local/lib/root/libRIO.so /usr/local/lib/root/libNet.so /usr/local/lib/root/libHist.so /usr/local/lib/root/libGraf.so /usr/local/lib/root/libGraf3d.so /usr/local/lib/root/libGpad.so /usr/local/lib/root/libROOTDataFrame.so /usr/local/lib/root/libTree.so /usr/local/lib/root/libTreePlayer.so /usr/local/lib/root/libRint.so /usr/local/lib/root/libPostscript.so /usr/local/lib/root/libMatrix.so /usr/local/lib/root/libPhysics.so /usr/local/lib/root/libMathCore.so /usr/local/lib/root/libThread.so /usr/local/lib/root/libMultiProc.so /usr/local/lib/root/libROOTVecOps.so weisskopf.o lindhard.o cascadeProd.o edepmath.o rootUtil.o`

It looks basically the same in the way that  it references the relevant `ROOT` libraries, like `libCore.so`,  what gives?
