### 2022-01-29 fd75c60
[ANV] try using targets for ROOT libraries
