+++ CMakeLists.txt
@@ -159,9 +159,13 @@ add_executable(realizeCascades ${binDir}/realizeCascades.cpp)
 #target_link_libraries(realizeCascades PUBLIC ROOT::Core ROOT::RIO ROOT::Tree libncap)
 target_link_libraries(realizeCascades PUBLIC libncap)
 add_executable(fetchYieldModel EXCLUDE_FROM_ALL ${testBin}/fetchYieldModel.cpp)
+target_link_libraries(fetchYieldModel PUBLIC libncap)
 add_executable(printIsotopeInfo EXCLUDE_FROM_ALL ${testBin}/printIsotopeInfo.cpp)
+target_link_libraries(printIsotopeInfo PUBLIC libncap)
 add_executable(readLevelfile EXCLUDE_FROM_ALL ${testBin}/readLevelfile.cpp)
+target_link_libraries(readLevelfile PUBLIC libncap)
 add_executable(realizeAndSave EXCLUDE_FROM_ALL ${testBin}/realizeAndSave.cpp ${incDir}/cascadeProd.h)
+target_link_libraries(realizeAndSave PUBLIC libncap)
 
 install(TARGETS realizeCascades regexPlayground RUNTIME DESTINATION ${CMAKE_INSTALL_PREFIX}/bin)
 install(TARGETS libncap LIBRARY DESTINATION ${CMAKE_LIB_PREFIX}/lib)
