### 2022-01-30 fd1e65c
Try a slightly different troubleshooting line in travis build. Last one doesn't seem to be affecting linux builds so probably irrelevant
