### 2022-01-31 villano-lab/nrCascadeSim#64 @villaa
Still working on this-- really annoying, cannot reproduce what I see on Travis. I tried a whole new directory on my Macbook Pro Mojave (10.14) and did the whole process. Eventually I get:

`/usr/local/Cellar/cmake/3.22.1/bin/cmake -E cmake_link_script CMakeFiles/realizeCascades.dir/link.txt --verbose=1
/Library/Developer/CommandLineTools/usr/bin/c++  -D__GIT_VERSION="\"v1.3.2-76-gfd1e\"" -fprofile-arcs -ftest-coverage -isysroot /Library/Developer/CommandLineTools/SDKs/MacOSX10.14.sdk -Wl,-search_paths_first -Wl,-headerpad_max_install_names  -coverage CMakeFiles/realizeCascades.dir/bin/realizeCascades.cpp.o -o realizeCascades  -Wl,-rpath,/Users/villaa/travis_nrCascadeSim/nrCascadeSim_build/lib -Wl,-rpath,/Users/villaa/travis_nrCascadeSim/nrCascadeSim/root/lib weisskopf.o lindhard.o cascadeProd.o edepmath.o rootUtil.o lib/libncap.dylib weisskopf.o lindhard.o cascadeProd.o edepmath.o rootUtil.o /Users/villaa/travis_nrCascadeSim/nrCascadeSim/root/lib/libTree.6.20.00.so /Users/villaa/travis_nrCascadeSim/nrCascadeSim/root/lib/libImt.6.20.00.so /Users/villaa/travis_nrCascadeSim/nrCascadeSim/root/lib/libNet.6.20.00.so /Users/villaa/travis_nrCascadeSim/nrCascadeSim/root/lib/libRIO.6.20.00.so /Users/villaa/travis_nrCascadeSim/nrCascadeSim/root/lib/libThread.6.20.00.so /Users/villaa/travis_nrCascadeSim/nrCascadeSim/root/lib/libCore.6.20.00.so `

So, very simillar to the latest Travis:

`/Applications/Xcode-11.4.1.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++   -D__GIT_VERSION="\"fd1e\"" -fprofile-arcs -ftest-coverage -isysroot /Applications/Xcode-11.4.1.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX10.15.sdk -Wl,-search_paths_first -Wl,-headerpad_max_install_names  -coverage CMakeFiles/realizeCascades.dir/bin/realizeCascades.cpp.o  -o realizeCascades  -Wl,-rpath,/Users/travis/build/nrCascadeSim_build/lib -Wl,-rpath,/Users/travis/build/villano-lab/nrCascadeSim/root/lib weisskopf.o lindhard.o cascadeProd.o edepmath.o rootUtil.o lib/libncap.dylib weisskopf.o lindhard.o cascadeProd.o edepmath.o rootUtil.o /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libTree.6.20.00.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libImt.6.20.00.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libNet.6.20.00.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libRIO.6.20.00.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libThread.6.20.00.so /Users/travis/build/villano-lab/nrCascadeSim/root/lib/libCore.6.20.00.so`

but the one on my Mac works!!?
