+++ CMakeLists.txt
@@ -151,6 +151,7 @@ add_executable(regexPlayground ${binDir}/regexPlayground.cpp)
 
 #link_libraries(weisskopf lindhard cascadeProd edepmath rootUtil PUBLIC ROOT::Core ROOT::RIO ROOT::Tree)
 link_libraries(weisskopf lindhard cascadeProd edepmath rootUtil ROOT::Core ROOT::RIO ROOT::Tree)
+#link_libraries(weisskopf lindhard cascadeProd edepmath rootUtil ${ROOT_LIBRARIES})
 add_library(libncap SHARED ${srcDir}/isotope_info.cpp ${srcDir}/rootUtil.cpp ${srcDir}/edepmath.cpp ${srcDir}/weisskopf.cpp ${srcDir}/lindhard.cpp ${srcDir}/cascadeProd.cpp ${incDir}/isotope_info.h ${incDir}/rootUtil.h ${incDir}/edepmath.h ${incDir}/weisskopf.h ${incDir}/lindhard.h ${incDir}/cascadeProd.h)
 
 #link_libraries(libncap PUBLIC ROOT::Core ROOT::RIO ROOT::Tree)
