### 2022-01-31 villano-lab/nrCascadeSim#64 @villaa
The only difference(s) are that on my Macbook it gets the right `__GIT_VERSION` and on Travis it _does not_. On Travis it only get's  the part after the dash, `fd1e`. Why?

Also on my Mac I accidentally used the wrong downloaded `ROOT` binary and as a result the `MacOSX10.14.sdk` was searched for but not found. But this only resulted in this warning:

```
warning: no such sysroot directory: '/Library/Developer/CommandLineTools/SDKs/MacOSX10.15.sdk'
```

but the program `realizeCascades` ran correctly anyway.
