### 2022-01-31 villano-lab/nrCascadeSim#64 @villaa
... and if I download the correct `ROOT` binary for my system, namely `root_v6.20.00.macosx64-10.14-clang110.tar.gz` instead of `root_v6.20.00.macosx64-10.15-clang110.tar.gz` it works on my Macbook without that errror.

AND I found out that travis was only using `fd1e` for the version b/c it has a different version of git where the --abbrev=4 option actually works.
