### 2022-01-31 villano-lab/nrCascadeSim#64 @villaa
AH HA! when I try to run `realizeCascades` after a _global_ installation at `/usr/local/bin` _and_ calling it from inside a second script `test_realizeCascades.sh` then I get this error on my Macbook Pro:

```(base) Anthonys-MacBook-Pro-2:tests villaa$ ./test_realizeCascades.sh
dyld: Library not loaded: @rpath/libTree.6.20.so
  Referenced from: /usr/local/bin/realizeCascades
  Reason: image not found
./test_realizeCascades.sh: line 4: 27887 Abort trap: 6           realizeCascades -n 100000 -o "output.root?reproducible=fixedname" -d 1 ../levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
Traceback (most recent call last):
  File "/Users/villaa/travis_nrCascadeSim/nrCascadeSim/tests/./isempty.py", line 1, in <module>
    import uproot
ModuleNotFoundError: No module named 'uproot'
```
