### 2022-01-31 villano-lab/nrCascadeSim#64 @villaa
When I go ahead and switch to a conda environment that has uproot I get:

```
(nrCascadeSim) Anthonys-MacBook-Pro-2:tests villaa$ ./test_realizeCascades.sh
dyld: Library not loaded: @rpath/libTree.6.20.so
  Referenced from: /usr/local/bin/realizeCascades
  Reason: image not found
./test_realizeCascades.sh: line 4: 27922 Abort trap: 6           realizeCascades -n 100000 -o "output.root?reproducible=fixedname" -d 1 ../levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
Traceback (most recent call last):
  File "/Users/villaa/anaconda3/envs/nrCascadeSim/lib/python3.8/site-packages/uproot/source/file.py", line 113, in _open
    self._file = numpy.memmap(self._file_path, dtype=self._dtype, mode="r")
  File "/Users/villaa/anaconda3/envs/nrCascadeSim/lib/python3.8/site-packages/numpy/core/memmap.py", line 225, in __new__
    f_ctx = open(os_fspath(filename), ('r' if mode == 'c' else mode)+'b')
FileNotFoundError: [Errno 2] No such file or directory: 'output.root'
```

Which seems consistent with the Travis Mac errors:

```
$ ./test_realizeCascades.sh

dyld: Library not loaded: @rpath/libTree.6.20.so

  Referenced from: /usr/local/bin/realizeCascades

  Reason: image not found

./test_realizeCascades.sh: line 4:  3321 Abort trap: 6           realizeCascades -n 100000 -o "output.root?reproducible=fixedname" -d 1 ../levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt

Traceback (most recent call last):

  File "/Users/travis/build/villano-lab/nrCascadeSim/tests/./isempty.py", line 2, in <module>

    file = uproot.open('output.root') #Change this to change tested file.

  File "/Users/travis/miniconda/envs/piptest/lib/python3.9/site-packages/uproot/rootio.py", line 54, in open

    return ROOTDirectory.read(openfcn(path), **options)

  File "/Users/travis/miniconda/envs/piptest/lib/python3.9/site-packages/uproot/rootio.py", line 51, in <lambda>

    openfcn = lambda path: MemmapSource(path, **kwargs)

  File "/Users/travis/miniconda/envs/piptest/lib/python3.9/site-packages/uproot/source/memmap.py", line 21, in __init__

    self._source = numpy.memmap(self.path, dtype=numpy.uint8, mode="r")

  File "/Users/travis/miniconda/envs/piptest/lib/python3.9/site-packages/numpy/core/memmap.py", line 228, in __new__

    f_ctx = open(os_fspath(filename), ('r' if mode == 'c' else mode)+'b')

FileNotFoundError: [Errno 2] No such file or directory: 'output.root'
```
