### 2022-02-01 villano-lab/nrCascadeSim#64 @villaa
I found that on the Mac systems, strangely, the executable `realizeCascades` works from the command line but _does not_ when called from within a script. When executed from within a script it leads to the above library error.

I stumbled on [this obscure matlab post](https://www.mathworks.com/matlabcentral/answers/552760-dyld-library-not-loaded-in-automatically-generated-shell-script). That lead me to [this post](https://www.mathworks.com/matlabcentral/answers/473971-how-do-i-configure-the-ld_library_path-on-linux-and-dyld_library_path-on-mac-os-x-to-point-to-mcr).

This suggests that `DYLD_LIBRARY_PATH` is not inherited in a daughter shell and needs  to be set _on the same line_ as the executable run  in the shell.  I made this change and it seems we are passing tests on the CMake branch.
