### 2022-02-01 c5ca9a3
[ANV] made modification to CMakeLists.txt b/c some reason didn't compile on my ubunntu laptop--despite passing all Travis tests
