+++ CMakeLists.txt
@@ -136,8 +136,8 @@ add_library(cascadeProd ${srcDir}/cascadeProd.cpp ${incDir}/cascadeProd.h ${head
 
 add_executable(regexPlayground ${binDir}/regexPlayground.cpp)
 
-#link_libraries(weisskopf lindhard cascadeProd edepmath rootUtil ROOT::Core ROOT::RIO ROOT::Tree ROOT::Thread ROOT::Rint ROOT::TreePlayer ROOT::ROOTVecOps ROOT::ROOTDataFrame ROOT::Postscript ROOT::Physics ROOT::Net ROOT::MultiProc ROOT::Matrix ROOT::MathCore ROOT::Imt ROOT::Hist ROOT::Graf ROOT::Graf3d ROOT::Gpad)
-link_libraries(weisskopf lindhard cascadeProd edepmath rootUtil ${ROOT_LIBRARIES})
+link_libraries(weisskopf lindhard cascadeProd edepmath rootUtil ROOT::Core ROOT::RIO ROOT::Tree ROOT::Thread ROOT::Rint ROOT::TreePlayer ROOT::ROOTVecOps ROOT::ROOTDataFrame ROOT::Postscript ROOT::Physics ROOT::Net ROOT::MultiProc ROOT::Matrix ROOT::MathCore ROOT::Imt ROOT::Hist ROOT::Graf ROOT::Graf3d ROOT::Gpad)
+#link_libraries(weisskopf lindhard cascadeProd edepmath rootUtil ${ROOT_LIBRARIES})
 add_library(libncap SHARED ${srcDir}/isotope_info.cpp ${srcDir}/rootUtil.cpp ${srcDir}/edepmath.cpp ${srcDir}/weisskopf.cpp ${srcDir}/lindhard.cpp ${srcDir}/cascadeProd.cpp ${incDir}/isotope_info.h ${incDir}/rootUtil.h ${incDir}/edepmath.h ${incDir}/weisskopf.h ${incDir}/lindhard.h ${incDir}/cascadeProd.h)
 
 
