+++ .travis/uploadtocodecov.sh
@@ -7,7 +7,7 @@ genhtml nrCascadeSim_coverage.info
 #Download the main uploader script
 curl -Os https://uploader.codecov.io/latest/linux/codecov
 #Check the uploader script
-curl https://keybase.io/codecovsecurity/pgp_keys.asc | gpg --no-default-keyring --keyring trustedkeys.gpg --impor$
+curl https://keybase.io/codecovsecurity/pgp_keys.asc | gpg --no-default-keyring --keyring trustedkeys.gpg --import
 curl -Os https://uploader.codecov.io/latest/linux/codecov.SHA256SUM
 curl -Os https://uploader.codecov.io/latest/linux/codecov.SHA256SUM.sig
 gpgv codecov.SHA256SUM.sig codecov.SHA256SUM
