### 2022-02-07 1c4734e
[ANV] try another configuration for present directory and lcov arguments
