+++ CMakeLists.txt
@@ -117,7 +117,7 @@ endif(NOT USE_GCOV)
 message("Use GCOV? ${USE_GCOV}")
 if(USE_GCOV)
 	SET(GCC_COVERAGE_COMPILE_FLAGS "-fprofile-arcs -ftest-coverage")
-	SET(GCC_COVERAGE_LINK_FLAGS    "-coverage")
+	SET(GCC_COVERAGE_LINK_FLAGS    "--coverage")
 	SET(CMAKE_CXX_FLAGS  "${CMAKE_CXX_FLAGS} ${GCC_COVERAGE_COMPILE_FLAGS}")
  	SET(CMAKE_EXE_LINKER_FLAGS  "${CMAKE_EXE_LINKER_FLAGS} ${GCC_COVERAGE_LINK_FLAGS}")
     message("Set up coverage flags.")
