### 2022-02-07 dd06117
[ANV] updated release notes
