### 2022-02-08 be83bd0
[ANV] update RELEASENOTES.md for fix coverage generation issues with CMake.
