+++ example-usecase/README.md
@@ -18,8 +18,10 @@ You can replace the file by running the script `generate.sh` or by the command b
 ## Files
 
 * `Yields_and_Resolutions.ipynb` - the notebook serving as the example described above.
+* `Z_G4Comparison.ipynb` - a notebook comparing data output with nrCascadeSim to data output with Geant4 in the context of applying yield models.
 * `environment.yml` - a conda environment yml file including everything needed to run this example.
 * `generate.sh` - a short shell script for running a new simulation using nrCascadeSim. Note that this will overwrite the provided file.
+* `rpune.py` - a python script for taking in Geant4 data and removing information that is not necessary for Z_G4Comparison.ipynb (see notebook).
 * `requirements.txt` - a pip requirements text file including all pip packages needed to run this example, as an alternative to the yml file.
 * `SorVsLin.svg` - a static image of one particular run of the simulation and analysis.
 * `standard.mplstyle` - formatting parameters read by the plotting library.
