+++ example-usecase/data/Nefile.txt
@@ -1,0 +1,1672 @@
+,EV,TS,P,Type,E1,D3,time1
+7001,6905,600001,100000,21010,0.0008483478565,0.0008483478565,208306.4979
+34037,33527,700001,100000,21010,0.001528019255,0.001528019255,643665.2203
+40094,39475,600001,100000,21010,0.0006556933917,0.0006556933917,115737.8085
+43774,43074,400001,100000,23010,8.998303451e-05,8.998303451e-05,124296.5292
+46782,46010,700001,100000,21010,0.000732059142,0.000732059142,190246.3407
+57337,56372,800001,100000,21010,0.001322947795,0.001322947795,665863.0331
+58052,57076,500001,100000,21010,0.0002797257344,0.0002797257344,2272.130944
+77730,76440,600001,100000,21010,0.003609386194,0.003609386194,431549.7101
+88653,87201,900001,100000,21010,0.001433597147,0.001433597147,125329.5963
+120512,118559,500001,100000,21010,0.0009695862921,0.0009695862921,110774.8972
+132676,130544,400001,100000,23010,0.0003643895541,0.0003643895541,336827.4086
+149445,146984,1100001,100000,21010,0.002294764914,0.002294764914,710149.7544
+165991,163243,600001,100000,21010,0.002691995925,0.002691995925,109249.1218
+180352,177367,600001,100000,21010,0.0002877243205,0.0002877243205,88333.16009
+197571,194352,800001,100000,21010,0.002142013715,0.002142013715,308739.9987
+4513,201172,1000001,100000,21010,0.001024627368,0.001024627368,143005.7085
+11862,208406,900001,100000,21010,0.0043753518,0.0043753518,132923.3317
+22915,219262,500001,100000,21010,0.0009052811937,0.0009052811937,4563.241932
+25376,221694,500001,100000,21010,0.001017381702,0.001017381702,725447.3156
+36014,232159,400001,100000,23010,5.893209134e-05,5.893209134e-05,105848.4998
+48036,243897,200001,100000,22010,0.00261849808,0.00261849808,405785.9348
+86913,282025,700001,100000,21010,0.0008215500425,0.0008215500425,219624.7372
+88071,283161,700001,100000,21010,0.0003835017614,0.0003835017614,53385.36157
+94810,289812,500001,100000,21010,0.0006596771029,0.0006596771029,411348.2475
+100207,295138,600001,100000,23010,0.0006479673838,0.0006479673838,67382.32153
+125471,319879,400001,100000,23010,0.0004130154887,0.0004130154887,595754.5765
+129405,323707,900001,100000,21010,0.001113745611,0.001113745611,203060.3649
+16788,409602,700001,100000,21010,0.000980088834,0.000980088834,259669.8267
+42167,434548,600001,100000,21010,0.0002246615093,0.0002246615093,632717.323
+45999,438323,500001,100000,23010,0.001005596549,0.001005596549,341610.231
+55293,447264,1100001,100000,21010,0.0009844497846,0.0009844497846,199962.6801
+63639,455485,500001,100000,21010,0.0001014830605,0.0001014830605,113536.0874
+73991,465652,900001,100000,21010,0.0003681247102,0.0003681247102,116927.9838
+86260,477696,700001,100000,21010,0.000582997835,0.000582997835,246592.0893
+86715,478144,900001,100000,21010,0.0002221163571,0.0002221163571,9201.298254
+175826,565740,700001,100000,21010,0.0005476773742,0.0005476773742,476920.2868
+184218,574020,400001,100000,21010,0.0003636172637,0.0003636172637,73322.94449
+22873,612032,400001,100000,21010,0.0002385365138,0.0002385365138,467165.8134
[…]
