### 2022-02-09 e83dbc0
Add pandas to all environment files as it is now used to import text-based data.
