+++ example-usecase/requirements.txt
@@ -5,6 +5,7 @@ ipykernel
 pytest
 uproot==3.12.0
 papermill
+pandas
 #Only really needed for testing - user can comment out but doesn't hurt anything
 nbconvert
 #hashlib
