+++ nrCascadeSim_withDocs.yml
@@ -18,4 +18,5 @@ dependencies:
     - scipy==1.6.2
     - matplotlib==3.3.2
     - breathe
+    - pandas
 #  - root
