### 2022-02-09 0a1b5ef
[ANV] toying with a scrub target which cleans and gets rid of lcov files
