+++ example-usecase/README.md
@@ -27,4 +27,5 @@ You can replace the file by running the script `generate.sh` or by the command b
 * `standard.mplstyle` - formatting parameters read by the plotting library.
 * `test_local.py` - script for CI to ensure that the notebook is running without errors.
 * `data` directory - contains some precompiled data for the notebook for those who do not wish to generate their own.
+* `plots` directory - contains plots generated by notebooks.
 * `python` directory - contains libraries used for the data processing.
