+++ docs/source/06_State_of_the_Field.rst
@@ -1,0 +1,44 @@
+==================
+State of the Field
+==================
+
+------
+Geant4
+------
+
+`Geant4  <https://geant4.web.cern.ch/>`_ is a widely popular toolkit for particle physics. 
+Unlike nrCascadeSim, Geant4 is a large toolkit with a broad range of applications 
+and less focus on any one particular application. 
+
+^^^^^^^^^^
+Comparison
+^^^^^^^^^^
+
+*Note: You can find the full code for this comparison in 
+`example-usecase/Z_G4Comparison.ipynb <https://github.com/villano-lab/nrCascadeSim/blob/master/example-usecase/Z_G4Comparison.ipynb>`_.*
+
+Most notably, Geant4 seems to be much more likely to generate single-step events, 
+in which all of the energy from the event is deposited at once, than nrCascadeSim.
+Under-generation of events with multiple energy deposits is problematic for modeling detectors with non-linear yield models -
+that is, for modeling detectors where the energy read out is not a linear function fo the energy deposited,
+making it necessary to take the yielded energy from each individual deposit separately.
+Where nrCascadeSim generated single-step events under 10% of the time for silicon, germanium, neon, and argon,
+Geant4 generated single-step events over 90% of the time for these elements.
+
+Below are plots comparing the output of nrCascadeSim to the output of Geant4. 
+Each plot has data generated by both programs for a single element.
+The energies plotted are the total energy deposits for a single event - 
+if an event had multiple steps, then the energies for each step were summed before adding them to the plot.
+The histograms are weighted to show the portion of events falling in each energy bin.
+
+.. image:: https://raw.githubusercontent.com/villano-lab/nrCascadeSim/master/example-usecase/plots/Silicon_Comparison.png
+   :width: 500
+
+.. image:: https://raw.githubusercontent.com/villano-lab/nrCascadeSim/master/example-usecase/plots/Germanium_Comparison.png
+   :width: 500
+
+.. image:: https://raw.githubusercontent.com/villano-lab/nrCascadeSim/master/example-usecase/plots/Neon_Comparison.png
+   :width: 500
+
+.. image:: https://raw.githubusercontent.com/villano-lab/nrCascadeSim/master/example-usecase/plots/Argon_Comparison.png
+   :width: 500
