### 2022-02-10 1c16235
[ANV] explicitly added all .gcno and .gcda files to make scrub target
