### 2022-02-10 93ba55b
Set fail on warning to false so it can build with files not on master.
