### 2022-02-10 #3993 @lucydot
@villaa I can see you are currently working on the Geant4 comparison testing (villano-lab/nrCascadeSim#75) - do you have a rough timescale for completion? Your software has travelled so far on the JOSS journey, I want to keep the train moving! 🚂
