### 2022-02-10 #3993 @gerudo7
> @gerudo7 is working on finalizing the Geant4 comparison and The comparison plots have been finalized and from what I understand the work remaining is to get a reasonable exposition of these facts into the documentation and the paper, is this correct @gerudo7 ? I would expect this to take just a few days to finish, do you agree @gerudo7 ?

@villaa Yes. The Geant4 comparison feature branch is just about ready to be merged back, so I think a few days is a reasonable timescale to incorporate that information into the paper.
