### 2022-02-12 c523966
[ANV]  updated release notes
