### 2022-02-12 ff40432
[ANV]  updated release notes
