+++ RELEASENOTES.md
@@ -1,3 +1,5 @@
+## Future Release (v1.4.0) Date XX.XX.XX
+
 ## Release (v1.3.3) Date 22.02.11
 
 * Hotfix 1.3.3 to add Geant4 comparison plots to master so docs render correctly before next release. 
