### 2022-02-12 8296b4d
[ANV] updated docs for  G4 comparison
