### 2022-02-12 a7b167d
[ANV] updated paper for G4 comparison
