### 2022-02-12 8bcbe36
[ANV] trying to resize figure
