### 2022-02-12 ec3e412
[ANV] updated figure size
