### 2022-02-12 55e1884
[ANV] added mention of open-source G4 and added to RELEASENOTES
