### 2022-02-12 villano-lab/nrCascadeSim#75 @villaa
We now have a comparison with G4 and are testing against G4 `v4.7.3` in the CI. This was added in PR #98
