+++ docs/source/08_Release_History.rst
@@ -2,6 +2,7 @@
 Release history
 ===============
 
+* 12.02.2022: Release of `v1.4.0 <https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.4.0>`_
 * 11.02.2022: Release of `v1.3.3 <https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.3.3>`_
 * 08.01.2022: Release of `v1.3.2 <https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.3.2>`_
 * 30.12.2021: Release of `v1.3.1 <https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.3.1>`_
