### 2022-02-12 #3993 @villaa
Ok @lucydot @altavir and @matthewfeickert we have just released `v1.4.0` -- the build system has been migrated to CMake and we have clearly documented comparisons to Geant4. These were the issues that were considered necessary for us to move on.

Huge thanks to @gerudo7 who took care of the Geant4 comparison and generally did a lot of work to get our code up to par.
