### 2022-02-13 #3993 @altavir
Looks good now. The comparison with GEANT4 shows a lot of differences, but it is for users to decide if the approximation is right or wrong in their case. The change of build configuration is also nice.
