### 2022-02-19 villano-lab/nrCascadeSim#102 @matthewfeickert
[JOSS Review] Remove recommendation to use `sudo make` as anti-pattern
I am submitting a **documentation issue**.

**The file(s) in question is/are:**

I'll link to them below using the commit hash to get a render and permalink.

**The problem is in the following category/categories:**
- [x] Clarity
- [x] Examples
- [ ] Broken links and images
- [ ] Typos, spelling, and grammar
- [ ] Undocumented Information
- [ ] Out-of-date
- [ ] Other

**Description of the problem:**
Describe whatever is wrong with the documentation or could otherwise be improved.

The [`v1.4.0` Executables of `nrCascadeSim` docs](https://nrcascadesim.readthedocs.io/en/v1.4.0/02_Executables.html#using-realizecascades-command-line-executable) still have one final instance of advocating for using `sudo make` that can be cleaned up.

https://github.com/villano-lab/nrCascadeSim/blob/83fac469648dc940585c9eee64dbe9488e0fe434/docs/source/02_Executables.rst?plain=1#L35

Otherwise great!
