### 2022-02-19 #3993 @matthewfeickert
These comments are a follow up to my review on 2022-01-17 (https://github.com/openjournals/joss-reviews/issues/3993#issuecomment-1015075303) and pertain to release [v1.4.0](https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.4.0) of the software.

## Notes

* [x] Change the `Makefile` to remove targeting `/usr/local` as a default. In many situations this is an unwritable location for users.

The change to CMake looks great and you have nice clean out-of-source builds now! :+1:

<summary>Example in Docker container:</summary>

```console
$ docker run --rm -ti -v $PWD:$PWD -w $PWD atlasamglab/stats-base:root6.24.06
root@ce876ebc60e7:/tmp/JOSS# git clone --depth 1 https://github.com/villano-lab/nrCascadeSim.git --branch v1.4.0 --single-branch
root@ce876ebc60e7:/tmp/JOSS# cmake -DCMAKE_INSTALL_PREFIX=/usr/local/venv -S nrCascadeSim -B build  # Using /usr/local/venv as it is already on PATH
root@ce876ebc60e7:/tmp/JOSS# cmake build -LH
root@ce876ebc60e7:/tmp/JOSS# cmake \
    --build build \
    --clean-first \
    --parallel $(($(nproc) - 1))
root@ce876ebc60e7:/tmp/JOSS# ls -lhtra build/
total 196K
drwxrwxr-x  4 1000 1000 4.0K Feb 19 03:50 ..
-rw-r--r--  1 root root  19K Feb 19 03:50 CMakeCache.txt
-rw-r--r--  1 root root 5.7K Feb 19 03:50 cmake_install.cmake
-rw-r--r--  1 root root  23K Feb 19 03:50 Makefile
-rwxr-xr-x  1 root root  25K Feb 19 03:50 regexPlayground
drwxr-xr-x  2 root root 4.0K Feb 19 03:50 lib
drwxr-xr-x  4 root root 4.0K Feb 19 03:50 .
-rwxr-xr-x  1 root root  99K Feb 19 03:50 realizeCascades
drwxr-xr-x 15 root root 4.0K Feb 19 03:50 CMakeFiles
root@ce876ebc60e7:/tmp/JOSS# cmake --build build --target install
root@ce876ebc60e7:/tmp/JOSS# command -v realizeCascades
/usr/local/venv/bin/realizeCascades
root@ce876ebc60e7:/tmp/JOSS# realizeCascades --help
Usage:  (null) options [ inputfile(s) ]
  -d, --seed          <integer>      Seed for random numbers
  -h, --help                         Print usage
  -n, --numgen        <number>       Number of traces to generate
  -o, --outfile       <filename>     Name the output file
  -s, --silent                       Silent, no standard out
  -v, --verbose       <level>        Print verbose messages at level <level>.
                                     Currently must use `--verbose=<level>` or `-v<level>` - no spaces.
  -V, --version                      Print version and exit
  -l, --log           <filename>     Log additional output to the specified file.
                                     If this option is not used, no logging will occur.
Report bugs to: villaa-at-gmail-dot-com
realizeCascades (nrCascadeSim) home page: <https://github.com/villano-lab/nrCascadeSim>
General help nrCascadeSim Software: <https://nrcascadesim.readthedocs.io/en/latest/>
root@ce876ebc60e7:/tmp/JOSS# time realizeCascades --verbose -n 100000 -o output.root nrCascadeSim/levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
nrCascadeSim/levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
Seed used: 3691720898
MT Result: 0.684558

real	0m1.881s
user	0m1.743s
sys	0m0.108s
root@ce876ebc60e7:/tmp/JOSS# cd build/
root@ce876ebc60e7:/tmp/JOSS/build# make uninstall
Scanning dependencies of target uninstall
Built target uninstall
root@ce876ebc60e7:/tmp/JOSS/build# ls /usr/local/venv/bin/realizeCascades
ls: cannot access '/usr/local/venv/bin/realizeCascades': No such file or directory
```

Small note: The [`v1.4.0` Download & Build docs](https://nrcascadesim.readthedocs.io/en/v1.4.0/01_Getting_Started.html#download-build) suggest to use the legacy in source build pattern and to change to Make right away

```
mkdir -p /path/to/build
cd /path/to/build
cmake -DCMAKE_INSTALL_PREFIX=/path/to/install /path/to/source
make
```

While not wrong, the legacy pattern isn't needed and even on old/stability focused operating systems like CentOS 7 in almost every situation in which software development is happening you'll have access to `devtoolset-7` or newer in which `cmake(3)` is available. So I'd recommend using the pattern of

```
$ cmake <options> -S <path to source> -B <path to build directory to make>
$ cmake <path to build directory> -LH  # See all of the options you've selected in a human readable form before build
$ cmake --build <path to build directory> --clean-first --parallel $(($(nproc) - 1))
```

as this gives the benefits of out-of-source builds, more transparency on what is happening, and allows for other build systems to be used other than Make (e.g. [Ninja](https://cmake.org/cmake/help/v3.23/generator/Ninja.html)).

For more on this please see the book [An Introduction to Modern CMake](https://cliutils.gitlab.io/modern-cmake/) by Henry Schreiner.

c.f. https://github.com/villano-lab/nrCascadeSim/issues/101

* [x] Add removal of the `.gcno` and `.gcda` files generated during the build to the `clean` Make target.
* [x] Remove recommendations of use of `sudo` to install software.

Note that the [`v1.4.0` Executables of `nrCascadeSim` docs](https://nrcascadesim.readthedocs.io/en/v1.4.0/02_Executables.html#using-realizecascades-command-line-executable) still have [one final instance of advocating for using `sudo make` that can be cleaned up](https://github.com/villano-lab/nrCascadeSim/blob/83fac469648dc940585c9eee64dbe9488e0fe434/docs/source/02_Executables.rst?plain=1#L35). Otherwise great!

c.f. https://github.com/villano-lab/nrCascadeSim/issues/102

* [x] Add at least one regression test to the test suite and plan to add more in the future.

* [x] In the State of the Field section it would be beneficial to mention that `Geant4` is open source software and to clarify if `FIFRELIN` is open source or not.

It would have been nice to clarify the `FIFRELIN` isn't an open source project, but I think the comparison that was done to `Geant4` `v10.7.3` is quite nice to have. :tada:

## Summary

@lucydot I am very impressed with the work that the team has done and I would give my full recommendation for publication with JOSS. I have opened two final GitHub issues as suggestions to the team for work in the future, but they should not be considered as further work for them to address in the review. :+1:
