### 2022-02-22 #3993 @lucydot
@villaa congratulations 🎈 we are onto the very final steps now :)

I'd like to give a huge thanks to @matthewfeickert and @altavir for their reviews. This software has been on a long and productive JOSS journey, steered forwards by your thoughtful reviews.

I'll do some checks now before passing this up to our editor-in-chief team for the final look over.
