### 2022-02-22 #3993 @villaa
@lucydot Ok, we've done those things:

- tagged release: v1.4.0
- Zenodo DOI: 10.5281/zenodo.6056681
-
