### 2022-02-22 #3993 @matthewfeickert
> I don't know if the above is something I'm supposed to do or you.

Nope you're good. One of the [journal editor-in-chiefs](https://joss.theoj.org/about#editorial_board) that is on duty this week will perform a final review of the draft generated on the joss-papers repo (linked above).

> Also, that link seems to be the PR and I don't see where the final proof is in it. I did look at the proof from one of the previous commands and it looked good.

Its there. It is just [a PDF artifact](https://github.com/openjournals/joss-papers/blob/ede3b57d6c84b6f942747b196a87bfa8c1129ea3/joss.03993/10.21105.joss.03993.pdf).
